### 2025-11-16 #9430 @editorialbot
[PRE REVIEW]: BaderKit: A Python Package for Grid-based Bader Charge Analysis
**Submitting author:** @SWeav02 (<a href="http://orcid.org/0000-0001-9358-3281">Samuel Weaver</a>)
**Repository:** https://github.com/SWeav02/baderkit
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.7.0
