### 2025-11-19 bf49dfd
small write method fix
