### 2025-11-19 0d874b3
Small changes before trying to add badelf
