### 2025-11-24 328a0f0
add convenience methods to enum
