### 2025-11-24 6003a43
add convenient regrid command
