### 2025-11-25 b7e10d5
move equal_spin to not be hidden variable
