### 2025-11-25 769d631
add sorting option for feature labeling
