### 2025-12-02 4df9570
fix major vasp formatting bug and add some docs
