+++ src/baderkit/core/labelers/bifurcation_graph/graph.py
@@ -27,8 +27,29 @@ from .nodes import IrreducibleNode, NodeBase, ReducibleNode
 
 class BifurcationGraph:
     """
-    A class for storing the nodes of a bifurcation graph. The nodes themselves
-    contain the information on their connectivity.
+    A convenience class for storing the nodes of a bifurcation graph and gathering
+    data on them.
+    
+    The nodes themselves contain the information on their connectivity.
+    
+    Parameters
+    ----------
+    structure : Structure
+        The PyMatGen structure of the chemical system.
+    labeler_type : str
+        The type of labeler used to mark irreducible features in the graph.
+    basin_maxima_frac : NDArray[float]
+        The fractional coordinates of the Bader basins in the system.
+    basin_charges : NDArray[float]
+        The integrated charges of the Bader basins in the system.
+    basin_volumes : NDArray[float]
+        The volumes of the Bader basins in the system.
+    crystalnn_kwargs : dict
+        The keyword arguments for the CrystalNN object used for finding neighbors.
+        Particularly important when calculating charges assigned to each atom.
+    atomic_radii : NDArray[float], optional
+        The radii of each atom in the system used to calculate distance beyond
+        the atom for each feature. The default is None.
     """
 
     def __init__(
@@ -41,6 +62,7 @@ class BifurcationGraph:
         crystalnn_kwargs: dict,
         atomic_radii: NDArray[float] = None,
     ):
+
         self._root_nodes = []
         self._nodes = []
         self._node_keys = {}
@@ -69,26 +91,82 @@ class BifurcationGraph:
         return f"BifurcationGraph(num_nodes={len(self._nodes)})"
 
     @property
-    def root_nodes(self):
+    def root_nodes(self) -> list:
+        """
+
+        Returns
+        -------
+        list
+            A list of nodes that have no parents
+
+        """
         return self._root_nodes
 
     @property
-    def nodes(self):
+    def nodes(self) -> list:
+        """
+
+        Returns
+        -------
+        list
+            The list of nodes in the graph.
+
+        """
         return self._nodes
 
-    def node_from_key(self, key):
+    def node_from_key(self, key: int) -> list:
+        """
+        Returns a node in the graph given its key.
+
+        Parameters
+        ----------
+        key : int
+            The integer key of the node.
+
+        Returns
+        -------
+        Node
+            The node with the corresponding key.
+
+        """
         return self._node_keys[key]
 
     @property
-    def irreducible_nodes(self):
+    def irreducible_nodes(self) -> list:
+        """
+
+        Returns
+        -------
+        list
+            The list of irreducible nodes in the graph i.e. the nodes that do
+            not have any children.
+
+        """
         return [i for i in self if not i.is_reducible]
 
     @property
-    def reducible_nodes(self):
+    def reducible_nodes(self) -> list:
+        """
+
+        Returns
+        -------
+        list
+            The list of reducible nodes in the graph i.e. the nodes that have
+            child nodes.
+
+        """
         return [i for i in self if i.is_reducible]
 
     @property
-    def sorted_reducible_nodes(self):
+    def sorted_reducible_nodes(self) -> list:
+        """
+
+        Returns
+        -------
+        list
+            A list of reducible nodes sorted by minimum value.
+
+        """
         nodes = self.reducible_nodes
         min_vals = [i.min_value for i in nodes]
         sorted_indices = np.argsort(min_vals)
@@ -96,7 +174,15 @@ class BifurcationGraph:
         return [nodes[i] for i in sorted_indices]
 
     @property
-    def unassigned_nodes(self):
+    def unassigned_nodes(self) -> list:
+        """
+
+        Returns
+        -------
+        list
+            The list of irreducible nodes that don't have an assigned feature type.
+
+        """
         return [
[…]
