### 2025-12-10 b83989b
update BadElfToolkit docs
