+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -48,7 +48,10 @@ Self = TypeVar("Self", bound="ElfLabeler")
 
 class ElfLabeler:
     """
-    A class for finding electride sites from an ELFCAR.
+    Labels chemical features present in the ELF and collects various properties
+    e.g charge, volume, elf value, etc.
+    
+    
     """
 
     _labeled_covalent = False
