### 2025-12-10 72da664
switch from quasi-atom to electride terminology
