+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -48,9 +48,9 @@ Self = TypeVar("Self", bound="ElfLabeler")
 
 class ElfLabeler:
     """
+    
     Labels chemical features present in the ELF and collects various properties
     e.g charge, volume, elf value, etc.
-    
     
     """
 
@@ -81,6 +81,8 @@ class ElfLabeler:
         vacuum_tol = False,
         **kwargs,
     ):
+        
+        
         # ensure the reference file is ELF
         if reference_grid.data_type != "elf":
             logging.warning(
@@ -125,18 +127,18 @@ class ElfLabeler:
         self._bifurcation_plot = None
         self._atom_elf_radii = None
         self._atom_elf_radii_types = None
-        self._quasi_atom_elf_radii = None
-        self._quasi_atom_elf_radii_types = None
+        self._electride_elf_radii = None
+        self._electride_elf_radii_types = None
         self._atom_nn_elf_radii = None
-        self._quasi_atom_nn_elf_radii = None
+        self._electride_nn_elf_radii = None
         self._atom_nn_elf_radii_types = None
-        self._quasi_atom_nn_elf_radii_types = None
+        self._electride_nn_elf_radii_types = None
         self._nearest_neighbor_data = None
-        self._quasi_atom_nearest_neighbor_data = None
+        self._electride_nearest_neighbor_data = None
         self._atom_features = None
-        self._quasi_atom_features = None
+        self._electride_features = None
         self._atom_max_values = None
-        self._quasi_atom_max_values = None
+        self._electride_max_values = None
 
         # TODO
         self._results_summary = None        
@@ -160,12 +162,29 @@ class ElfLabeler:
 
     @property
     def bifurcation_graph(self) -> BifurcationGraph:
+        """
+
+        Returns
+        -------
+        BifurcationGraph
+            A BifurcationGraph class representing features and bifurcations in the
+            ELF.
+
+        """
         if self._bifurcation_graph is None:
             self._get_bifurcation_graph()
         return self._bifurcation_graph
 
     @property
     def bifurcation_plot(self) -> go.Figure:
+        """
+
+        Returns
+        -------
+        go.Figure
+            A plotly graph object representing the bifurcation graph.
+
+        """
         if self._bifurcation_plot is None:
             self._bifurcation_plot = self.bifurcation_graph.get_plot()
         return self._bifurcation_plot
@@ -175,11 +194,19 @@ class ElfLabeler:
     ###########################################################################
     @property
     def structure(self) -> Structure:
+        """
+
+        Returns
+        -------
+        Structure
+            The PyMatGen Structure representing the system.
+
+        """
         return self.reference_grid.structure
 
     @property
-    def quasi_atom_structure(self) -> Structure:
-        return self.bifurcation_graph.quasi_atom_structure
+    def electride_structure(self) -> Structure:
+        return self.bifurcation_graph.electride_structure
 
     @property
     def labeled_structure(self) -> Structure:
@@ -192,10 +219,10 @@ class ElfLabeler:
         return self._nearest_neighbor_data
     
     @property
-    def quasi_atom_nearest_neighbor_data(self):
-        if self._quasi_atom_nearest_neighbor_data is None:
-            self._quasi_atom_nearest_neighbor_data = self._get_nearest_neighbor_data(self.quasi_atom_structure)
-        return self._quasi_atom_nearest_neighbor_data
+    def electride_nearest_neighbor_data(self):
+        if self._electride_nearest_neighbor_data is None:
+            self._electride_nearest_neighbor_data = self._get_nearest_neighbor_data(self.electride_structure)
+        return self._electride_nearest_neighbor_data
 
     @property
     def atom_elf_radii(self) -> NDArray[np.float64]:
[…]
