### 2025-12-11 0c4a345
ElfLabeler docs and results summaries
