+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -1,11 +1,8 @@
 # -*- coding: utf-8 -*-
 
-"""
-Contains the main class for running elf topology analysis.
-"""
-
 import logging
 import math
+import json
 import warnings
 from pathlib import Path
 from typing import Literal, TypeVar
@@ -30,33 +27,122 @@ from .elf_labeler_numba import get_feature_edges, get_min_avg_feat_surface_dists
 from .elf_radii import ElfRadiiTools
 
 Self = TypeVar("Self", bound="ElfLabeler")
+
 # TODO:
-
+# - Add modified feature volume normalized to structure volume (likely better
+# metric for hp-electrides)
 # - Add option for calculating atom radius from volume rather than bonds
-
+# - Add electride dimensionality calculation here. Ideally use numba method
 # - add a method to print a more traditional bifurcation plot?
-# - create method to get all radii around each atom using new radii method:
-# 1. get atom neighbors (up to distance or dynamically)
-# 2. find radii for closest and add to official neighbors
-# 3. go to next closest and get radii. See if they fit under planes made
-# by other radii
-# 4. Continue until all neighbors found. (Whats the best stop condition?)
-# - rework BadELF voxel assignment using Numba. That should just be so much
-# faster. I could probably even reimplement exact voxel splitting
-# - Update simmate workflows (and badelf -_- )
 
 
 class ElfLabeler:
     """
     Labels chemical features present in the ELF and collects various properties
-    e.g charge, volume, elf value, etc.
+    e.g charge, volume, elf value, etc. 
+    This is originally designed for analyzing electride materials, but could be
+    quite useful for bond analysis as well.
     
+    This class is designed only for single spin or total spin charge densities 
+    and ELF. For spin-dependent systems, use the SpinElfLabeler instead.
     
+    Parameters
+    ----------
+    charge_grid : Grid
+        The charge density grid used for integrating charge.
+    reference_grid : Grid
+        The ELF grid used to partition volumes.
+    ignore_low_pseudopotentials : bool, optional
+        Whether or not to ignore errors associated with the use of pseudopotentials
+        with limited valence electrons. The default is False.
+    shared_shell_ratio : float, optional
+        The ratio used to determine if shallow nodes surrounding an atom should
+        be considered as a single shell feature. 
+        
+        Highly symmetric features such
+        as atom shells will often split into smaller features at a high isosurface
+        value very close to the maximum value in these smaller features. This
+        ratio refers to the range of values the ELF feature exists as a 
+        single feature divided by the highest ELF value in the feature.
+        
+        As covalency increases, this ratio will decrease as the small features
+        grow and eventually form large covalent features that align along a bond. 
+        Thus, this parameter is in some ways a cutoff for considering a 
+        feature a covalent bond versus an ionic shell. The default is 0.75.
+    covalent_molecule_ratio : float, optional
+        The ratio used to determine if an ELF feature belongs to one or multiple
+        atomic shell systems. 
+        
+        The covalent-ionic spectrum can be visualized in the ELF by comparing
+        the ELF values at which Shell-like features form and fully surround
+        atoms. "Fully surround" in this case means there is a void space
+        within the shell that is not the same surface as the outside of the shell.
+        In ionic bonds, the outermost shells will fully surround a single atom
+        at a relatively high ELF value and will not connect to other features
+        to surround additional atoms until considerably lower ELF values. 
+        In contrast, homogenous covalent bonds will never fully surround 1 atom, 
+        instead surrounding at least two when they first form a shell-like
+        shape. Heterogenous covalent bonds sit somewhere between the two,
+        surrounding a single atom for a small range of values before connecting
+        to a feature surrounding additional atoms.
+        
+        The ratio here is the range of ELF values where the feature belongs
[…]
