### 2025-12-19 bc688a0
add badelf
