### 2025-12-19 1b20c1c
clean up badelf some
