+++ src/baderkit/core/__init__.py
@@ -4,3 +4,5 @@ from .bader import Bader
 from .toolkit import Grid, Structure
 
 from .labelers import ElfLabeler, SpinElfLabeler
+
+from .badelf import Badelf, SpinBadelf
