+++ src/baderkit/core/badelf/badelf.py
@@ -2,7 +2,6 @@
 
 import json
 import logging
-import math
 import os
 import warnings
 from pathlib import Path
@@ -10,17 +9,14 @@ from typing import Literal
 
 import numpy as np
 from numpy.typing import NDArray
-import psutil
-from pymatgen.analysis.local_env import CrystalNN
 from pymatgen.io.vasp import Potcar
 from scipy.ndimage import label
 from tqdm import tqdm
 
 from baderkit.core.utilities.file_parsers import Format
-from baderkit.core import Structure, ElfLabeler, Grid, Bader, SpinElfLabeler
+from baderkit.core import Structure, ElfLabeler, Grid, SpinElfLabeler
 from baderkit.core.labelers.bifurcation_graph.enum_and_styling import FeatureType, FEATURE_DUMMY_ATOMS
 from baderkit.core.bader.methods.shared_numba import get_min_avg_surface_dists, get_edges
-from baderkit.core.labelers.elf_labeler.elf_radii import ElfRadiiTools
 from baderkit.core.badelf.badelf_numba import get_badelf_assignments
 
 class Badelf:
@@ -41,34 +37,33 @@ class Badelf:
     charge_grid : Grid
         A badelf app Grid like object used for summing charge. Usually
         contains charge density.
-    threads : int, optional
-        The number of threads to use for voxel assignment.
-        Defaults to 0.9*the total number of threads available.
-    algorithm : Literal["badelf", "voronelf", "zero-flux"], optional
-        The algorithm to use for partitioning electrides from the nearby
+    method : Literal["badelf", "voronelf", "zero-flux"], optional
+        The method to use for partitioning electrides from the nearby
         atoms.
             'badelf' (default)
                 Separates electrides using zero-flux surfaces then uses
                 planes at atom radii to separate atoms. This may give more reasonable
-                results for atoms, particularly in ionic solids.
+                results for atoms, particularly in ionic solids. Radii are
+                calculated directly from the ELF.
             'voronelf'
                 Separates both electrides and atoms using planes at atomic/electride
                 radii. This is not recommended for electrides that are not
                 spherical, but may provide better results for those that are.
+                Radii are calculated directly from the ELF.
             'zero-flux'
                 Separates electrides and atoms using zero-flux surface. This
                 is the most traditional ELF analysis, but may display some
                 bias towards atoms with higher ELF values. Results for electride
                 sites are identical to BadELF, and the method can be significantly
                 faster.
-    shared_feature_splitting_method : Literal["plane", "pauling", "equal", "dist", "nearest"], optional
+    shared_feature_splitting_method : Literal["pauling", "equal", "dist", "nearest"], optional
         The method of assigning charge from shared ELF features
-        such as covalent or metallic bonds. The default is "plane".
-            'plane' (default)
-                Distributes charge based on the voronoi surface belonging to 
-                each atom. Charge will not be assigned to electrides unless
-                the voronelf method is used, and this method cannot be used
-                with the zero-flux method.
+        such as covalent or metallic bonds. This parameter is only used with the
+        zero-flux method.
+            'weighted_dist' (default)
+                Fraction increases with decreasing distance to each atom. The
+                fraction is further weighted by the radius of each atom
+                calculated from the ELF
             'pauling'
                 Distributes charge to neighboring atoms (calculated using CrystalNN)
                 based on the pualing electronegativity of each species normalized
@@ -83,24 +78,6 @@ class Badelf:
                 of (1/dist) to each neighboring atom.
             'nearest'
                 Gives all charge to the nearest atom or electride site.
-    labeled_structure : Structure, optional
-        A pymatgen structure object with dummy atoms representing
[…]
