+++ src/baderkit/core/labelers/elf_labeler/elf_radii.py
@@ -1,9 +1,4 @@
 # -*- coding: utf-8 -*-
-"""
-This is a reimplementation of the ionic radius finder I created for BadELF in
-[Simmate](https://github.com/jacksund/simmate/blob/main/src/simmate/apps/badelf/core/partitioning.py)
-"""
-
 
 import numpy as np
 from numpy.typing import NDArray
@@ -13,7 +8,13 @@ from pymatgen.analysis.local_env import CrystalNN
 
 
 from .elf_radii_numba import get_all_atom_elf_radii
-from baderkit.core.utilities.voronoi import get_voronoi_planes_exact, reduce_voronoi_planes_conservative
+from baderkit.core.utilities.voronoi import (
+    get_planes_on_surface,
+    get_canonical_bonds,
+    generate_symmetric_bonds,
+    )
+
+from scipy.spatial import HalfspaceIntersection
 
 class ElfRadiiTools:
     def __init__(
@@ -38,38 +39,14 @@ class ElfRadiiTools:
 
         # NOTE: atom labels should correspond to labeled structure
 
-    def get_all_neigh_elf_radii_and_type(
+    def _get_elf_radii_and_type(
         self,
         site_indices,
         neigh_indices,
         neigh_frac_coords,
-        neigh_dists,
+        pair_dists,
+        reversed_bonds,
     ):
-        # first, get the unique bonds. We do this here because numba does not
-        # currently allow np.unique axis argument
-        # 1. get equivalent indices by symmetry for each atom/neighbor
-        equivalent_atoms = self.structure.equivalent_atoms
-        equiv_sites = equivalent_atoms[site_indices]
-        equiv_neighs = equivalent_atoms[neigh_indices]
-
-        # Create an array that is sorted such that the lower index of the bond
-        # always come first. We will use this to avoid calculating the radius for
-        # reverse bonds
-        sorted_pair_array = np.column_stack((equiv_sites, equiv_neighs))
-        sorted_indices = np.argsort(sorted_pair_array, axis=1)
-        # get the indices that need to be flipped (we need these later)
-        flipped_mask = sorted_indices[:,0] == 1
-        # flip those indices
-        flipped_indices = np.where(flipped_mask)[0]
-        sorted_pair_array[flipped_indices]=np.fliplr(sorted_pair_array[flipped_indices])
-        sorted_pair_array = np.column_stack((sorted_pair_array, neigh_dists))
-
-        # get the unique bonds as a map pointing from a bond index to the equivalent
-        # bond
-        unique_bonds, index, inverse = np.unique(
-            sorted_pair_array, return_inverse=True, return_index=True, axis=0
-        )
-        equivalent_bonds = index[inverse]
         
         # create a map of covalent labels
         covalent_labels = np.zeros(len(self.feature_structure), dtype=np.bool_)
@@ -82,13 +59,83 @@ class ElfRadiiTools:
             neigh_indices=neigh_indices,
             site_frac_coords=self.structure.frac_coords,
             neigh_frac_coords=neigh_frac_coords,
-            neigh_dists=neigh_dists,
-            equivalent_bonds=equivalent_bonds,
-            reversed_bonds=flipped_mask,
+            neigh_dists=pair_dists,
+            reversed_bonds=reversed_bonds,
             data=self.cubic_coeffs,
             feature_labels=self.feature_labels,
             covalent_labels=covalent_labels,
+            equivalent_atoms=self.structure.equivalent_atoms,
         )
+    
+    def _get_neigh_info_from_pymatgen(
+        self,
+        neigh_info: list
+            ):
+        # for each site, get all neighbors within a sphere of twice the largest
+        # CrystalNN neighbor distance. Get relavent information
+        neigh_indices = []
[…]
