+++ src/baderkit/core/utilities/voronoi.py
@@ -4,193 +4,530 @@ import numpy as np
 from numpy.typing import NDArray
 from numba import njit, prange
 
+###############################################################################
+# Plane related functions
+###############################################################################
 
 @njit(fastmath=True, cache=True)
 def get_plane_dist(
     point: NDArray,
-    plane_vector: NDArray,
-    plane_point: NDArray,
+    plane_equation: NDArray,
 ):
     """
-    Gets the sign associated with a point compared with a plane.
+    Gets the signed distance of a point from a plane
 
     Args:
         point (ArrayLike):
             A point in cartesian coordinates to compare with a plane
-        plane_vector (ArrayLike):
-            The vector normal to the plane of interest
-        plane_point (ArrayLike):
-            A point on the plane of interest
+        plane_equation (ArrayLike):
+            The plane equation (A*x+b <= 0)
 
     Returns:
         The distance of the point to the plane.
     """
-    # get all of the points in cartesian coordinates
-    x, y, z = plane_point
-    a, b, c = plane_vector
-    x1, y1, z1 = point
+
+    a, b, c, d = plane_equation
+    x, y, z = point
     value_of_plane_equation = round(
-        (a * (x - x1) + b * (y - y1) + c * (z - z1)), 12
+        a*x + b*y + c*z + d, 12
     )
-    # positive value is "below" (opposite normal)
+    # negative value is "below" (opposite normal)
     return value_of_plane_equation
 
 @njit(parallel=True, cache=True)
-def reduce_voronoi_planes_conservative(
-    site_indices,
-    plane_points,
-    plane_vectors,
-    angle_tol=1e-6,
-    dist_tol=1e-12,
-):
-    n = len(site_indices)
-    important = np.ones(n, dtype=np.bool_)
-
-    for i in prange(n):
-        si = site_indices[i]
-        ni = plane_vectors[i]
-        pi = plane_points[i]
-
-        for j in range(n):
-            if j == i or site_indices[j] != si:
-                continue
-
-            nj = plane_vectors[j]
-            pj = plane_points[j]
-
-            # check near-parallel
-            dot = np.dot(ni, nj)
-            if dot < 1 - angle_tol:
-                continue
-
-            # project pi onto plane j normal
-            d = np.dot(pi - pj, nj)
-
-            if d < -dist_tol:
-                important[i] = False
-                break
-
-    return important
-
-@njit(cache=True)
-def intersect_three_planes(p1, n1, p2, n2, p3, n3, tol=1e-12):
-    """
-    Returns (success, point)
-    """
-    A = np.vstack((n1, n2, n3))
-    det = np.linalg.det(A)
-    if abs(det) < tol:
-        return False, np.zeros(3)
-
-    b = np.array([
-        np.dot(n1, p1),
-        np.dot(n2, p2),
-        np.dot(n3, p3),
-    ])
-
-    x = np.linalg.solve(A, b)
-    return True, x
-
-@njit(cache=True)
-def satisfies_all_planes(
-        x, 
-        i,
-        j,
-        k,
-        si,
-        site_indices,
-        important_indices,
-        important,
-        plane_points, 
-        plane_vectors, 
+def get_planes_on_surface(
+    plane_equations,
+    vertices,
+    tol=1e-12,
         ):
-    for idx in important_indices:
-        # skip points that don't belong to this site or are part of the group
-        # of planes making up this intercept
-        if site_indices[idx] != si:
-            continue
-        if idx == i or idx == j or idx == k:
-            continue
-        # if this is a previous plane that we've found to not be important, we
-        # can skip it
-        if idx < i and not important[idx]:
-            continue
-        # check if this intercept is outside this plane. If so, it is not important
-        if get_plane_dist(
-                point=x, 
-                plane_point=plane_points[idx], 
[…]
