### 2025-12-23 7d42643
remove symmetry.py
