+++ src/baderkit/core/utilities/symmetry1.py
@@ -1,278 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from numba import njit, prange
-import numpy as np
-from numpy.typing import NDArray
-
-@njit(cache=True)
-def find_site_in_tol(
-    chunked_coords,
-    site_coords,
-    tol,
-        ):
-    chunked = np.round(site_coords / tol).astype(np.int64)
-    # Match wrapped coord to site
-    index = -1
-    for j, chunked_coord in enumerate(chunked_coords):
-        
-        d = chunked - chunked_coord
-        if np.sum(d) == 0:
-            index = j
-            break
-
-    return index
-
-@njit(cache=True)
-def get_canonical_displacement(bond_displacement, tol):
-    # wrap into cell
-    bond_displacement -= np.round(bond_displacement)
-    
-    # quantize to tolerance
-    v = np.round(bond_displacement / tol).astype(np.int64)
-
-    # choose lexicographically positive representative
-    if (
-        v[0] < 0 or
-        (v[0] == 0 and v[1] < 0) or
-        (v[0] == 0 and v[1] == 0 and v[2] < 0)
-    ):
-        v[0] = -v[0]
-        v[1] = -v[1]
-        v[2] = -v[2]
-
-    return v
-
-
-@njit(cache=True)
-def get_canonical_bond(
-        site_idx,
-        neigh_idx,
-        neigh_coords,
-        all_frac_coords, 
-        equivalent_atoms,
-        rotation_matrices,
-        translation_vectors,
-        included_ops,
-        pair_dist,
-        tol,
-        ):
-    
-    # get frac coords for the site
-    site_coords = all_frac_coords[site_idx]
-    
-    # create a placeholder for the best canonical bond
-    best_rep = np.full(7, np.iinfo(np.int64).max, dtype=np.int64)
-    
-    if equivalent_atoms[site_idx] <= equivalent_atoms[neigh_idx]:
-        best_rep[0] = 0 # bond is not inverted
-        best_rep[1] = equivalent_atoms[site_idx]
-        best_rep[2] = equivalent_atoms[neigh_idx]
-    else:
-        best_rep[0] = 1 # bond is inverted
-        best_rep[2] = equivalent_atoms[site_idx]
-        best_rep[1] = equivalent_atoms[neigh_idx]
-    
-    # get quantized pair distance to retain information on bond length
-    best_rep[6] = round(pair_dist/tol)
-
-    # iterate over valid symmetry operations
-    for trans_idx, valid in enumerate(included_ops):
-        # skip operations that don't transform to the lowest index equivalent atom
-        # (checked ahead of time for speed)
-        if not valid:
-            continue
-        
-        # get transformed site and neighbor
-        matrix = rotation_matrices[trans_idx]
-        vector = translation_vectors[trans_idx]
-        
-        trans_site_coords = matrix @ site_coords + vector
-        trans_neigh_coords = matrix @ neigh_coords + vector
-        
-        # get the displacement vector
-        displacement = trans_neigh_coords - trans_site_coords
-        
-        # canonize
-        displacement = get_canonical_displacement(displacement, tol)
-        
-        # check if this is a better representation than the current one
-        if displacement[0] > best_rep[3]:
-            continue
-        elif displacement[1] > best_rep[4]:
-            continue
-        elif displacement[2] > best_rep[5]:
-            continue
-        
-        # if we're still here, this is equal or better than the best rep
-        best_rep[3:6] = displacement
-
-    # Choose unique canonical representative
-    return best_rep
-
-
-@njit(cache=True)
-def perform_symmop(coords, rotation_matrix: NDArray, translation_vector: NDArray):
-    return rotation_matrix @ coords + translation_vector
-
-        
-@njit(parallel=True, cache=True)
-def get_canonical_bonds(
-        site_indices,
[…]
