### 2025-12-24 9f4762a
badelf index fixing
