+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
+import json
 import logging
 import math
-import json
 import warnings
 from pathlib import Path
 from typing import Literal, TypeVar
@@ -19,7 +19,6 @@ from baderkit.core.labelers.bifurcation_graph import (
     DomainSubtype,
     FeatureType,
 )
-
 from baderkit.core.utilities.coord_env import check_all_covalent
 from baderkit.core.utilities.file_parsers import Format
 
@@ -39,13 +38,13 @@ Self = TypeVar("Self", bound="ElfLabeler")
 class ElfLabeler:
     """
     Labels chemical features present in the ELF and collects various properties
-    e.g charge, volume, elf value, etc. 
+    e.g charge, volume, elf value, etc.
     This is originally designed for analyzing electride materials, but could be
     quite useful for bond analysis as well.
-    
-    This class is designed only for single spin or total spin charge densities 
+
+    This class is designed only for single spin or total spin charge densities
     and ELF. For spin-dependent systems, use the SpinElfLabeler instead.
-    
+
     Parameters
     ----------
     charge_grid : Grid
@@ -57,35 +56,35 @@ class ElfLabeler:
         with limited valence electrons. The default is False.
     shared_shell_ratio : float, optional
         The ratio used to determine if shallow nodes surrounding an atom should
-        be considered as a single shell feature. 
-        
+        be considered as a single shell feature.
+
         Highly symmetric features such
         as atom shells will often split into smaller features at a high isosurface
         value very close to the maximum value in these smaller features. This
-        ratio refers to the range of values the ELF feature exists as a 
+        ratio refers to the range of values the ELF feature exists as a
         single feature divided by the highest ELF value in the feature.
-        
+
         As covalency increases, this ratio will decrease as the small features
-        grow and eventually form large covalent features that align along a bond. 
-        Thus, this parameter is in some ways a cutoff for considering a 
+        grow and eventually form large covalent features that align along a bond.
+        Thus, this parameter is in some ways a cutoff for considering a
         feature a covalent bond versus an ionic shell. The default is 0.75.
     covalent_molecule_ratio : float, optional
         The ratio used to determine if an ELF feature belongs to one or multiple
-        atomic shell systems. 
-        
+        atomic shell systems.
+
         The covalent-ionic spectrum can be visualized in the ELF by comparing
         the ELF values at which Shell-like features form and fully surround
         atoms. "Fully surround" in this case means there is a void space
         within the shell that is not the same surface as the outside of the shell.
         In ionic bonds, the outermost shells will fully surround a single atom
         at a relatively high ELF value and will not connect to other features
-        to surround additional atoms until considerably lower ELF values. 
-        In contrast, homogenous covalent bonds will never fully surround 1 atom, 
+        to surround additional atoms until considerably lower ELF values.
+        In contrast, homogenous covalent bonds will never fully surround 1 atom,
         instead surrounding at least two when they first form a shell-like
         shape. Heterogenous covalent bonds sit somewhere between the two,
         surrounding a single atom for a small range of values before connecting
         to a feature surrounding additional atoms.
-        
+
         The ratio here is the range of ELF values where the feature belongs
         to a parent feature surrounding multiple atoms divided by the range of values
         where it belongs to a parent feature surrounding only 1 atoms. This
@@ -118,7 +117,7 @@ class ElfLabeler:
         The minimum charge a feature must have to be considered an electride
[…]
