### 2025-12-29 9934320
add cli tests
