+++ docs/api_reference/core/badelf_spin.md
@@ -1,0 +1,1 @@
+::: src.baderkit.core.badelf.SpinBadelf
