+++ src/baderkit/core/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from .toolkit import Grid, Structure # isort:skip
+from .toolkit import Grid, Structure  # isort:skip
 
-from .bader import Bader # isort:skip
+from .bader import Bader  # isort:skip
 
-from .labelers import ElfLabeler, SpinElfLabeler # isort:skip
+from .labelers import ElfLabeler, SpinElfLabeler  # isort:skip
 
-from .badelf import Badelf, SpinBadelf # isort:skip
+from .badelf import Badelf, SpinBadelf  # isort:skip
