### 2026-01-02 8343b1f
fix floating point issues with tests
