### 2026-01-03 4e1e7c0
initial doc updates
