+++ docs/bader/background.md
@@ -26,7 +26,7 @@ sometimes termed an *attractor*. Each attractor typically (though not always)
 correspond to an atom, and the charge and oxidation state of the atom can be determined 
 by integrating the charge density within this region.
 
-![bader_separation](images/bader_separation_wb.png)
+![bader_separation](/images/bader_separation_wb.png)
 
 In practice, it is often difficult and computationally expense to thoroughly 
 sample the zero-flux surface defining basins. To avoid this problem, 
@@ -41,6 +41,6 @@ surface.
 
 Since the development of their original algorithm the Henkelman group and others have
 developed several improved methods for performing this steepest ascent. Each has
-its own advantages and disadvantages. We recommend reading through our [Methods and Benchmarks](/baderkit/implementation)
+its own advantages and disadvantages. We recommend reading through our [Methods and Benchmarks](/baderkit/bader/methods)
 page to determine the best one for your use case.
     
