+++ docs/bader/usage.md
@@ -1,0 +1,178 @@
+## Introduction
+
+The `Bader` class reproduces the methods for Bader charge analysis available in the 
+[Henkelman group's](https://theory.cm.utexas.edu/henkelman/code/bader/) excellent 
+Fortran code within the Python ecosystem. It is built on
+top of the [PyMatGen](https://pymatgen.org/) package, allowing for
+easy extension to other projects. Under the hood, BaderKit runs on [Numba](https://numba.pydata.org/numba-doc/dev/index.html) 
+and [NumPy](https://numpy.org/doc/stable/index.html) to parallelize and vectorize
+calculations. This allows BaderKit to reach speeds [comparable or faster](/methods/#__tabbed_2_1)
+than the original code.
+
+## Basic Use
+
+BaderKit can be used through the command line interface or through
+Python script. This page covers only the most simple use case of running 
+Bader charge analysis on a VASP `CHGCAR` or Gaussian `cube` file. For more 
+advance usage, see our [API reference](/baderkit/api_reference/core/bader) 
+and [Examples](/baderkit/examples) pages.
+
+
+=== "Command Line"
+
+    1. Activate your environment with BaderKit installed. If you are not using an
+    environment manager, skip to step 2.
+    
+        ```bash
+        conda activate my_env
+        ```
+        
+    2. Navigate to the directory with your charge density file.
+    
+        ```bash
+        cd /path/to/directory
+        ```
+    
+    3. Run the bader analysis. Replace 'chargefile' with the name of your file.
+    
+        ```bash
+        baderkit run chargefile
+        ```
+    
+    Output files for atoms and bader basins will be written automatically to 
+    `bader_atom_summary.tsv` and `bader_basin_summary.tsv` respectively. Each file includes rows for each atom or basin with columns for:
+
+    - atom labels/assignments
+    - coordinates (fractional) 
+    - charges (*e*) 
+    - volumes (Å<sup>3</sup>)
+    - minimum surface distances (Å)
+    - distances to nearest atom (Å) <small>**basins only**</small>
+    
+    Additional arguments and options such as those for printing output files or using different 
+    algorithms can be viewed by running the help command.
+    ```bash
+    baderkit run --help
+    ```
+
+=== "Python"
+    
+    1. Import the `Bader` class.
+    
+        ```python
+        from baderkit.core import Bader
+        ```
+    
+    2. Use the `Bader` class' `from_dynamic` method to read a `CHGCAR` or `cube`
+    file.
+    
+        ```python
+        # instantiate the class
+        bader = Bader.from_dynamic("path/to/charge_file")
+        ```
+    
+    3. To run the analysis, we can call any class property. Try getting a complete
+    summary in dictionary format.
+        ```python
+        results = bader.to_json()
+        ```
+    
+    4. Now try getting individual properties. For more details on each property,
+    see the [API reference](../api_reference/core/bader/#src.baderkit.core.bader.Bader).
+        ```python
+        atom_charges = bader.atom_charges # Total atom charges
+        atom_labels = bader.atom_labels # Atom assignments for each point in the grid
+        basin_volumes = bader.basin_volumes # The volumes of each bader basin
+        maxima_coords = bader.basin_maxima_frac # Frac coordinates of each attractor
+        ```
+    
+    5. BaderKit also provides convenience methods for writing results to file. First,
+    let's write a summary of the full analysis.
+    
+        ```python
+        bader.write_json()
+        ```
+    
+    6. Now let's write the volume assigned to one of our atoms.
+    
+        ```python
+        bader.write_atom_volumes(atom_indices = [0])
+        ```
+    
+    !!! Tip
[…]
