+++ docs/elf_labeler/background.md
@@ -1,0 +1,69 @@
+# Background
+
+!!! Note
+    This page assumes prior knowledge of Bader's method for quantifying atomic charge and oxidation states. See [Bader](../../bader/background) for more information.
+
+## The Electron Localization Function
+
+The Electron Localization Function (ELF), as its name suggests, is a measure of the localization of electrons in a system. It was originally developed in 1990 by [Becke and Edgecombe](https://doi.org/10.1063/1.458517) and was very quickly adopted as a tool for the qualitative and quantitative analysis of chemical features and bonding
+
+![elf features](/images/elf_features.png)
+
+There are many interpretations of the ELF, but our favorite comes from [J.K. Burdett and T.A. McCormick](https://pubs.acs.org/doi/10.1021/jp9820774). For those who are not so math oriented, don't worry. The equations simplify very quickly.
+
+For a system with spin $\sigma$, the charge density, $\rho_{\sigma}$, and a set of occupied molecular orbitals, ${\psi_{i\sigma}}$, the ELF is given by
+
+$$
+\begin{equation}
+\text{ELF} = \frac{1}{1 + (\frac{D_{\sigma}}{D_{\sigma,gas}})^2}\
+\label{eq:1}
+\end{equation}
+$$
+
+where,
+
+$$
+\begin{equation}
+D_{\sigma}/D_{\sigma,gas} = 0.3483\rho^{-5/3}_{\sigma}\Big[\sum_{i}|\nabla\psi_{i\sigma}|^2 - \frac{1}{8}|\nabla\rho_{\sigma}|^2/\rho_{\sigma}\Big]
+\label{eq:2}
+\end{equation}
+$$
+
+The $D_{\sigma}$ is the first term in the Taylor expansion of the spherically averaged $\sigma$-spin pair probability, and the subscript "gas" refers to the corresponding value for the homogenous electron gas. 
+
+The ELF ranges from 0 to 1 and has a value of 0.5 for the homogenous electron gas. Therefore, an ELF value of 1.0 corresponds to a fully localized region and a value of 0.5 to a fully delocalized region. Values below 0.5 are less intuitive. The key point that will allow us to simplify our current interpretation of the ELF is that the first term of $\eqref{eq:2}$ is the kinetic energy density at a given point and the latter term is proportional to the Weiszäcker functional (the kinetic energy density at a point if only one occupied orbital were to contribute to $\rho$). The first term is always greater than the second and in practice is typically much larger. This means that for most systems, the dominant part of equation $\eqref{eq:2}$ is
+
+$$
+\sum_i\frac{|\nabla\psi_{i\sigma}|^2}{\rho^{5/3}}
+$$
+
+Thus, the ELF is primarily dependent on the kinetic energy density (numerator) and charge density (denomenator). As this term is inverted in equation $\eqref{eq:1}$, the ELF increases with charge density and decreases with kinetic energy. This is chemically intuitive, as it suggests high localization where there are many electrons or where electrons are not moving significantly. Another useful note is that the ELF tends to decrease heavily in regions where there are a significant number of nodes in the occupied orbitals. The increase in charge density in regions with nodes is far outweighed by the contribution of nodes to the kinetic energy. Thus an ELF value below 0.5 can be understood as a region with very little electron density (i.e. interstitial areas between atoms) or regions with a great number of nodes.
+
+---
+
+## Topology of the ELF
+
+!!! note
+ This discussion is largely based on this review by Carlo Gatti: [Chemical bonding in crystals: new directions](https://doi.org/10.1524/zkri.220.5.399.65073).
+ 
[…]
