+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -28,6 +28,7 @@ from .elf_radii import ElfRadiiTools
 Self = TypeVar("Self", bound="ElfLabeler")
 
 # TODO:
+# - Add distinction between core and valence shells using POTCAR if desired
 # - Add modified feature volume normalized to structure volume (likely better
 # metric for hp-electrides)
 # - Add option for calculating atom radius from volume rather than bonds
@@ -107,21 +108,25 @@ class ElfLabeler:
         bonds and metallic bonds with covalent-like behavior may form ring
         like shapes due to digeneracy with maxima slightly of the bond, decreasing
         this angle. The default is 135.
+    max_metal_depth : float, optional
+        The maximum depth a feature can have and be considered metallic. Features
+        between multiple atoms with a depth lower than this value will be marked
+        as metallic. The default is 0.1.
     min_electride_elf_value : float, optional
         The minimum ELF value for a feature to be considered an electride rather
-        than a metallic bond. The default is 0.5.
+        than a multi-centered bond. The default is 0.5.
     min_electride_depth : float, optional
         The minimum range of ELF values a feature must exist distinctly to
-        be considered an electride rather than a metal. The default is 0.2.
+        be considered an electride rather than a multi-centered bond. The default is 0.2.
     min_electride_charge : float, optional
         The minimum charge a feature must have to be considered an electride
-        rather than a metal. The default is 0.5.
+        rather than a multi-centered bond. The default is 0.5.
     min_electride_volume : float, optional
         The minimum volume a feature must have to be considered an electride
-        rather than a metal. The default is 10.
+        rather than a multi-centered bond. The default is 10.
     min_electride_dist_beyond_atom : float, optional
         The minimum distance beyond the atoms radius that an electride must
-        sit to be considered an electride rather than a metal. The radius
+        sit to be considered an electride rather than a multi-centered bond. The radius
         is calculated as the minimum in the ELF between two atoms. If a covalent
         or metallic bond is present, the maximum is used instead. The default is 0.3.
     crystalnn_kwargs : dict, optional
@@ -140,7 +145,7 @@ class ElfLabeler:
     """
 
     _labeled_covalent = False
-    _labeled_metallic = False
+    _labeled_multi = False
     spin_system = "total"
 
     def __init__(
@@ -153,6 +158,7 @@ class ElfLabeler:
         combine_shells: bool = True,
         min_covalent_charge: float = 0.6,
         min_covalent_angle: float = 135,
+        max_metal_depth: float = 0.1,
         min_electride_elf_value: float = 0.5,
         min_electride_depth: float = 0.2,
         min_electride_charge: float = 0.5,
@@ -211,18 +217,18 @@ class ElfLabeler:
         self._bifurcation_plot = None
         self._atom_elf_radii = None
         self._atom_elf_radii_types = None
-        self._electride_elf_radii = None
-        self._electride_elf_radii_types = None
+        self._atom_elf_radii_e = None
+        self._atom_elf_radii_types_e = None
         self._atom_nn_elf_radii = None
-        self._electride_nn_elf_radii = None
+        self._atom_nn_elf_radii_e = None
         self._atom_nn_elf_radii_types = None
-        self._electride_nn_elf_radii_types = None
+        self._atom_nn_elf_radii_types_e = None
         self._nearest_neighbor_data = None
-        self._electride_nearest_neighbor_data = None
+        self._nearest_neighbor_data_e = None
         self._atom_feature_indices = None
         self._atom_max_values = None
         self._atom_nn_planes = None
-        self._electride_nn_planes = None
+        self._atom_nn_planes_e = None
 
[…]
