### 2026-01-07 e89e4b7
fix small pymatgen version issue
