### 2026-01-08 d948c0c
small updates for Simmate compatability
