+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -3,6 +3,7 @@
 import json
 import logging
 import math
+import os
 import warnings
 from pathlib import Path
 from typing import Literal, TypeVar
@@ -567,6 +568,7 @@ class ElfLabeler:
                     neigh_indices,
                     neigh_coords,
                     radii,
+                    dists,
                     bond_types,
                     plane_points,
                     plane_vectors,
@@ -577,6 +579,7 @@ class ElfLabeler:
                     site_indices,
                     neigh_indices,
                     neigh_coords,
+                    dists,
                 )
                 self._atom_nn_planes_e = (plane_points, plane_vectors)
 
@@ -619,7 +622,7 @@ class ElfLabeler:
                     node.coord_number == 1
                     and node.feature_type not in FeatureType.bare_types
                 ):
-                    atom_features[node.coord_electride_indices[0]].append(feat_idx)
+                    atom_features[node.coord_indices_e[0]].append(feat_idx)
             # add electride indices
             electride_num = 0
             for i, node in enumerate(self.bifurcation_graph.irreducible_nodes):
@@ -784,7 +787,7 @@ class ElfLabeler:
             The atoms coordinated with each feature.
 
         """
-        return self._get_feature_properties("coord_atom_indices")
+        return self._get_feature_properties("coord_indices")
 
     @property
     def feature_coord_nums(self) -> list:
@@ -799,7 +802,7 @@ class ElfLabeler:
         return np.array(self._get_feature_properties("coord_number"), dtype=np.int64)
 
     @property
-    def feature_coord_atom_dists(self) -> list:
+    def feature_coord_dists(self) -> list:
         """
 
         Returns
@@ -808,7 +811,7 @@ class ElfLabeler:
             The distance to each coordinated atom from each feature.
 
         """
-        return self._get_feature_properties("coord_atom_dists")
+        return self._get_feature_properties("coord_dists")
 
     @property
     def feature_coord_atoms_e(self) -> list:
@@ -820,7 +823,7 @@ class ElfLabeler:
             The coordinated atoms to each feature, including electrides as quasi-atoms
 
         """
-        return self._get_feature_properties("coord_electride_indices")
+        return self._get_feature_properties("coord_indices_e")
 
     @property
     def feature_coord_nums_e(self) -> list:
@@ -833,9 +836,7 @@ class ElfLabeler:
             electrides as quasi-atoms.
 
         """
-        return np.array(
-            self._get_feature_properties("electride_coord_number"), dtype=np.int64
-        )
+        return np.array(self._get_feature_properties("coord_number_e"), dtype=np.int64)
 
     @property
     def feature_coord_atoms_dists_e(self):
@@ -847,7 +848,7 @@ class ElfLabeler:
             The distance to each coordinated atom/electride from each feature.
 
         """
-        return self._get_feature_properties("coord_electride_dists")
+        return self._get_feature_properties("coord_dists_e")
 
     @property
     def feature_min_surface_dists(self) -> list:
@@ -1047,7 +1048,7 @@ class ElfLabeler:
                     f"No neighboring atoms found for feature with index {feature_idx}. Feature assigned to nearest atom."
                 )
                 # assign all charge/volume to the closest atom
-                nearest = np.argmin(self.feature_coord_atom_dists[feature_idx])
+                nearest = np.argmin(self.feature_coord_dists[feature_idx])
                 atom_charge[nearest] += charge
                 atom_volume[nearest] += volume
 
@@ -1078,7 +1079,7 @@ class ElfLabeler:
                 if use_electrides:
                     dists = self.feature_coord_atoms_dists_e[feature_idx].copy()
                 else:
-                    dists = self.feature_coord_atom_dists[feature_idx].copy()
[…]
