### 2026-01-08 9a39d85
remove windows/mac from lint tests
