### 2026-01-08 36fa25e
add small tolerance to weight label assignments
The mac os tests are giving different results only for the average surface distances which suggests to me the labels are being assigned improperly. I've added a tolerance to hopefully fix this.
