### 2026-01-08 b2da150
add more detailed test
The previous effort didn't work, so I've added additional asserts to see if I can get at the issue.
