### 2026-01-08 5afa60d
small bugfix in weight method
