### 2026-01-08 919a209
small weight tweaks
