### 2026-01-08 ac35f1f
tweak weight method
