### 2026-01-08 59ec9b7
add badelf and elf-labeler :carrot:  (#42)
Adds BadELF and ElfLabeler functionality and documentation.
