+++ src/baderkit/core/labelers/bifurcation_graph/graph_numba/bifurcation_base.py
@@ -1,0 +1,539 @@
+# -*- coding: utf-8 -*-
+import numpy as np
+from numba import njit, prange
+
+from baderkit.core.utilities.basic import coords_to_flat, wrap_point, wrap_point_w_shift
+from baderkit.core.utilities.union_find import (
+    find_root_no_compression,
+    find_root_with_shift,
+    union_w_roots,
+    union_with_shift,
+)
+
+###############################################################################
+# Ring/Cage Identification
+###############################################################################
+
+
+@njit(cache=True)
+def get_pt_ring_cage(frac_coords, tol=1e-3):
+    """
+    Checks if a set of points is similar to a point, ring, or cage. This is a very
+    rough estimate made only from the maxima. A more rigorous method might incorporate
+    voxels slightly below these values.
+    """
+
+    # if we only have 1 or 2 points we cannot make a ring or cage
+    if len(frac_coords) < 3:
+        return 0
+    # if we have exactly 3 we will always have a ring
+    elif len(frac_coords) == 3:
+        return 1
+
+    # reference coord used for unwrapping
+    ref0 = 0.0
+    ref1 = 0.0
+    ref2 = 0.0
+    ref_set = False
+
+    # create an array to store unwrapped coords
+    unwrapped_frac_coords = np.empty_like(frac_coords, dtype=np.float64)
+
+    # scan all maxima and pick those that belong to this target_group
+    for idx, (c0, c1, c2) in enumerate(frac_coords):
+
+        # first seen -> set reference for unwrapping
+        if not ref_set:
+            ref0, ref1, ref2 = c0, c1, c2
+            ref_set = True
+
+        # unwrap coordinate relative to reference: unwrapped = coord - round(coord - ref)
+        # Using np.round via float -> use built-in round for numba compatibility
+        # but call round(x) (returns float)
+        unwrapped_frac_coords[idx, 0] = c0 - round(c0 - ref0)
+        unwrapped_frac_coords[idx, 1] = c1 - round(c1 - ref1)
+        unwrapped_frac_coords[idx, 2] = c2 - round(c2 - ref2)
+
+    # get average of points
+    avg_pt = np.zeros(3, dtype=np.float64)
+    for frac_coord in unwrapped_frac_coords:
+        avg_pt += frac_coord
+    avg_pt /= len(unwrapped_frac_coords)
+
+    # Center points
+    pts = unwrapped_frac_coords - avg_pt
+    # Singular value decomposition (variance across each direction)
+    _, s, vh = np.linalg.svd(pts, full_matrices=False)
+    # Smallest singular value gives out-of-plane deviation
+    deviation = s[-1] / s[0]
+    if deviation < tol:
+        # we are fairly planar and consider this to be a ring
+        return 1
+    else:
+        # we are not planar and consider this a cage
+        return 2
+
+
+###############################################################################
+# Saddle Point Identification
+###############################################################################
+
+
+@njit(cache=True)
+def trans_to_idx(i, j, k, size):
+    val_range = size * 2 + 1
+    return (i + size) * (val_range**2) + (j + size) * val_range + (k + size)
+
+
+@njit(cache=True, inline="always")
+def shift_to_index(cx, cy, cz):
+    # each value can range from -2 to 2, essentially giving us a base 5 system
+    index = (cx + 2) * 25 + (cy + 2) * 5 + (cz + 2)
+    return index
+
+
+@njit(cache=True, inline="always")
+def index_to_shift(index):
+    cx = index // 25 - 2
+    cy = (index % 25) // 5 - 2
+    cz = index % 5 - 2
+    return cx, cy, cz
+
+
+@njit(cache=True, inline="always")
+def get_connections_in_box(size):
+    shifts = []
[…]
