### 2026-01-12 de4f3c5
fix radius finder bugs :potato: (#43)
Fixes several bugs in the ELF radii finder
