### 2026-01-12 9423b16
add vacuum to spin badelf
