### 2026-01-19 e2deb5b
update to new JOSS standards
