### 2026-01-19 #9430 @SWeav02
@zhubonan
I've updated the paper and pushed to the dedicated `paper` branch on the repo.
Thanks!
