### 2026-01-21 abf9198
make shallow cutoffs optional
