### 2026-01-21 0ba2a91
Add vacuum tolerance for BadELF :dash: (#44)
Adds option to mask vacuum regions in the ELF using a cutoff for the charge density
