### 2026-01-21 259f99f
fix bug in Bader json writeout
