### 2026-01-21 c7ec024
require non-negative total charge
