+++ src/baderkit/core/labelers/elf_labeler/elf_labeler.py
@@ -51,8 +51,11 @@ class ElfLabeler:
     ----------
     charge_grid : Grid
         The charge density grid used for integrating charge.
-    reference_grid : Grid
+    elf_grid : Grid
         The ELF grid used to partition volumes.
+    total_charge_grid : Grid, optional
+        The total charge density used for bader integrations and vacuum masks. If
+        not provided, the charge_grid will be used instead.
     ignore_low_pseudopotentials : bool, optional
         Whether or not to ignore errors associated with the use of pseudopotentials
         with limited valence electrons. The default is False.
@@ -162,7 +165,8 @@ class ElfLabeler:
     def __init__(
         self,
         charge_grid: Grid,
-        reference_grid: Grid,
+        elf_grid: Grid,
+        total_charge_grid: Grid | None = None,
         ignore_low_pseudopotentials: bool = False,
         shared_shell_ratio: float = 0.75,
         covalent_molecule_ratio: float = 0.2,
@@ -186,14 +190,18 @@ class ElfLabeler:
         **kwargs,
     ):
 
-        # ensure the reference file is ELF
-        if reference_grid.data_type != "elf":
+        # ensure the elf file is ELF
+        if elf_grid.data_type != "elf":
             logging.warning(
-                "A non-ELF reference file has been detected. Results may not be valid."
+                "A non-ELF elf file has been detected. Results may not be valid."
             )
 
         self.charge_grid = charge_grid
-        self.reference_grid = reference_grid
+        self.elf_grid = elf_grid
+        
+        if total_charge_grid is None:
+            total_charge_grid = charge_grid
+        self.total_charge_grid = total_charge_grid
 
         self.ignore_low_pseudopotentials = ignore_low_pseudopotentials
         self.crystalnn_kwargs = crystalnn_kwargs
@@ -248,12 +256,21 @@ class ElfLabeler:
         self._electrides_per_formula = None
         self._electrides_per_reduced_formula = None
 
-        # create a bader object
-        self.bader = Bader(
+        # create a bader object for the elf
+        self.elf_bader = Bader(
             charge_grid=charge_grid,
-            reference_grid=reference_grid,
+            reference_grid=elf_grid,
             vacuum_tol=vacuum_tol,
             use_reference_vacuum=False,
+            **kwargs,
+        )
+        
+        # create a bader object for the charge density
+        self.elf_bader = Bader(
+            charge_grid=charge_grid,
+            reference_grid=total_charge_grid,
+            vacuum_tol=vacuum_tol,
+            use_reference_vacuum=True,
             **kwargs,
         )
 
@@ -305,7 +322,7 @@ class ElfLabeler:
             The PyMatGen Structure representing the system.
 
         """
-        return self.reference_grid.structure
+        return self.elf_grid.structure
 
     @property
     def electride_structure(self) -> Structure:
@@ -682,7 +699,7 @@ class ElfLabeler:
             The charge assigned to the vacuum.
 
         """
-        return self.bader.vacuum_charge
+        return self.elf_bader.vacuum_charge
 
     @property
     def vacuum_volume(self) -> float:
@@ -694,7 +711,7 @@ class ElfLabeler:
             The total volume assigned to the vacuum.
 
         """
-        return self.bader.vacuum_volume
+        return self.elf_bader.vacuum_volume
 
     @property
     def vacuum_mask(self) -> NDArray[bool]:
@@ -706,7 +723,7 @@ class ElfLabeler:
             A mask representing the voxels that belong to the vacuum.
 
         """
-        return self.bader.vacuum_mask
+        return self.elf_bader.vacuum_mask
 
     @property
     def num_vacuum(self) -> int:
@@ -718,7 +735,7 @@ class ElfLabeler:
[…]
