+++ src/baderkit/core/bader/bader.py
@@ -35,34 +35,36 @@ class Bader:
     Parameters
     ----------
     charge_grid : Grid
-        A Grid object with the charge density that will be integrated.
-    reference_grid : Grid | None
-        A Grid object whose values will be used to construct the basins. If
-        None, defaults to the charge_grid.
+        The Grid object with the charge density that will be integrated.
+    total_charge_grid : Grid | None, optional
+        The Grid object used for determining vacuum regions in the system. For
+        pseudopotential codes this represents the total electron density and should
+        be provided whenever possible. If None, defaults to the charge_grid.
+    reference_grid : Grid | None, optional
+        The Grid object whose values will be used to construct the basins. This
+        should typically only be set when partitioning functions other than the 
+        charge density (e.g. ELI-D, ELF, etc.).If None, defaults to the 
+        total_charge_grid.
     method : str | Method, optional
         The algorithm to use for generating bader basins.
     vacuum_tol : float | bool, optional
         If a float is provided, this is the value below which a point will
         be considered part of the vacuum. If a bool is provided, no vacuum
-        will be used on False, and the default tolerance will be used on True.
+        will be used on False, and the default tolerance (0.001) will be used on True.
     basin_tol : float, optional
-        The value below which a basin will not be considered significant. This
-        is used only used to avoid writing out data that is likely not valuable.
-    use_reference_vacuum : bool, optional
-        Whether or not to use the reference file to determine regions of vacuum
-        (low charge density). This should generally be set to True unless the
-        ELF is being used as the reference. The default is True.
+        The total charge value below which a basin will not be considered significant.
+        This is used only in write methods to avoid writing insignicant regions.
 
     """
 
     def __init__(
         self,
         charge_grid: Grid,
+        total_charge_grid: Grid | None = None,
         reference_grid: Grid | None = None,
         method: str | Method = Method.weight,
         vacuum_tol: float | bool = 1.0e-3,
         basin_tol: float = 1.0e-3,
-        use_reference_vacuum: bool = True,
         **kwargs,
     ):
 
@@ -78,13 +80,25 @@ class Bader:
             )
 
         self._charge_grid = charge_grid
-
-        # if no reference is provided, use the base charge grid
+        
+        # if no total charge is provided, use the base charge grid
+        if total_charge_grid is None:
+            total_charge_grid = charge_grid
+        
+        # if no reference grid is provided, use the total charge grid
         if reference_grid is None:
-            reference_grid = charge_grid.copy()
+            reference_grid = total_charge_grid
+            
+        # ensure all grids have the same shape
+        for x in range(3):
+            assert charge_grid.shape[x] == total_charge_grid.shape[x] == reference_grid.shape[x], "Differing grid sizes found. All grids must have the same shape."
+
+        self._total_charge_grid = total_charge_grid
         self._reference_grid = reference_grid
 
-        self._use_reference_vacuum = use_reference_vacuum
+        # check if the total charge grid has values below 0
+        if vacuum_tol is not False:
+            assert total_charge_grid.total.min() >= 0, "The charge grid used to detect vacuum has values below 0. This typically results from too low of a grid density and causes incorrect partitions."
 
         # if vacuum tolerance is True, set it to the same default as above
         if vacuum_tol is True:
@@ -192,6 +206,23 @@ class Bader:
     def charge_grid(self, value: Grid):
         self._charge_grid = value
         self._reset_properties()
+        
+    @property
+    def total_charge_grid(self) -> Grid:
+        """
+
+        Returns
+        -------
+        Grid
+            A Grid object whose values are used to determine vacuum regions.
+
+        """
+        return self._total_charge_grid
+    
+    @total_charge_grid.setter
[…]
