### 2026-01-22 2893a63
add surface area estimate
