+++ src/baderkit/core/__init__.py
@@ -4,6 +4,4 @@ from .toolkit import Grid, Structure  # isort:skip
 
 from .bader import Bader  # isort:skip
 
-from .labelers import ElfLabeler, SpinElfLabeler  # isort:skip
-
-from .badelf import Badelf, SpinBadelf  # isort:skip
+from .elf_analysis import BifurcationGraph, ElfLabeler, SpinElfLabeler, Badelf, SpinBadelf  # isort:skip

[…]
