+++ src/baderkit/core/base/base_analysis.py
@@ -1,0 +1,563 @@
+# -*- coding: utf-8 -*-
+
+import copy
+import json
+from pathlib import Path
+from abc import ABC, abstractclassmethod
+from typing import Literal, TypeVar
+import logging
+
+import numpy as np
+from numpy.typing import NDArray
+
+from baderkit.core.toolkit import Grid, Structure
+from baderkit.core.utilities.file_parsers import Format
+
+# # This allows for Self typing and is compatible with python 3.10
+Self = TypeVar("Self", bound="BaseAnalysis")
+
+# TODO:
+# - Add handling of non-nuclear attractors (e.g. those in Li metal)
+
+
+class BaseAnalysis(ABC):
+    """
+    This is the base class that most analysis classes pull from. It covers the absolute
+    basic methods and ensures similar structure throughout the package.
+
+    Parameters
+    ----------
+    charge_grid : Grid
+        The Grid object with the charge density that will be integrated.
+    total_charge_grid : Grid | None, optional
+        The Grid object used for determining vacuum regions in the system. For
+        pseudopotential codes this represents the total electron density and should
+        be provided whenever possible. If None, defaults to the charge_grid.
+    reference_grid : Grid | None, optional
+        The Grid object whose values will be used to construct the basins. This
+        should typically only be set when partitioning functions other than the 
+        charge density (e.g. ELI-D, ELF, etc.).If None, defaults to the 
+        total_charge_grid.
+    vacuum_tol : float | bool, optional
+        If a float is provided, this is the value below which a point will
+        be considered part of the vacuum. If a bool is provided, no vacuum
+        will be used on False, and the default tolerance (0.001) will be used on True.
+
+    """
+    
+    _base_reset_props = [
+        "vacuum_mask",
+        "num_vacuum",
+        "vacuum_charge",
+        "vacuum_volume",
+        "structure"
+        ]
+    _reset_props = []
+
+    def __init__(
+        self,
+        charge_grid: Grid,
+        total_charge_grid: Grid | None = None,
+        reference_grid: Grid | None = None,
+        vacuum_tol: float | bool = 1.0e-3,
+        **kwargs,
+    ):
+
+        self._charge_grid = charge_grid
+        
+        # if no total charge is provided, use the base charge grid
+        if total_charge_grid is None:
+            total_charge_grid = charge_grid
+        
+        # if no reference grid is provided, use the total charge grid
+        if reference_grid is None:
+            reference_grid = total_charge_grid
+            
+        # ensure all grids have the same shape
+        for x in range(3):
+            assert charge_grid.shape[x] == total_charge_grid.shape[x] == reference_grid.shape[x], "Differing grid sizes found. All grids must have the same shape."
+
+        self._total_charge_grid = total_charge_grid
+        self._reference_grid = reference_grid
+
+        # check if the total charge grid has values below 0
+        if vacuum_tol is not False:
+            assert total_charge_grid.total.min() >= 0, "The charge grid used to detect vacuum has values below 0. This typically results from too low of a grid density and causes incorrect partitions."
+
+        # if vacuum tolerance is True, set it to the same default as above
+        if vacuum_tol is True:
+            self._vacuum_tol = 1.0e-3
+        else:
+            self._vacuum_tol = vacuum_tol
+
+        # set hidden class variables. This allows us to cache properties and
+        # still be able to recalculate them if needed, though that should only
+        # be done by advanced users
+        self._reset_properties()
+
+    ###########################################################################
+    # Set Properties
+    ###########################################################################
+
+    def _reset_properties(
+        self,
+        include_properties: list[str] = None,
+        exclude_properties: list[str] = [],
+    ):
+        if include_properties is None:
+            include_properties = self._reset_props + self._base_reset_props
+        # get our final list of properties
+        reset_properties = [
[…]
