### 2026-01-28 4da8cf1
critical point ideas
