+++ src/baderkit/core/critical_points/critical_points_numba1.py
@@ -1,0 +1,828 @@
+from numba import njit, prange
+import numpy as np
+
+def get_facet_transforms(parity):
+    facet_transforms = []
+    opposites = []
+    for i, p in enumerate(parity):
+        trans = np.zeros(3, dtype=np.int8)
+        if p == 1:
+            trans[i] = -1
+            facet_transforms.append(trans.copy())
+            opposites.append(1)
+            
+            trans[i] = 1
+            facet_transforms.append(trans.copy())
+            opposites.append(-1)
+    return facet_transforms, opposites
+
+def get_cofacet_transforms(parity):
+    cofacet_transforms = []
+    for i, p in enumerate(parity):
+        trans = np.zeros(3, dtype=np.int8)
+        if p == 0:
+            trans[i] = -1
+            cofacet_transforms.append(trans.copy())
+            trans[i] = 1
+            cofacet_transforms.append(trans.copy())
+    # add 0,0,0 for 3d parity
+    if not cofacet_transforms:
+        cofacet_transforms.append(np.zeros(3, dtype=np.int8))
+    return cofacet_transforms
+
+def get_trans_maps():
+    trans_to_int = np.zeros((3,3,3), dtype=np.int8)
+    int_to_trans = np.zeros((6,3), dtype=np.int8)
+    for i in range(6):
+        j = i // 2
+        value = (-1) ** (i & 1) # -1 for even, 1 for odd
+        int_to_trans[i, j] = value
+        x, y, z = [0 if k != j else value for k in (0, 1, 2)]
+        trans_to_int[x,y,z] = i
+    return trans_to_int, int_to_trans
+
+def get_parity_maps():
+    parity_to_int = np.empty((3,3,3), dtype=np.int8)
+    int_to_parity = np.empty((8,3), dtype=np.int8)
+    parities = [
+        [0,0,0], # 0D
+        [0,0,1], # 1D
+        [0,1,0], # 1D
+        [0,1,1], # 2D
+        [1,0,0], # 1D
+        [1,0,1], # 2D
+        [1,1,0], # 2D
+        [1,1,1], # 3D
+        ]
+    parity_dims = np.array([0,1,1,2,1,2,2,3], dtype=np.int8)
+    for parity_idx, (x,y,z) in enumerate(parities):
+        parity_to_int[x,y,z] = parity_idx
+        int_to_parity[parity_idx] = [x,y,z]
+    return parity_to_int, int_to_parity, parity_dims
+
+def get_pointer_maps():
+    """
+    Assumes:
+        bit 7 (value 128): traversal flag (0/1)
+        bit 6 (value 64): crit (0/1)
+        bits 3-5 (8,16,32): face (0..7)
+        bits 0-2 (1,2,4): cofacet (0..7)
+    """
+    pointers_to_int = np.zeros((2, 8, 8), dtype=np.uint8)
+    int_to_pointers = np.zeros((255, 3), dtype=np.uint8)
+    
+    count = 0
+    for crit in range(2):
+        for face in range(8):
+            for cofacet in range(8):
+                pointers_to_int[crit, face, cofacet] = count
+                int_to_pointers[count] = [crit, face, cofacet]
+                count += 1
+
+    return pointers_to_int, int_to_pointers
+
+def get_cofacet_map():
+    
+    parities = np.array([
+        [0,0,0],  # 0D
+        [0,0,1],  # 1D
+        [0,1,0],  # 1D
+        [0,1,1],  # 2D
+        [1,0,0],  # 1D
+        [1,0,1],  # 2D
+        [1,1,0],  # 2D
+        [1,1,1],  # 3D
+    ])
+    parity_cofacet_trans = []
+    for parity in parities:
+        facet_transforms = get_cofacet_transforms(parity)
[…]

[…]
