### 2026-01-29 845f932
working persistence reduction
