+++ src/baderkit/core/critical_points/bader_based.py
@@ -4,372 +4,60 @@ from baderkit.core import Bader
 from baderkit.core.utilities.basic import wrap_point
 import numpy as np
 from numpy.typing import NDArray
+from numba import njit, prange
 
-PARITIES = np.array([
-    [0,0,0], # 0D
-    [0,0,1], # 1D
-    [0,1,0], # 1D
-    [0,1,1], # 2D
-    [1,0,0], # 1D
-    [1,0,1], # 2D
-    [1,1,0], # 2D
-    [1,1,1], # 3D
-    ], dtype=np.uint8)
-
-PARITY_DIMS = np.array([0,1,1,2,1,2,2,3], dtype=np.uint8)
-PARITY_TO_INT = np.zeros((2,2,2), dtype=np.int8)
-for idx, (i,j,k ) in enumerate(PARITIES):
-    PARITY_TO_INT[i,j,k] = idx
-
-CUBE_VERTICES = [
-    [-1,-1,-1],
-    [1,-1,-1],
-    [1,1,-1],
-    [-1,1,-1],
-    [-1,-1,1],
-    [1,-1,1],
-    [1,1,1],
-    [-1,1,1]
-    ]
-POLY_VERTICES = [
-    [-1,-1],
-    [1,-1],
-    [1,1],
-    [-1,1],
-    ]
-EDGE_VERTICES = [
-    [-1],
-    [1],
-    ]
-VERTEX_VERTICES = [[]]
-VERTICES = [CUBE_VERTICES, POLY_VERTICES, EDGE_VERTICES, VERTEX_VERTICES]
-
-CUBE_TO_POLY = np.array([
-    [0, 1, 2, 3],
-    [4, 5, 6, 7],
-    [0, 1, 5, 4],
-    [3, 2, 6, 7],
-    [0, 4, 7, 3],
-    [1, 5, 6, 2],
-    ],dtype=np.uint8)
-POLY_TO_EDGE = np.array([
-    [0,1],
-    [1,2],
-    [2,3],
-    [3,0],
-    ],dtype=np.uint8)
-EDGE_TO_VERTEX = np.array([
-    [0],
-    [1],
-    ],dtype=np.uint8)
-empty = np.array([[]], dtype=np.uint8)
-# queried by number of vertices
-FACET_MAPS = (
-    empty,
-    empty, 
-    EDGE_TO_VERTEX, 
-    empty, 
-    POLY_TO_EDGE, 
-    empty, 
-    empty, 
-    empty, 
-    CUBE_TO_POLY,
-    )
-
-def parity_to_cofacets(parity: NDArray[int]) -> NDArray[int]:
+@njit
+def szudzik_pair(a, b):
     """
-
-    Parameters
-    ----------
-    parity : NDArray[int]
-        The parity of a cell.
-
-    Returns
-    -------
-    NDArray[int]
-        A Nx3 array where each row is the relative transform from a cell with
-        the input parity to each of its cofacets.
-
+    Szudzik pairing for uint16 inputs.
+    Returns uint32.
     """
-    # get where the parity is 1
-    zero = np.where(parity==0)[0]
-    # get transformations
-    transforms = []
-    for i in zero:
-        transform = np.zeros(3, dtype=np.int8)
-        for j in range(3):
-            if j == i:
-                transform[j] = 1
-                transforms.append(transform.copy())
-                transform[j] = -1
-                transforms.append(transform.copy())
-    # for 3D parity, we add a blank list to ensure array has the same typing
-    if len(transforms) == 0:
-        return np.array([[]], dtype=np.int8)
-    return np.array(transforms, dtype=np.int8)
+    a = np.uint16(a)
+    b = np.uint16(b)
+    if a >= b:
+        return a * a + a + b
+    else:
+        return b * b + a
 
[…]
