### 2026-01-30 05bf989
working ongrid/neargrid periodic awareness
