+++ src/baderkit/core/bader/methods/shared_numba.py
@@ -9,7 +9,6 @@ from numpy.typing import NDArray
 from baderkit.core.utilities.basic import (
     coords_to_flat,
     flat_to_coords,
-    merge_frac_coords_weighted,
     wrap_point,
     wrap_point_w_shift,
 )
@@ -139,6 +138,10 @@ def get_best_neighbor_with_shift(
     bni = i
     bnj = j
     bnk = k
+    # create initial best shift
+    bsi = 0
+    bsj = 0
+    bsk = 0
     # For each neighbor get the difference in value and if its better
     # than any previous, replace the current best
     for (si, sj, sk), dist in zip(neighbor_transforms, neighbor_dists):
@@ -343,13 +346,14 @@ def get_basin_charges_and_volumes(
     volumes = np.zeros(maxima_num, dtype=np.float64)
     vacuum_charge = 0.0
     vacuum_volume = 0.0
+    vacuum_val = maxima_num
     # iterate in parallel over each voxel
     for i in range(nx):
         for j in range(ny):
             for k in range(nz):
                 charge = data[i, j, k]
                 label = labels[i, j, k]
-                if label < 0:
+                if label == vacuum_val:
                     vacuum_charge += charge
                     vacuum_volume += 1
                 else:
@@ -362,164 +366,6 @@ def get_basin_charges_and_volumes(
     vacuum_charge = vacuum_charge / total_points
     return charges, volumes, vacuum_charge, vacuum_volume
 
-# @njit(parallel=True, cache=True)
-# def perturb_same_value_adjacencies(
-#     labels: NDArray[int],
-#     data: NDArray[np.float64],
-#     neighbor_transforms: NDArray[np.int64],
-#     vacuum_mask: NDArray[np.bool_],
-#         ):
-#     nx, ny, nz = data.shape
-#     ny_nz = ny*nz
-#     # create a mask for the location of same valued connections
-#     same_values = np.zeros_like(labels, dtype=np.bool_)
-#     # loop over each voxel in parallel
-#     for i in prange(nx):
-#         for j in range(ny):
-#             for k in range(nz):
-#                 # if this voxel is part of the vacuum, continue
-#                 if vacuum_mask[i, j, k]:
-#                     continue
-#                 # get this voxels value
-#                 value = data[i, j, k]
-#                 # iterate over the neighboring voxels
-#                 for si, sj, sk in neighbor_transforms:
-#                     # wrap points
-#                     ii, jj, kk = wrap_point(i + si, j + sj, k + sk, nx, ny, nz)
-#                     if data[ii,jj,kk] == value:
-#                         idx = coords_to_flat(i,j,k, ny_nz, nz)
-#                         same_values[idx] = True
-#                         break
-#     breakpoint()
-#     same_indices = np.where(same_values)[0]
-#     # initialize indices
-#     labels[same_indices] = same_indices
-#     # create unions and find roots
-#     for idx in same_indices:
-#         i, j, k = flat_to_coords(idx, ny_nz, nz)
-#         for si, sj, sk in neighbor_transforms:
-#             # wrap points
-#             ii, jj, kk = wrap_point(i + si, j + sj, k + sk, nx, ny, nz)
-#             if data[ii,jj,kk] == value:
-#                 # union
-#                 neigh_idx = coords_to_flat(ii, jj, kk, ny_nz, nz)
-#                 union(labels, idx, neigh_idx)
-#                 break
-    
-#     # get roots
-#     roots = np.array(len(same_indices), dtype=np.int64)
-#     for sidx in prange(len(same_indices)):
-#         roots[sidx] = find_root_no_compression(labels, sidx)
-#     breakpoint()
-#     # get the unique roots
-#     unique_roots = np.unique(roots)
-    
[…]
