+++ joss_paper/paper.bib
@@ -35,6 +35,19 @@
     url={https://doi.org/10.1063/1.3553716}
 }
 
+@article{Otero-de-la-Roza:2014,
+    title={Critic2: A program for real-space analysis of quantum chemical interactions in solids}, 
+    volume={185}, 
+    DOI={10.1016/j.cpc.2013.10.026}, 
+    number={3}, 
+    journal={Computer Physics Communications}, 
+    author={Otero-de-la-Roza, A and Johnson, Erin R and Luaña, V}, 
+    year={2014}, 
+    month={May}, 
+    pages={1007-1018},
+    url={https://doi.org/10.1016/j.cpc.2013.10.026}
+}
+
 @misc{pybader,
     author = {A. M. Kerrigan},
     title = {pybader},
