+++ joss_paper/paper.md
@@ -27,15 +27,20 @@ bibliography: paper.bib
 
 # Summary
 
-The concept of oxidation states has existed for centuries, guiding and informing the decisions of generations of scientists. However, oxidation states are not observable, and cannot be uniquely derived from first principle calculations. This has led researchers to develop a variety of methods to recover oxidation states, each with their own unique theory and methodology. Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules, which derives oxidation states directly from a systems electron charge density. Though there are several popular implementations of this method, they are either outdated or slow. The `BaderKit` package brings Bader charge analysis into the modern Python ecosystem, reworking the most popular grid-based algorithms to run in parallel.
+The concept of oxidation states has existed for centuries, guiding and informing the decisions of generations of scientists. However, oxidation states are not observable, and cannot be uniquely derived from first principle calculations. This has led researchers to develop a variety of methods to recover oxidation states, each with their own unique theory and methodology. Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules, which derives oxidation states directly from a systems electron charge density. The `BaderKit` package brings Bader charge analysis into the modern Python ecosystem, and reworks the most popular grid-based algorithms to run in parallel.
 
-# Statement of need
+# Statement of Need
 
-Bader charge analysis is among the most widely used methods in chemistry and materials science, with thousands of articles referencing the method each year. By far the most popular implementation is the code developed by the Henkelman group at UT Austin [@Henkelman:2006]. While fast and memory-efficient, the code is written in Fortran making it cumbersome to incorporate into modern workflows. It is well over a decade old and was not designed for modern multi-threaded CPUs. There have been some previous attempts to alleviate these issues. In particular, `pybader` [@pybader] partially implemented the method into parallelized Python code. However, `pybader` is typically slower than the original code and requires a significant amount of boilerplate code. `BaderKit` aims to improve upon the areas where the original Fortran code and `pybader` fall short. It is fast, extensive, easy to use, and designed to be easily inserted into modern workflows.
+Bader Quantum Theory of Atoms in Molecules (QTAIM) charge analysis is among the most widely used methods in chemistry and materials science, with thousands of articles referencing the method each year. This popularity has given rise to many packages for performing QTAIM analysis. Despite the variety of implementations, none are well equipped for modern high-throughput workflows. Most are written in Fortran making them cumbersome to automate and adapt for purposes outside their original scope. The algorithms implemented in these codes are serial and do not utilize modern multi-core CPUs. `BaderKit` aims to resolve these issues, providing a fast, parallelized, and easily extended QTAIM implementation written in Python.
+
+
+# State of the Field
+
[…]
