### 2026-01-30 #9430 @SWeav02
> Apologies [@SWeav02](https://github.com/SWeav02), we missed one section from the list above. In addition to the sections already mentioned, JOSS papers now also require:

I've updated the paper again and pushed to the `paper` branch. Thanks!
