### 2026-01-30 b9bece1
working neargrid-weight periodic
