### 2026-01-30 b3004e8
working weight periodic
