### 2026-02-02 cb114ce
add minima basin finder
