+++ src/baderkit/core/critical_points/bader_based.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from baderkit.core import Bader
-from baderkit.core.utilities.basic import wrap_point
+from baderkit.core.utilities.basic import wrap_point, wrap_point_w_shift
 import numpy as np
 from numpy.typing import NDArray
 from numba import njit, prange
@@ -220,35 +220,97 @@ def get_edges_thin(
 
     return edges
 
+@njit(parallel=True)
+def get_edges_periodic(
+    labeled_array: NDArray[np.int64],
+    images: NDArray[np.int64],
+    neighbor_transforms: NDArray[np.int64],
+    vacuum_mask: NDArray[np.bool_],
+):
+    """
+    In a 3D array of labeled voxels, finds the voxels that neighbor at
+    least one voxel with a different label.
 
-bader_up = Bader.from_vasp("CHGCAR")
+    Parameters
+    ----------
+    labeled_array : NDArray[np.int64]
+        A 3D array where each entry represents the basin label of the point.
+    neighbor_transforms : NDArray[np.int64]
+        The transformations from each voxel to its neighbors.
+    vacuum_mask : NDArray[np.bool_]
+        A 3D array representing the location of the vacuum
+
+    Returns
+    -------
+    edges : NDArray[np.bool_]
+        A mask with the same shape as the input grid that is True at points
+        on basin edges.
+
+    """
+    nx, ny, nz = labeled_array.shape
+
+    # create 3D array to store edges
+    edges = np.zeros_like(labeled_array, dtype=np.bool_)
+    # loop over each voxel in parallel
+    for i in prange(nx):
+        for j in range(ny):
+            for k in range(nz):
+                # if this voxel is part of the vacuum, continue
+                if vacuum_mask[i, j, k]:
+                    continue
+                # get this voxels label and image
+                label = labeled_array[i, j, k]
+                image = images[i, j, k]
+
+                # iterate over the neighboring voxels
+                for si, sj, sk in neighbor_transforms:
+                    # wrap points
+                    ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(i + si, j + sj, k + sk, nx, ny, nz)
+                    # skip vacuum
+                    if vacuum_mask[ii,jj,kk]:
+                        continue
+                    # get neighbors label and image
+                    neigh_label = labeled_array[ii, jj, kk]
+                    neigh_image = images[ii, jj, kk]
+                    # if the neighbor has a different label, this is an edge
+                    if neigh_label != label:
+                        edges[i,j,k] = True
+                        break
+                    # if the neighbor has a different image it may be an edge.
+                    # We don't count this if the neighbor wrapped around
+                    if image == neigh_image:
+                        continue
+                    if ssi != 0 or ssj != 0 or ssk != 0:
+                        continue
+                    edges[i,j,k] = True
+                    break
+
+    return edges
+
+
+bader = Bader.from_vasp("CHGCAR")
 
 # get basins
-labels = bader_up.basin_labels
+labels_up = bader.basin_labels
+images_up = bader.basin_images
 
-# get reverse
-reverse_grid = bader_up.reference_grid.copy()
-reverse_grid.total *= -1
-reverse_grid.total += bader_up.reference_grid.total.max() + bader_up.vacuum_tol
-reverse_grid.total[bader_up.vacuum_mask] = 0
[…]
