### 2026-02-02 2ebbd7d
update bader property names
