+++ src/baderkit/core/bader/methods/neargrid/neargrid_method.py
@@ -119,8 +119,8 @@ class NeargridMethod(MethodBase):
         images = images.reshape(shape)
         # get all results
         results = {
-            "basin_labels": labels,
-            "basin_images": images,
+            "maxima_basin_labels": labels,
+            "maxima_basin_images": images,
         }
         # assign charges/volumes, etc.
         logging.info("Assigning Charges and Volumes")
