+++ src/baderkit/core/critical_points/bader_based.py
@@ -5,6 +5,7 @@ from baderkit.core.utilities.basic import wrap_point, wrap_point_w_shift
 import numpy as np
 from numpy.typing import NDArray
 from numba import njit, prange
+import itertools
 
 @njit
 def szudzik_pair(a, b):
@@ -287,58 +288,575 @@ def get_edges_periodic(
 
     return edges
 
+PARITIES = np.array(
+    [
+        [0, 0, 0],  # 0D
+        [0, 0, 1],  # 1D
+        [0, 1, 0],  # 1D
+        [0, 1, 1],  # 2D
+        [1, 0, 0],  # 1D
+        [1, 0, 1],  # 2D
+        [1, 1, 0],  # 2D
+        [1, 1, 1],  # 3D
+    ],
+    dtype=np.uint8,
+)
+
+PARITY_DIMS = np.array([0, 1, 1, 2, 1, 2, 2, 3], dtype=np.uint8)
+
+# Conversion from parity to its index
+PARITY_TO_INT = np.zeros((2, 2, 2), dtype=np.int8)
+for idx, (i, j, k) in enumerate(PARITIES):
+    PARITY_TO_INT[i, j, k] = idx
+
+# Generate transformations to vertices for each parity type.
+CUBE_VERTICES = [
+    [-1, -1, -1],
+    [1, -1, -1],
+    [1, 1, -1],
+    [-1, 1, -1],
+    [-1, -1, 1],
+    [1, -1, 1],
+    [1, 1, 1],
+    [-1, 1, 1],
+]
+POLY_VERTICES = [
+    [-1, -1],
+    [1, -1],
+    [1, 1],
+    [-1, 1],
+]
+EDGE_VERTICES = [
+    [-1],
+    [1],
+]
+VERTEX_VERTICES = [[]]
+VERTICES = [CUBE_VERTICES, POLY_VERTICES, EDGE_VERTICES, VERTEX_VERTICES]
+def parity_to_vertices(parity: NDArray[int]) -> NDArray[int]:
+    """
+
+    Parameters
+    ----------
+    parity : NDArray[int]
+        The parity of a cell.
+
+    Returns
+    -------
+    NDArray[int]
+        A Nx3 array where each row is the relative transform from a cell with
+        the input parity to each of its vertices.
+
+    """
+    zero = np.where(parity == 0)[0]
+    vertex_type = len(zero)
+    base_transforms = VERTICES[vertex_type]
+    transforms = []
+    for trans in base_transforms:
+        new_trans = trans.copy()
+        for idx in zero:
+            new_trans.insert(idx, 0)
+        transforms.append(new_trans)
+    return np.array(transforms)
+
+
+def get_parity_vertices() -> list[NDArray]:
+    """
+
+    Returns
+    -------
+    transforms : list[NDArray]
+        The relative transforms for each parity that generate a corresponding
+        cells vertices.
+
+    """
+    transforms = []
+    for parity in PARITIES:
+        transforms.append(parity_to_vertices(parity))
+    return tuple(transforms)
+
+
+PARITY_VERTICES = get_parity_vertices()
+
+@njit(parallel=True)
+def get_edges_double_grid(
+    labeled_array: NDArray[np.int64],
+    images: NDArray[np.int64],
+    neighbor_transforms: NDArray[np.int64],
+    vacuum_mask: NDArray[np.bool_],
+):
+
+    nx, ny, nz = labeled_array.shape
+    nx2 = nx*2
+    ny2 = ny*2
+    nz2 = nz*2
+
+
+    # create 3D array to store edges. This will be twice the size of the original
+    # grids
+    edges = np.zeros((nx2, ny2, nz2), dtype=np.bool_)
+    # loop over each voxel in parallel
[…]
