+++ src/baderkit/core/bader/bader.py
@@ -65,14 +65,21 @@ class Bader(BaseAnalysis):
 
     _reset_props = [
         # assigned by _run_bader
-        "basin_labels",
-        "basin_images",
-        "basin_maxima_frac",
-        "basin_maxima_charge_values",
-        "basin_maxima_ref_values",
-        "basin_maxima_vox",
+        "maxima_basin_labels",
+        "maxima_basin_images",
+        "maxima_frac",
+        "maxima_vox",
+        "maxima_charge_values",
+        "maxima_ref_values",
         "basin_charges",
         "basin_volumes",
+        # assigned by _run_minima_bader
+        "minima_basin_labels",
+        "minima_basin_images",
+        "minima_frac",
+        "minima_vox",
+        "minima_charge_values",
+        "minima_ref_values",
         # Assigned by calling the property
         "basin_min_surface_distances",
         "basin_avg_surface_distances",
@@ -82,8 +89,6 @@ class Bader(BaseAnalysis):
         "basin_contact_surface_areas",
         "atom_surface_areas",
         "atom_contact_surface_areas",
-        "minima_basin_labels",
-        "minima_basin_images",
         # Assigned by run_atom_assignment
         "basin_atoms",
         "basin_atom_dists",
@@ -229,11 +234,11 @@ class Bader(BaseAnalysis):
             ])
 
     ###########################################################################
-    # Calculated Properties
+    # Maxima Basin Properties
     ###########################################################################
 
     @property
-    def basin_labels(self) -> NDArray[float]:
+    def maxima_basin_labels(self) -> NDArray[float]:
         """
 
         Returns
@@ -243,29 +248,12 @@ class Bader(BaseAnalysis):
             representing the basin the voxel belongs to.
 
         """
-        if self._basin_labels is None:
+        if self._maxima_basin_labels is None:
             self._run_bader()
-        return self._basin_labels
+        return self._maxima_basin_labels
     
     @property
-    def minima_basin_labels(self) -> NDArray[float]:
-        """
-
-        Returns
-        -------
-        NDArray[float]
-            The equivalent of bader basins for the desending gradient to local
-            minima. This is each minima's ascending manifold and can be used
-            in combination with the bader basins to locate important topological
-            features.
-
-        """
-        if self._minima_basin_labels is None:
-            self._run_minima_bader()
-        return self._minima_basin_labels
-    
-    @property
-    def basin_images(self) -> NDArray[int]:
+    def maxima_basin_images(self) -> NDArray[int]:
         """
 
         Returns
@@ -305,26 +293,12 @@ class Bader(BaseAnalysis):
                 26: [ 1,  1,  1]
 
         """
-        if self._basin_images is None:
+        if self._maxima_basin_images is None:
             self._run_bader()
-        return self._basin_images
+        return self._maxima_basin_images
     
     @property
-    def minima_basin_images(self) -> NDArray[float]:
-        """
-
-        Returns
-        -------
-        NDArray[float]
-            The equivalent of the basin_images property for minima basins.
-
-        """
-        if self._minima_basin_images is None:
-            self._run_minima_bader()
-        return self._minima_basin_images
-
-    @property
-    def basin_maxima_frac(self) -> NDArray[float]:
+    def maxima_frac(self) -> NDArray[float]:
         """
 
         Returns
@@ -333,12 +307,12 @@ class Bader(BaseAnalysis):
             The fractional coordinates of each attractor.
 
         """
-        if self._basin_maxima_frac is None:
+        if self._maxima_frac is None:
             self._run_bader()
-        return self._basin_maxima_frac
+        return self._maxima_frac
 
     @property
[…]
