### 2026-02-03 349f37c
initial working saddle finder
