+++ src/baderkit/core/critical_points/critical_points_numba.py
@@ -1,244 +1,170 @@
-from baderkit.core.utilities.basic import wrap_point_w_shift
+from baderkit.core.utilities.basic import wrap_point_w_shift, wrap_point, coords_to_flat
+import itertools
 import numpy as np
 from numpy.typing import NDArray
 from numba import njit, prange
 
-# Things to do:
-    # get grid with labels for each type of critical point and manifold
-    # Label ascending/descending manifolds
-    # create connection map between critical points
-    # determine what else is useful to make easily available. 
-        # - integral lines?
-        # - bader bonded atoms
-        # - etc.
+PARITIES = np.array(
+    [
+        [0, 0, 0],  # 3D
+        [0, 0, 1],  # 2D
+        [0, 1, 0],  # 2D
+        [0, 1, 1],  # 1D
+        [1, 0, 0],  # 2D
+        [1, 0, 1],  # 1D
+        [1, 1, 0],  # 1D
+        [1, 1, 1],  # 0D
+    ],
+    dtype=np.uint8,
+)
 
-@njit(parallel=True)
-def get_manifolds(
-    labeled_array: NDArray[np.int64],
-    images: NDArray[np.int64],
-    image_to_int: NDArray[np.int64],
-    int_to_image: NDArray[np.int64],
-    neighbor_transforms: NDArray[np.int64],
-    vacuum_mask: NDArray[np.bool_],
-):
+PARITY_DIMS = np.array([3,2,2,1,2,1,1,0], dtype=np.uint8)
+
+# Conversion from parity to its index
+PARITY_TO_INT = np.zeros((2, 2, 2), dtype=np.int8)
+for idx, (i, j, k) in enumerate(PARITIES):
+    PARITY_TO_INT[i, j, k] = idx
+
+###############################################################################
+# Voxel Transforms
+###############################################################################
+# Generate transformations to vertices for each parity type.
+VERTEX_VOXELS = [
+    [-1, -1, -1],
+    [1, -1, -1],
+    [1, 1, -1],
+    [-1, 1, -1],
+    [-1, -1, 1],
+    [1, -1, 1],
+    [1, 1, 1],
+    [-1, 1, 1],
+]
+EDGE_VOXELS = [
+    [-1, -1],
+    [1, -1],
+    [1, 1],
+    [-1, 1],
+]
+FACE_VOXELS = [
+    [-1],
+    [1],
+]
+VOXEL_VOXELS = [[]]
+VOXELS = [VERTEX_VOXELS, EDGE_VOXELS, FACE_VOXELS, VOXEL_VOXELS]
+
+def parity_to_voxels(parity: NDArray[int]) -> NDArray[int]:
     """
-    Finds the edges of maxima/minima 3-manifolds. Points that have 1 different
-    neighbor lie on the dividing surface of the 3-manifold and correspond to
-    2-manifolds. Points that have more than 1 different neighbor lie on the bounding
-    edges of these 2-manifolds making up 1-manifolds. The complete description
-    is as follows:
-        maxima 3-manifolds:
-            bounding 2-manifolds are descending connections from type-2 saddles
-            to type-1 saddles
-            bounding 1-manifolds are descending connections from type-1 saddles
-            to minima
-        minima 3-manifolds:
-            bounding 2-manifolds are ascending connections from type-1 saddles
-            to type-2 saddles
-            bounding 1-manifolds are ascending connections from type-2 saddles
-            to maxima
 
     Parameters
     ----------
-    labeled_array : NDArray[np.int64]
-        A 3D array where each entry represents the basin label of the point.
-    images : NDArray[np.int64]
-        A 3D array where each entry represents the periodic image of the maximum
-        or minimum this point belongs to
-    image_to_int : NDArray[np.int64]
[…]
