### 2026-02-03 a65bb6b
make persistence checks image aware
