+++ src/baderkit/core/bader/methods/shared_numba.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import math
+import itertools
 
 import numpy as np
 from numba import njit, prange
@@ -218,12 +219,85 @@ def get_edges(
                         break
     return edges
 
+IMAGE_TO_INT = np.empty([3,3,3], dtype=np.int64)
+INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
+for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
+    IMAGE_TO_INT[i,j,k] = shift_idx
+
+@njit(parallel=True, cache=True)
+def get_edges_w_images(
+    labeled_array: NDArray[np.int64],
+    images: NDArray[np.int64],
+    neighbor_transforms: NDArray[np.int64],
+    vacuum_mask: NDArray[np.bool_],
+):
+    """
+    In a 3D array of labeled voxels, finds the voxels that neighbor at
+    least one voxel with a different label.
+
+    Parameters
+    ----------
+    labeled_array : NDArray[np.int64]
+        A 3D array where each entry represents the basin label of the point.
+    images : NDArray[np.int64]
+        A 3D array where each entry represents the periodic image of the basin
+        the point belongs to
+    neighbor_transforms : NDArray[np.int64]
+        The transformations from each voxel to its neighbors.
+    vacuum_mask : NDArray[np.bool_]
+        A 3D array representing the location of the vacuum
+
+    Returns
+    -------
+    edges : NDArray[np.bool_]
+        A mask with the same shape as the input grid that is True at points
+        on basin edges.
+
+    """
+    nx, ny, nz = labeled_array.shape
+    # create 3D array to store edges
+    edges = np.zeros_like(labeled_array, dtype=np.bool_)
+    
+    # loop over each voxel in parallel
+    for i in prange(nx):
+        for j in range(ny):
+            for k in range(nz):
+                # if this voxel is part of the vacuum, continue
+                if vacuum_mask[i, j, k]:
+                    continue
+                # get this voxels label and image
+                label = labeled_array[i, j, k]
+                image = images[i,j,k]
+                # iterate over the neighboring voxels
+                for si, sj, sk in neighbor_transforms:
+                    # wrap points
+                    ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(i + si, j + sj, k + sk, nx, ny, nz)
+                    # skip vacuum neighs
+                    if vacuum_mask[ii,jj,kk]:
+                        continue
+                    # get neighbors label and image
+                    neigh_label = labeled_array[ii, jj, kk]
+                    neigh_image = images[ii, jj, kk]
+                    if neigh_label !=label:
+                        edges[i,j,k] = True
+                        break
+                    
+                    # adjust neigh image
+                    si1, sj1, sk1 = INT_TO_IMAGE[neigh_image]
+                    si1 += ssi
+                    sj1 += ssj
+                    sk1 += ssk
+                    neigh_image = IMAGE_TO_INT[si1,sj1,sk1]
+                    if neigh_image != image:
+                        edges[i,j,k] = True
+                        break
+    return edges
+
 @njit(cache=True)
 def get_neighboring_basin_connections(
     labeled_array: NDArray[np.int64],
     data: NDArray[np.float64],
     neighbor_transforms: NDArray[np.int64],
-    vacuum_mask: NDArray[np.bool_],
     edge_mask: NDArray[np.bool_],
     label_num: int,
         ):
@@ -240,7 +314,7 @@ def get_neighboring_basin_connections(
         for j in range(ny):
             for k in range(nz):
                 # if this voxel is part of the vacuum, continue
-                if vacuum_mask[i, j, k]:
+                if not edge_mask[i, j, k]:
                     continue
                 # get this voxels label
[…]
