+++ src/baderkit/core/critical_points/bader_based.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
-
+import time
 from baderkit.core import Bader
-from baderkit.core.utilities.basic import wrap_point, wrap_point_w_shift
+from baderkit.core.utilities.basic import wrap_point, wrap_point_w_shift, coords_to_flat, flat_to_coords
+from baderkit.core.utilities.union_find import union, find_root
 import numpy as np
 from numpy.typing import NDArray
 from numba import njit, prange
@@ -160,59 +161,132 @@ INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
 for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
     IMAGE_TO_INT[i,j,k] = shift_idx
 
-@njit(cache=True)
+# @njit(cache=True)
+# def get_cell_type(
+#     i,j,k,
+#     nx,ny,nz,
+#     labels,
+#     images,
+#     vacuum_mask,
+#         ):
+#     # get this cells parity
+#     pi = i&1
+#     pj = j&1
+#     pk = k&1
+#     parity = PARITY_TO_INT[pi,pj,pk]
+    
+#     # get transforms to voxels
+#     transforms = PARITY_VOXELS[parity]
+
+#     # iterate over the neighboring voxels
+#     current_labels = [-1]
+#     current_images = [-1]
+#     unique = 0
+#     for si, sj, sk in transforms:
+#         if unique == 3:
+#             break
+#         # wrap points
+#         ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(int((i + si)/2), int((j + sj)/2), int((k + sk)/2), nx, ny, nz)
+#         # skip vacuum points
+#         if vacuum_mask[ii,jj,kk]:
+#             continue
+
+#         # get neighbors label and image
+#         neigh_label = labels[ii, jj, kk]
+#         neigh_image = images[ii, jj, kk]
+        
+#         # adjust neigh image
+#         si1, sj1, sk1 = INT_TO_IMAGE[neigh_image]
+#         si1 += ssi
+#         sj1 += ssj
+#         sk1 += ssk
+#         neigh_image = IMAGE_TO_INT[si1,sj1,sk1]
+        
+#         # check if this is a new label/image
+#         new = True
+#         for label, image in zip(current_labels, current_images):
+#             if neigh_label == label and neigh_image == image:
+#                 new = False
+#                 break
+#         if new:
+#             current_labels.append(neigh_label)
+#             current_images.append(neigh_image)
+#             unique +=1
+
+#     return unique
+@njit(cache=True, inline='always')
 def get_cell_type(
-    i,j,k,
-    nx,ny,nz,
+    i, j, k,
+    nx, ny, nz,
     labels,
     images,
     vacuum_mask,
-        ):
-    # get this cells parity
-    pi = i&1
-    pj = j&1
-    pk = k&1
-    parity = PARITY_TO_INT[pi,pj,pk]
+):
+    # get the parity of this point
+    pi = i & 1
+    pj = j & 1
+    pk = k & 1
+    parity = PARITY_TO_INT[pi, pj, pk]
     
-    # get transforms to voxels
+    # get the corresponding transforms
     transforms = PARITY_VOXELS[parity]
 
-    # iterate over the neighboring voxels
-    current_labels = [-1]
-    current_images = [-1]
+    # initialize potential labels
+    label0 = -1; image0 = -1
+    label1 = -1; image1 = -1
     unique = 0
-    for si, sj, sk in transforms:
+
+    # iterate over transforms
+    for trans in range(transforms.shape[0]):
+        # if we've found more than two neighbors, immediately break
         if unique == 3:
             break
-        # wrap points
[…]
