+++ src/baderkit/core/critical_points/critical_points_numba.py
@@ -1,4 +1,10 @@
-from baderkit.core.utilities.basic import wrap_point_w_shift, wrap_point, coords_to_flat
+from baderkit.core.utilities.basic import (
+    wrap_point_w_shift, 
+    wrap_point, 
+    coords_to_flat, 
+    flat_to_coords,
+    )
+from baderkit.core.utilities.union_find import union, find_root_no_compression, find_root
 import itertools
 import numpy as np
 from numpy.typing import NDArray
@@ -103,6 +109,15 @@ INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
 for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
     IMAGE_TO_INT[i,j,k] = shift_idx
 
+FACE_TRANSFORMS = np.array([
+    [1,0,0],
+    [-1,0,0],
+    [0,1,0],
+    [0,-1,0],
+    [0,0,1],
+    [0,0,-1],
+    ], dtype=np.int64)
+
 @njit(cache=True)
 def get_cell_type(
     i,j,k,
@@ -168,10 +183,10 @@ def get_manifold_labels(
     vacuum_mask: NDArray[np.bool_],
 ):
     """
-    0: maxima
-    1: minima
-    2: 1-saddle
-    3: 2-saddle
+    0: minima
+    1: 1-saddle
+    2: 2-saddle
+    3: maxima
     4: 1-saddle ascending 2-manifold
     5: 1-saddle descending 1-manifold
     6: 2-saddle ascending 1-manifold
@@ -188,23 +203,34 @@ def get_manifold_labels(
 
     # create 3D array, twice the size of the original, to store critical points
     # and manifolds
-    manifold_labels = np.empty((nx2,ny2,nz2), dtype=np.uint8)
+    manifold_labels = np.full((nx2,ny2,nz2), np.iinfo(np.uint8).max, dtype=np.uint8)
     
-    # initialize maxima and minima
+    # initialize maxima and minima and get the double grid groups
+    new_maxima_groups = []
     for group in maxima_groups:
         # adjust to double grid coords
         group2 = group*2
+        new_group = []
         for i,j,k in group2:
+            new_group.append(np.array((i,j,k), dtype=np.uint16))
             for si, sj, sk in INT_TO_IMAGE:
                 ii, jj, kk = wrap_point(i+si, j+sj, k+sk, nx2, ny2, nz2)
-                manifold_labels[ii, jj, kk] = 0
+                manifold_labels[ii, jj, kk] = 3
+                new_group.append(np.array(ii,jj,kk), dtype=np.uint16)
+        new_maxima_groups.append(new_group)
+        
+    new_minima_groups = []
     for group in minima_groups:
         # adjust to double grid coords
         group2 = group*2
+        new_group = []
         for i,j,k in group2:
+            new_group.append(np.array((i,j,k), dtype=np.uint16))
             for si, sj, sk in INT_TO_IMAGE:
                 ii, jj, kk = wrap_point(i+si, j+sj, k+sk, nx2, ny2, nz2)
-                manifold_labels[ii, jj, kk] = 1
+                manifold_labels[ii, jj, kk] = 0
+                new_group.append(np.array(ii,jj,kk), dtype=np.uint16)
+        new_minima_groups.append(new_group)
     
     # loop over each voxel in parallel
     for i in prange(nx2):
@@ -212,7 +238,7 @@ def get_manifold_labels(
             for k in range(nz2):
                 
                 # skip maxima/minima
-                if manifold_labels[i,j,k] == 0 or manifold_labels[i,j,k] == 1:
+                if manifold_labels[i,j,k] == 0 or manifold_labels[i,j,k] == 3:
                     continue
                 
[…]
