### 2026-02-06 9fb394c
working saddle point finder
