+++ src/baderkit/core/critical_points/critical_points_numba.py
@@ -118,57 +118,72 @@ FACE_TRANSFORMS = np.array([
     [0,0,-1],
     ], dtype=np.int64)
 
-@njit(cache=True)
+@njit(cache=True, inline='always')
 def get_cell_type(
-    i,j,k,
-    nx,ny,nz,
+    i, j, k,
+    nx, ny, nz,
+    parity,
     labels,
     images,
     vacuum_mask,
-        ):
-    # get this cells parity
-    pi = i&1
-    pj = j&1
-    pk = k&1
-    parity = PARITY_TO_INT[pi,pj,pk]
+):
     
-    # get transforms to voxels
+    # get the corresponding transforms
     transforms = PARITY_VOXELS[parity]
 
-    # iterate over the neighboring voxels
-    current_labels = [-1]
-    current_images = [-1]
+    # initialize potential labels
+    label0 = -1; image0 = -1
+    label1 = -1; image1 = -1
     unique = 0
-    for si, sj, sk in transforms:
+
+    # iterate over transforms
+    for trans in range(transforms.shape[0]):
+        # if we've found more than two neighbors, immediately break
         if unique == 3:
             break
-        # wrap points
-        ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(int((i + si)/2), int((j + sj)/2), int((k + sk)/2), nx, ny, nz)
-        # skip vacuum points
-        if vacuum_mask[ii,jj,kk]:
-            continue
+        
+        # get shifts
+        si = transforms[trans, 0]
+        sj = transforms[trans, 1]
+        sk = transforms[trans, 2]
 
-        # get neighbors label and image
+        # move to original grid (same as (i+si)/2)
+        ii = (i + si) >> 1
+        jj = (j + sj) >> 1
+        kk = (k + sk) >> 1
+
+        # wrap around periodic edges and store shift
+        ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(
+            ii, jj, kk, nx, ny, nz
+        )
+
+        if vacuum_mask[ii, jj, kk]:
+            continue
+        
+        # get the label and image of this neighbor
         neigh_label = labels[ii, jj, kk]
         neigh_image = images[ii, jj, kk]
-        
-        # adjust neigh image
-        si1, sj1, sk1 = INT_TO_IMAGE[neigh_image]
-        si1 += ssi
-        sj1 += ssj
-        sk1 += ssk
-        neigh_image = IMAGE_TO_INT[si1,sj1,sk1]
-        
-        # check if this is a new label/image
-        new = True
-        for label, image in zip(current_labels, current_images):
-            if neigh_label == label and neigh_image == image:
-                new = False
-                break
-        if new:
-            current_labels.append(neigh_label)
-            current_images.append(neigh_image)
-            unique +=1
+
+        # update image to be relative to the current points transformation
+        si1 = INT_TO_IMAGE[neigh_image, 0] + ssi
+        sj1 = INT_TO_IMAGE[neigh_image, 1] + ssj
+        sk1 = INT_TO_IMAGE[neigh_image, 2] + ssk
+        neigh_image = IMAGE_TO_INT[si1, sj1, sk1]
+
+        # compare to any previous labels and update our unique number
+        if unique == 0:
+            label0 = neigh_label
+            image0 = neigh_image
+            unique = 1
+        elif unique == 1:
+            if neigh_label != label0 or neigh_image != image0:
+                label1 = neigh_label
+                image1 = neigh_image
+                unique = 2
+        elif unique == 2:
+            if ((neigh_label != label0 or neigh_image != image0) and
[…]
