+++ src/baderkit/core/critical_points/critical_points_numba1.py
@@ -1,852 +1,968 @@
-from numba import njit, prange
+import time
+from baderkit.core.utilities.basic import (
+    wrap_point_w_shift, 
+    wrap_point, 
+    coords_to_flat, 
+    flat_to_coords,
+    )
+from baderkit.core.utilities.union_find import union, find_root_no_compression, find_root
+from baderkit.core.bader.methods.shared_numba import get_edges_w_images, get_best_neighbor
+from baderkit.core import Bader
+import itertools
 import numpy as np
+from numpy.typing import NDArray
+from numba import njit, prange
 
+PARITIES = np.array(
+    [
+        [0, 0, 0],  # 3D
+        [0, 0, 1],  # 2D
+        [0, 1, 0],  # 2D
+        [0, 1, 1],  # 1D
+        [1, 0, 0],  # 2D
+        [1, 0, 1],  # 1D
+        [1, 1, 0],  # 1D
+        [1, 1, 1],  # 0D
+    ],
+    dtype=np.uint8,
+)
 
-def get_facet_transforms(parity):
-    facet_transforms = []
-    opposites = []
-    for i, p in enumerate(parity):
-        trans = np.zeros(3, dtype=np.int8)
-        if p == 1:
-            trans[i] = -1
-            facet_transforms.append(trans.copy())
-            opposites.append(1)
+PARITY_DIMS = np.array([3,2,2,1,2,1,1,0], dtype=np.uint8)
 
-            trans[i] = 1
-            facet_transforms.append(trans.copy())
-            opposites.append(-1)
-    return facet_transforms, opposites
+# Conversion from parity to its index
+PARITY_TO_INT = np.zeros((2, 2, 2), dtype=np.int8)
+for idx, (i, j, k) in enumerate(PARITIES):
+    PARITY_TO_INT[i, j, k] = idx
 
+###############################################################################
+# Voxel Transforms
+###############################################################################
+# Generate transformations to vertices for each parity type.
+VERTEX_VOXELS = [
+    [-1, -1, -1],
+    [1, -1, -1],
+    [1, 1, -1],
+    [-1, 1, -1],
+    [-1, -1, 1],
+    [1, -1, 1],
+    [1, 1, 1],
+    [-1, 1, 1],
+]
+EDGE_VOXELS = [
+    [-1, -1],
+    [1, -1],
+    [1, 1],
+    [-1, 1],
+]
+FACE_VOXELS = [
+    [-1],
+    [1],
+]
+VOXEL_VOXELS = [[]]
+VOXELS = [VERTEX_VOXELS, EDGE_VOXELS, FACE_VOXELS, VOXEL_VOXELS]
 
-def get_cofacet_transforms(parity):
-    cofacet_transforms = []
-    for i, p in enumerate(parity):
-        trans = np.zeros(3, dtype=np.int8)
-        if p == 0:
-            trans[i] = -1
-            cofacet_transforms.append(trans.copy())
-            trans[i] = 1
-            cofacet_transforms.append(trans.copy())
-    # add 0,0,0 for 3d parity
-    if not cofacet_transforms:
-        cofacet_transforms.append(np.zeros(3, dtype=np.int8))
-    return cofacet_transforms
-
-
-def get_trans_maps():
-    trans_to_int = np.zeros((3, 3, 3), dtype=np.int8)
-    int_to_trans = np.zeros((6, 3), dtype=np.int8)
-    for i in range(6):
-        j = i // 2
-        value = (-1) ** (i & 1)  # -1 for even, 1 for odd
-        int_to_trans[i, j] = value
[…]

[…]
