### 2026-02-08 fea305d
working critical points connections
