+++ src/baderkit/core/critical_points/flood_fill.py
@@ -154,14 +154,19 @@ def get_possible_saddles(
                     )
                 
                 if num_neighs == 1 and opp_num_neighs > 1:
+                    # saddle 1
                     edges[i,j,k] = 1
                 elif num_neighs > 1 and opp_num_neighs == 1:
+                    # saddle 2
                     edges[i,j,k] = 2
                 elif num_neighs == 1 and opp_num_neighs <1:
+                    # edge of maxima manifold
                     edges[i,j,k] = 3
                 elif num_neighs < 1 and opp_num_neighs == 1:
+                    # edge of minima manifold
                     edges[i,j,k] = 4
                 elif num_neighs == 1 and opp_num_neighs == 1:
+                    # edge of both maxima/minima manifold
                     edges[i,j,k] = 5
     return edges
 
@@ -319,7 +324,7 @@ def get_possible_saddle_vals(
                 
     return edge_mask, possible_saddles, saddle_connections, connection_vals
 
-# @njit
+@njit
 def get_saddles(
     connection_values,
     saddle_coords,
@@ -350,170 +355,103 @@ def get_saddles(
     
     return saddles, best_vals
 
-# @njit(parallel=True)
-# def get_possible_1saddles(
-#     labels: NDArray[np.int64],
-#     images: NDArray[np.int64],
-#     data: NDArray[np.float64],
-#     neighbor_transforms: NDArray[np.int64],
-#     edge_mask: NDArray[np.uint8],
-# ):
-#     nx, ny, nz = labels.shape
-#     ny_nz = ny*nz
+@njit
+def get_saddle_connections(
+    saddle1_coords,
+    saddle2_coords,
+    neighbor_transforms: NDArray[np.int64],
+    edge_mask: NDArray[np.uint8],
+):
+    nx, ny, nz = edge_mask.shape
     
-#     # get the points that may be saddles
-#     possible_saddles = np.argwhere(edge_mask==1)
+    # get the number of possible edges
+    num_edges = len(np.where(np.isin(edge_mask, (1,2,5)))[0])
+    num_edges += len(saddle1_coords) + len(saddle2_coords)
     
-#     # create an array to track connections between these points.
-#     # For each entry we will have:
-#         # 1: the lower label index
-#         # 2: the higher label index
-#         # 3: the connection image from lower basin to higher basin
-#     saddle_connections = np.empty((len(possible_saddles),3),dtype=np.uint16)
-#     connection_vals = np.empty(len(possible_saddles), dtype=np.float64)
-#     # create a mask to track important connections
-#     important = np.ones(len(possible_saddles), dtype=np.bool)
-#     max_val = np.iinfo(np.uint16).max
-#     for idx in prange(len(possible_saddles)):
-#         i,j,k = possible_saddles[idx]
-#         lower, higher, shift, connection_value = get_saddle_connection(
-#             i, j, k,
-#             nx, ny, nz,
-#             ny_nz,
-#             labels,
-#             images,
-#             data,
-#             neighbor_transforms,
-#             edge_mask,
-#             max_val,
-#         )
-#         if lower == max_val:
-#             # note this wasn't a true saddle
-#             important[idx] = False
-#             continue
-#         saddle_connections[idx, 0] = lower
-#         saddle_connections[idx, 1] = higher
-#         saddle_connections[idx, 2] = shift
-#         connection_vals[idx] = connection_value
-        
-#     # get only the connections that are important
-#     important = np.where(important)[0]
-#     possible_saddles = possible_saddles[important]
-#     saddle_connections = saddle_connections[important]
-#     connection_vals = connection_vals[important]
+    # create an empty queue for storing which points are next
[…]
