### 2026-02-08 #9943 @editorialbot
[REVIEW]: BaderKit: A Python Package for Grid-based Bader Charge Analysis
**Submitting author:** @SWeav02 (<a href="http://orcid.org/0000-0001-9358-3281">Samuel Weaver</a>)
**Repository:** https://github.com/SWeav02/baderkit
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.9.1
