### 2026-02-10 5e904c3
working newton refinement
