+++ src/baderkit/plotting/core/grid.py
@@ -1,0 +1,543 @@
+# -*- coding: utf-8 -*-
+
+from pathlib import Path
+
+import numpy as np
+import pyvista as pv
+from numpy.typing import NDArray
+
+
+from baderkit.core import Grid
+from .structure import StructurePlotter
+
+    
+class GridPlotter(StructurePlotter):
+    def __init__(
+        self,
+        grid: Grid,
+        **structure_kwargs,
+        # downscale: int | None = 400,
+    ):
+        """
+        A convenience class for creating plots of crystal structures and isosurfaces
+        using pyvista's package for VTK.
+
+        Parameters
+        ----------
+        grid : Grid
+            The Grid object to use for isosurfaces. The structure will be pulled
+            from this grid.
+
+        Returns
+        -------
+        None.
+
+        """
+        # apply StructurePlotter kwargs
+        super().__init__(structure=grid.structure, **structure_kwargs)
+
+        # Grid specific items
+        # if downscale is not None:
+        #     if grid.voxel_resolution > downscale:
+        #         # downscale the grid for speed
+        #         logging.info("Grid is above desired resolution. Downscaling.")
+        #         grid = grid.regrid(downscale)
+        self.grid = grid
+        self._show_surface = True
+        self._show_caps = True
+        self._surface_opacity = 0.8
+        self._cap_opacity = 0.8
+        self._colormap = "viridis"
+        self._use_solid_surface_color = False
+        self._use_solid_cap_color = False
+        self._surface_color = "#BA8E23"
+        self._cap_color = "#BA8E23"
+
+        # wrap values around to get one extra voxel on the far side of each axis.
+        values = np.pad(grid.total, pad_width=((0, 1), (0, 1), (0, 1)), mode="wrap")
+        self.shape = values.shape
+        self.values = values.ravel(order="F")
+        self._min_val = self.values.min()
+        # make min val slightly above 0
+        self._min_val += 0.0000001 * self._min_val
+        self._max_val = self.values.max()
+        # determine default iso if not provided
+        self._iso_val = self._min_val  # np.mean(grid.total)
+        # generate the structured grid
+        indices = np.indices(self.shape).reshape(3, -1, order="F").T
+        self.points = grid.grid_to_cart(indices)
+        self.structured_grid = self._make_structured_grid(self.values)
+        # generate the surface
+        self.surface = self.structured_grid.extract_surface()
+        # update plotter
+        self.plotter = self._create_grid_plot()
+
+    def _make_structured_grid(self, values: NDArray[float]) -> pv.StructuredGrid:
+        """
+        Creates a pyvista StructuredGrid object for making isosurfaces. This
+        should generally not be called directly.
+
+        Parameters
+        ----------
+        values : NDArray[float]
+            A 3xN array of values representing the data in the structured grid.
+            These should be raveled/reshaped using Fortran's conventions (order='F'')
+
+        Returns
+        -------
+        structured_grid : pv.StructuredGrid
+            A pyvista StructuredGrid with values representing the grid data.
+
+        """
+        structured_grid = pv.StructuredGrid()
+        structured_grid.points = self.points
+        structured_grid.dimensions = self.shape
+        structured_grid["values"] = values
+        return structured_grid
+
+    @property
+    def show_surface(self) -> bool:
+        """
+
+        Returns
+        -------
+        bool
+            whether or not to display the isosurface.
+
+        """
+        return self._show_surface
+
+    @show_surface.setter
+    def show_surface(self, show_surface: bool):
+        if "iso" in self.plotter.actors.keys():
+            actor = self.plotter.actors["iso"]
+            actor.visibility = show_surface
+        self._show_surface = show_surface
+
+    @property
+    def show_caps(self) -> bool:
+        """
+
[…]
