+++ src/baderkit/core/critical_points/__init__.py
@@ -1,1 +1,3 @@
 # -*- coding: utf-8 -*-
+
+from .critical_points import CriticalPoints
