### 2026-02-11 98144fb
start plotting edges
