### 2026-02-12 975fbd1
initial working edges
