### 2026-02-12 c72e468
working but slow
