+++ src/baderkit/plotting/core/critical_points.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+from itertools import product# -*- coding: utf-8 -*-
 
 import numpy as np
 import pyvista as pv
@@ -35,20 +35,19 @@ class CriticalPointsPlotter(GridPlotter):
         self.critical_graph = critical_points.morse_graph
         
         self._visible_critical_points = [i for i in range(len(self.critical_graph.nodes))]
-        self._critical_point_radii = [0.25 for s in self.critical_graph.nodes]
+        self._critical_point_radii = [0.75 for s in self.critical_graph.nodes]
         self._critical_point_colors = [CRIT_COLORS.get(self.critical_graph.nodes[cp]["type"], "#FFFFFF") for cp in self.critical_graph.nodes]
+        
+        self._visible_edges = [i for i in range(len(self.critical_graph.edges))]
+        self._edge_thicknesses = [4.0 for s in self.critical_graph.edges]
+        self._edge_colors = ["black" for edge in self.critical_graph.edges]
+        
         
         self.plotter = self._create_critical_plot()
         
         self.show_caps = False
         self.show_surface = False
         self.visible_atoms = []
-    # visible critical points
-    # visible connections
-    # critical radii
-    # connection thickness
-    # critical colors
-    # connection colors
     
 
     @property
@@ -67,13 +66,37 @@ class CriticalPointsPlotter(GridPlotter):
     def visible_critical_points(self, visible_critical_points: set[int]):
         # update visibility of critical points
         for i in range(len(self.critical_graph)):
-            actor = self.plotter.actors[f"crit_{i}"]
+            actor = self.plotter.actors[f"crit_{i}{self._actor_suffix}"]
             if i in visible_critical_points:
                 actor.visibility = True
             else:
                 actor.visibility = False
         # set visible points
         self._visible_critical_points = visible_critical_points
+        
+    @property
+    def visible_edges(self) -> list[int]:
+        """
+
+        Returns
+        -------
+        list[int]
+            A list of critical points to include in the plot
+
+        """
+        return self._visible_edges
+
+    @visible_edges.setter
+    def visible_edges(self, visible_edges: set[int]):
+        # update visibility of critical points
+        for i in range(len(self.critical_graph)):
+            actor = self.plotter.actors[f"edge_{i}{self._actor_suffix}"]
+            if i in visible_edges:
+                actor.visibility = True
+            else:
+                actor.visibility = False
+        # set visible points
+        self._visible_edges = visible_edges
 
     @property
     def critical_point_colors(self) -> list[str]:
@@ -94,9 +117,34 @@ class CriticalPointsPlotter(GridPlotter):
         for crit_idx, old_color, new_color in zip(self.critical_points, self.critical_point_colors, colors):
             if old_color == new_color:
                 continue
-            actor = self.plotter.actors[f"crit_{crit_idx}"]
-            actor.prop.color = new_color
+            for suffix in ["", "_wrapped"]:
+                actor = self.plotter.actors[f"crit_{crit_idx}{suffix}"]
+                actor.prop.color = new_color
         self._critical_point_colors = colors
+        
+    @property
+    def edge_colors(self) -> list[str]:
+        """
+
+        Returns
+        -------
+        list[str]
+            The colors to use for each atom as hex codes.
+
+        """
+        return self._edge_colors
+
+    @edge_colors.setter
+    def edge_colors(self, colors: list[str]):
+        # for each site, check if the radius has changed and if it has remove it
+        # then remake
+        for edge_idx, (old_color, new_color) in enumerate(zip(self.edge_colors, colors)):
[…]
