### 2026-02-15 b05e8d2
some improvements for ELF crit points
