+++ src/baderkit/plotting/core/critical_points.py
@@ -2,10 +2,11 @@ from itertools import product# -*- coding: utf-8 -*-
 
 import numpy as np
 import pyvista as pv
+from numpy.typing import NDArray
 
 from baderkit.core.critical_points import CriticalPoints
 from .grid import GridPlotter
-from baderkit.plotting.core.defaults import CRIT_COLORS
+from baderkit.plotting.core.defaults import CRIT_COLORS, CONNECTION_COLORS
     
 class CriticalPointsPlotter(GridPlotter):
     def __init__(
@@ -30,25 +31,34 @@ class CriticalPointsPlotter(GridPlotter):
         """
         # apply StructurePlotter kwargs
         grid = critical_points.reference_grid
-        super().__init__(grid=grid, **grid_kwargs)
+        
+        # set initial critical point objects
         self.critical_points = critical_points
-        self.critical_graph = critical_points.morse_graph
+        self.morse_graph = critical_points.morse_graph
         
-        self._visible_critical_points = [i for i in range(len(self.critical_graph.nodes))]
-        self._critical_point_radii = [0.75 for s in self.critical_graph.nodes]
-        self._critical_point_colors = [CRIT_COLORS.get(self.critical_graph.nodes[cp]["type"], "#FFFFFF") for cp in self.critical_graph.nodes]
+        # set critical coord properties
+        self._visible_critical_points = np.array([1.0 for i in range(4)], dtype=float)
+        self._critical_point_radii = np.array([0.75 for s in range(4)])
+        self._critical_point_colors = np.array([pv.Color(CRIT_COLORS.get(crit_type, "#FFFFFF")).float_rgb for crit_type in range(4)])
         
-        self._visible_edges = [i for i in range(len(self.critical_graph.edges))]
-        self._edge_thicknesses = [4.0 for s in self.critical_graph.edges]
-        self._edge_colors = ["black" for edge in self.critical_graph.edges]
+        # set edge properties
+        self._visible_connections = np.array([1.0 for i in range(3)])
+        self._connection_thickness = 4.0
+        self._connection_colors = np.array([pv.Color(CONNECTION_COLORS.get(crit_type, "#FFFFFF")).float_rgb for crit_type in range(3)])
         
         
-        self.plotter = self._create_critical_plot()
+        # meshes
+        self._wrapped_crit_poly = None
+        self._crit_poly = None
+        self._map_wrapped_to_crits = None
+        self._map_to_crits = None
+        self._connection_meshes = None
+        self._wrapped_connection_meshes = None
+        super().__init__(grid=grid, **grid_kwargs)
         
+        # update to hide by default        
         self.show_caps = False
         self.show_surface = False
-        self.visible_atoms = []
-    
 
     @property
     def visible_critical_points(self) -> list[int]:
@@ -64,18 +74,19 @@ class CriticalPointsPlotter(GridPlotter):
 
     @visible_critical_points.setter
     def visible_critical_points(self, visible_critical_points: set[int]):
-        # update visibility of critical points
-        for i in range(len(self.critical_graph)):
-            actor = self.plotter.actors[f"crit_{i}{self._actor_suffix}"]
-            if i in visible_critical_points:
-                actor.visibility = True
-            else:
-                actor.visibility = False
-        # set visible points
+        # convert to array
+        visible_critical_points = np.array(visible_critical_points, dtype=np.float64)
+        
+        # make sure we have the write shape
+        assert self.visible_critical_points.shape == (4,), "Length must be 4"
+
+        # set visible atoms
         self._visible_critical_points = visible_critical_points
+        # update
+        self._update_critical_meshes()
         
     @property
-    def visible_edges(self) -> list[int]:
[…]

[…]
