### 2026-02-17 SWeav02/baderkit#45 @mdavezac
JOSS review
Thanks for submitting to JOSS.
Here are a few points I'm curious about:

1. It is not so easy to check the code works in anyway since example CHGCAR (and INCAR? and POSCAR? KPOINTS? POTCARs I think cannot be distributed however) are not provided explicitly. Pointing to the CHGCAR in the tests could be one solution, thought the more the merrier.
2. The outputs are never shown, whether the csv files from the cli or the output of the python objects. Showing "how" is good, showing "what" is missing.
4. When are POTCARs required? oxydation numbers apparently, but it's not really documented that I can tell
5. The code fails when reading the charge densities given at https://theory.cm.utexas.edu/henkelman/code/bader/. Can't say whether it's a real bug, but again, examples are not really provided here.
6. I can't for the life of me figure out figure 2. What crystal structure is this? For the purpose of the JOSS review, it would be good to provide the inputs so we can reproduce the claim.
7. What system is being studied in figure 1? NaCl presumably, from the benchmark files. But the paper is less explicit. For the purpose of JOSS, CHGCARs would be welcome so we can reproduce the claim (if not the density files themselves, whithout VASP)
8. Presumably, figure 2 is done for a fixed system with a given number of atoms/basins and increasingly fines direct space meshes. However, I can't quite guess whether the strategies between the henkelman code and baderkit would yield different behaviors with respect to the number of atoms/basins.
9. New practitioners may appreciate some pointers or known trends with respect charge density cutoff and kpoint convergence in the documentation, e.g. should they expect to greatly increase them compared to total energy convergence, or should they not worry too much?
10. Out of curiosity, does the algorithm extend easily to GPUs?
11. Might be my system (Linux, NixOS), but installing "baderkit[gui]" works, yet running baderkit gui complains that "baderkit[gui]" is not installed.
