+++ src/baderkit/core/utilities/interpolation.py
@@ -444,6 +444,10 @@ def spline_grad(i, j, k, data, h=0.5, is_frac=False):
         i = i * nx
         j = j * ny
         k = k * nz
+        
+    i = float(i)
+    j = float(j)
+    k = float(k)
 
     fxp = interp_spline(i + h, j, k, data, False)
     fxm = interp_spline(i - h, j, k, data, False)
@@ -459,6 +463,7 @@ def spline_grad(i, j, k, data, h=0.5, is_frac=False):
     gz = (fzp - fzm) / (2.0 * h)
 
     return gx, gy, gz
+   
 
 @njit(fastmath=True)
 def spline_hess(i, j, k, data, h=0.5, is_frac=False):
@@ -471,27 +476,37 @@ def spline_hess(i, j, k, data, h=0.5, is_frac=False):
         i = i * nx
         j = j * ny
         k = k * nz
+        
+    i = float(i)
+    j = float(j)
+    k = float(k)
+    
+    h2 = 2.0*h
+    hh = h*h
+    hh4 = 4.0*hh
+
 
     f0 = interp_spline(i, j, k, data, False)
+    f02 = f0*2
 
     # Second derivatives
     f_xx = (
         interp_spline(i + h, j, k, data, False)
-        - 2.0 * f0
+        - f02
         + interp_spline(i - h, j, k, data, False)
-    ) / (h * h)
+    ) / hh
 
     f_yy = (
         interp_spline(i, j + h, k, data, False)
-        - 2.0 * f0
+        - f02
         + interp_spline(i, j - h, k, data, False)
-    ) / (h * h)
+    ) / hh
 
     f_zz = (
         interp_spline(i, j, k + h, data, False)
-        - 2.0 * f0
+        - f02
         + interp_spline(i, j, k - h, data, False)
-    ) / (h * h)
+    ) / hh
 
     # Mixed partials
     f_xy = (
@@ -499,21 +514,21 @@ def spline_hess(i, j, k, data, h=0.5, is_frac=False):
         - interp_spline(i + h, j - h, k, data, False)
         - interp_spline(i - h, j + h, k, data, False)
         + interp_spline(i - h, j - h, k, data, False)
-    ) / (4.0 * h * h)
+    ) / hh4
 
     f_xz = (
         interp_spline(i + h, j, k + h, data, False)
         - interp_spline(i + h, j, k - h, data, False)
         - interp_spline(i - h, j, k + h, data, False)
         + interp_spline(i - h, j, k - h, data, False)
-    ) / (4.0 * h * h)
+    ) / hh4
 
     f_yz = (
         interp_spline(i, j + h, k + h, data, False)
         - interp_spline(i, j + h, k - h, data, False)
         - interp_spline(i, j - h, k + h, data, False)
         + interp_spline(i, j - h, k - h, data, False)
-    ) / (4.0 * h * h)
+    ) / hh4
 
     H = np.empty((3, 3))
     H[0, 0] = f_xx
@@ -526,10 +541,259 @@ def spline_hess(i, j, k, data, h=0.5, is_frac=False):
 
     return H
 
+@njit(fastmath=True)
+def spline_grad_and_hess(i, j, k, data, h=0.5, is_frac=False):
+    """
+    Compute both gradient and Hessian of a spline-interpolated scalar field
+    with respect to grid coordinates (i, j, k), minimizing redundant interpolations.
+    """
+    nx, ny, nz = data.shape
+    if is_frac:
+        i = i * nx
+        j = j * ny
+        k = k * nz
+        
+    i = float(i)
+    j = float(j)
[…]
