+++ src/baderkit/core/utilities/interpolation.py
@@ -16,7 +16,7 @@ from baderkit.core.utilities.basic import (
 ###############################################################################
 
 
-@njit(inline="always", cache=True, fastmath=True)
+#@njit(inline="always", cache=True, fastmath=True)
 def interp_nearest(i, j, k, data, is_frac=True):
     nx, ny, nz = data.shape
     if is_frac:
@@ -36,7 +36,7 @@ def interp_nearest(i, j, k, data, is_frac=True):
 ###############################################################################
 # Linear interpolation
 ###############################################################################
-@njit(inline="always", cache=True, fastmath=True)
+#@njit(inline="always", cache=True, fastmath=True)
 def interp_linear(i, j, k, data, is_frac=True):
     nx, ny, nz = data.shape
 
@@ -87,7 +87,7 @@ def interp_linear(i, j, k, data, is_frac=True):
 ###############################################################################
 
 
-@njit(cache=True, inline="always", fastmath=True)
+#@njit(cache=True, inline="always", fastmath=True)
 def cubic_bspline_weights(di, dj, dk):
     weights = np.empty((3, 4), dtype=np.float64)
     for d_idx, d in enumerate((di, dj, dk)):
@@ -104,7 +104,7 @@ def cubic_bspline_weights(di, dj, dk):
     return weights
 
 
-@njit(cache=True, fastmath=True)
+#@njit(cache=True, fastmath=True)
 def interp_spline(i, j, k, data, is_frac=True):
     nx, ny, nz = data.shape
 
@@ -155,7 +155,7 @@ def interp_spline(i, j, k, data, is_frac=True):
 ###############################################################################
 
 
-@njit(cache=True)
+#@njit(cache=True)
 def interpolate_point(
     point,
     method,
@@ -175,7 +175,7 @@ def interpolate_point(
     return value
 
 
-@njit(parallel=True, cache=True)
+#@njit(parallel=True, cache=True)
 def interpolate_points(points, method, data, is_frac=True):
     out = np.empty(len(points))
     if method == "nearest":
@@ -198,7 +198,7 @@ def interpolate_points(points, method, data, is_frac=True):
     return out
 
 
-@njit(cache=True)
+#@njit(cache=True)
 def linear_slice(
     data,
     p1: NDArray[float],
@@ -219,7 +219,7 @@ def linear_slice(
 ###############################################################################
 # Parabolic Fitting
 ###############################################################################
-@njit(fastmath=True, cache=True)
+#@njit(fastmath=True, cache=True)
 def refine_frac_extrema_parabolic(grid, frac_coords, lattice, use_minima=False):
     """
     Numerically stable refinement of a local maximum or minimum on a 3D periodic grid.
@@ -381,7 +381,7 @@ def refine_frac_extrema_parabolic(grid, frac_coords, lattice, use_minima=False):
 
 
 
-@njit(parallel=True, cache=True)
+#@njit(parallel=True, cache=True)
 def refine_extrema_parabolic(
     extrema_coords,
     extrema_children,
@@ -431,7 +431,7 @@ def refine_extrema_parabolic(
 # Newton Refinement
[…]
