### 2026-02-20 35d442b
seemingly working betti nums
