+++ untitled1.py
@@ -1,759 +1,49 @@
 # -*- coding: utf-8 -*-
 """
-Created on Tue Feb 10 11:30:01 2026
+Created on Thu Feb 19 11:22:49 2026
 
 @author: sammw
 """
-import itertools
-from baderkit.core import Bader
-from numba import njit, prange
 import numpy as np
-from numpy.typing import NDArray
-from baderkit.core.bader.methods.shared_numba import get_extrema
-from baderkit.core.utilities.basic import wrap_point, coords_to_flat, wrap_point_w_shift, flat_to_coords, get_norm
-from baderkit.core.utilities.union_find import union, find_root
-from baderkit.core.utilities.interpolation import interp_linear, interp_spline
-from baderkit.core.critical_points.critical_points import CriticalPoints
 
-IMAGE_TO_INT = np.empty([3,3,3], dtype=np.int64)
-INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
-for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
-    IMAGE_TO_INT[i,j,k] = shift_idx
+from baderkit.core import Bader
+from baderkit.core.critical_points.hessian_based import find_saddle_points
+from baderkit.core.critical_points import CriticalPoints
 
-FACE_TRANSFORMS = np.array([
-    [1,0,0],
-    [-1,0,0],
-    [0,1,0],
-    [0,-1,0],
-    [0,0,1],
-    [0,0,-1],
-    ], dtype=np.int64)
+bader = Bader.from_vasp(reference_filename="ELFCAR", maxima_persistence_tol=0.03)
 
-@njit(cache=True)
-def get_ongrid_gradient_cart(i, j, k, data, dir2car):
-    nx, ny, nz = data.shape
-    
-    c000 = data[i,j,k]
-    c100 = data[(i + 1) % nx,j,k]
-    c_100 = data[(i - 1) % nx,j,k]
-    c010 = data[i,(j + 1) % ny,k]
-    c0_10 = data[i,(j - 1) % ny,k]
-    c001 = data[i,j,(k + 1) % nz]
-    c00_1 = data[i,j,(k - 1) % nz]
+# maxima_groups, minima_groups = bader.get_persistence_groups()
 
-    # central differences in voxel coordinates
-    gi = (c100 - c_100) / (2.0)
-    gj = (c010 - c0_10) / (2.0)
-    gk = (c001 - c00_1) / (2.0)
-
-    # optional extrema clamping
-    if c100 <= c000 and c_100 <= c000:
-        gi = 0.0
-    if c010 <= c000 and c0_10 <= c000:
-        gj = 0.0
-    if c001 <= c000 and c00_1 <= c000:
-        gk = 0.0
-
-    # convert to fractional-coordinate gradient
-    gi *= nx
-    gj *= ny
-    gk *= nz
-
-    # convert to Cartesian gradient
-    gx = dir2car[0, 0] * gi + dir2car[0, 1] * gj + dir2car[0, 2] * gk
-    gy = dir2car[1, 0] * gi + dir2car[1, 1] * gj + dir2car[1, 2] * gk
-    gz = dir2car[2, 0] * gi + dir2car[2, 1] * gj + dir2car[2, 2] * gk
-
-    return gx, gy, gz
-
-@njit(parallel=True, cache=True)
-def get_residuals(
-    data,
-    neighbor_transforms,
-    dir2car,
-):
-    nx, ny, nz = data.shape
-    # create 3D array to store edges
-    residuals = np.empty_like(data, dtype=np.float64)
-    
-    # loop over each voxel in parallel
-    for i in prange(nx):
-        for j in range(ny):
[…]
