+++ src/baderkit/core/critical_points/critical_points.py
@@ -12,7 +12,8 @@ import numpy as np
 import pandas as pd
 from numpy.typing import NDArray
 from networkx import MultiDiGraph
-from scipy.spatial import ConvexHull
+from scipy.spatial import ConvexHull, Delaunay
+from scipy.ndimage import binary_erosion, distance_transform_edt, label
 
 from baderkit.core.base.base_analysis import BaseAnalysis
 from baderkit.core.toolkit import Structure, Grid
@@ -26,6 +27,7 @@ from .critical_points_numba import (
     INT_TO_IMAGE,
     IMAGE_TO_INT
 )
+from .betti_numbers import get_all_betti_numbers
 from .hessian_based import find_saddle_points, refine_saddle_points
 
 # This allows for Self typing and is compatible with python 3.10
@@ -84,6 +86,8 @@ class CriticalPoints(BaseAnalysis):
         "maxima_vox",
         "maxima_persistence_groups",
         "minima_persistence_groups",
+        "maxima_group_types",
+        "minima_group_types",
         "saddle1_vox",
         "saddle2_vox",
         "critical_vox",
@@ -194,6 +198,12 @@ class CriticalPoints(BaseAnalysis):
         if self._maxima_persistence_groups is None:
             self._get_extrema_persistence_groups()
         return self._maxima_persistence_groups
+    
+    @property
+    def maxima_group_types(self):
+        if self._maxima_group_types is None:
+            self._get_extrema_persistence_groups()
+        return self._maxima_group_types
 
     @property
     def minima_vox(self) -> NDArray:
@@ -202,6 +212,12 @@ class CriticalPoints(BaseAnalysis):
     @property
     def minima_frac(self) -> NDArray:
         return self.bader.minima_frac
+    
+    @property
+    def minima_group_types(self):
+        if self._minima_group_types is None:
+            self._get_extrema_persistence_groups()
+        return self._minima_group_types
     
     @property
     def minima_persistence_groups(self):
@@ -338,8 +354,8 @@ class CriticalPoints(BaseAnalysis):
             data=self.bader.reference_grid.total,
             matrix=self.bader.reference_grid.matrix,
             allowed_points=allowed_points,
+            eig_rel_tol=1e-4,
             )
-        breakpoint()
         
         # refine and get exact coords
         saddle1_coords, saddle2_coords = refine_saddle_points(
@@ -357,20 +373,114 @@ class CriticalPoints(BaseAnalysis):
         self._saddle1_frac = saddle1_frac
         self._saddle2_frac = saddle2_frac
         
-    def _get_extrema_persistence_groups(self):
-        
-        maxima_groups, minima_groups = self.bader.get_persistence_groups()
+    
+    @staticmethod
+    def _classify_extrema_group(
+        group_set,
+        pad=1,
+        voxel_tol=1.0,
+        coplanar_tol=1e-3,
+    ):
+        """
+        Returns:
+            0 : blob / filled object
+            1 : ring-like (coplanar hollow)
+            2 : cage-like (non-coplanar hollow)
+            3 : thin shell / surface blob (no interior)
+        """
+    
 
-        # make lists to store new groups
-        new_maxima_groups = []
+        # Build local voxel volume
+        P = np.asarray(list(group_set), dtype=float)
+        if P.shape[0] < 4:
+            return 0
+    
+        mins = np.floor(P.min(axis=0)).astype(int) - pad
+        maxs = np.ceil(P.max(axis=0)).astype(int) + pad
+    
+        shape = (maxs - mins + 1)
+        vol = np.zeros(shape, dtype=bool)
+    
+        idx = (P.astype(int) - mins)
[…]
