### 2026-02-23 92191e4
mostly working connections
