### 2026-02-24 d9f18af
partial ring support
