+++ untitled1.py
@@ -16,6 +16,21 @@ bader = Bader.from_vasp(reference_filename="ELFCAR", maxima_persistence_tol=0.03
 # maxima_groups, minima_groups = bader.get_persistence_groups()
 
 critical_points = CriticalPoints(bader)
+
+maxima, minima = bader.get_persistence_groups()
+test_grid = bader.reference_grid.copy()
+mask = np.zeros(test_grid.shape, dtype=bool)
+for coords in maxima:
+    mask[coords[:,0],coords[:,1],coords[:,2]] = True
+test_grid.total=mask
+test_grid.write_vasp("ELFCAR_test")
+# TODO:
+    # make sure ring labeling works.
+    # fix ring to use limited subset
+    # fix issues with bad critical point connections by adding increasing
+    # range of transformations
+
+
 plotter = CriticalPointsPlotter(critical_points)
 
 manifold_labels = critical_points.manifold_labels
