+++ src/baderkit/core/critical_points/saddle_connections.py
@@ -24,17 +24,23 @@ INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
 for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
     IMAGE_TO_INT[i,j,k] = shift_idx
     
+@njit(cache=True)
 def get_saddle_most_connections(
     i,j,k,
     labels,
     images,
     neighbor_transforms,
+    neighbor_dists,
     vacuum_mask,
         ):
     nx,ny,nz = labels.shape
     neigh_labels = []
     neigh_label_counts = []
-    for trans, (si, sj, sk) in  enumerate(neighbor_transforms):
+    max_dist = neighbor_dists.max()
+    for trans, ((si, sj, sk), dist) in  enumerate(zip(
+            neighbor_transforms,
+            neighbor_dists,
+             )):
         # get the neighbor
         ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(
             i+si, j+sj, k+sk, nx, ny, nz
@@ -51,11 +57,11 @@ def get_saddle_most_connections(
         comb = tuple((neigh_label, neigh_image))
         if not comb in neigh_labels:
             neigh_labels.append(comb)
-            neigh_label_counts.append(1)
+            neigh_label_counts.append(1/(1+dist/max_dist))
         else:
             for label_idx, label in enumerate(neigh_labels):
                 if comb == label:
-                    neigh_label_counts[label_idx] += 1
+                    neigh_label_counts[label_idx] += 1/(1+dist/max_dist)
                     break
     # get two most found neighbors
     neigh_label_counts = np.array(neigh_label_counts, dtype=np.uint8)
@@ -63,19 +69,21 @@ def get_saddle_most_connections(
     return neigh_labels[b0][0], neigh_labels[b0][1], neigh_labels[b1][0], neigh_labels[b1][1]
     
     
-    
+@njit(cache=True)
 def get_saddle_extrema_connections(
     labels,
     images,
     saddle_coords,
+    saddle_indices,
     neighbor_transforms: NDArray[np.int64],
+    neighbor_dists,
     vacuum_mask,
 ):
     nx, ny, nz = labels.shape
 
     connections = []
     problem_saddles = []
-    for saddle_idx in range(len(saddle_coords)):
+    for saddle_idx in saddle_indices:
         i,j,k = saddle_coords[saddle_idx]
         
         label0 = -1; image0 = -1
@@ -118,6 +126,7 @@ def get_saddle_extrema_connections(
                     labels, 
                     images, 
                     neighbor_transforms, 
+                    neighbor_dists,
                     vacuum_mask)
                 break
 
@@ -128,6 +137,128 @@ def get_saddle_extrema_connections(
             problem_saddles.append(saddle_idx)
 
     return connections, problem_saddles
+
+
+@njit(cache=True)
+def get_saddle_saddle_connections(
+    saddle1_coords,
+    saddle2_coords,
+    neighbor_transforms: NDArray[np.int64],
+    edge_mask: NDArray[np.uint8],
+):
+    nx, ny, nz = edge_mask.shape
+    
+    # get the number of possible edges
+    num_edges = len(np.where(np.isin(edge_mask, (1,2,6)))[0])
+    num_edges += len(saddle1_coords) + len(saddle2_coords)
+    
+    # create an empty queue for storing which points are next
+    queue = np.empty((num_edges, 3), dtype=np.uint32)
+    
+    # create arrays to store flood filled labels
+    max_val = np.iinfo(np.int32).max
[…]
