+++ src/baderkit/core/bader/bader.py
@@ -23,6 +23,7 @@ from .methods.shared_numba import (
     get_persistence_groups,
     get_persistence_cutoffs
 )
+from baderkit.core.utilities.betti_numbers import get_all_betti_numbers_scanning
 
 # This allows for Self typing and is compatible with python 3.10
 Self = TypeVar("Self", bound="Bader")
@@ -431,30 +432,24 @@ class Bader(BaseAnalysis):
         if self._maxima_persistence_values is None:
             # get groups
             maxima_groups = self.maxima_voxel_groups
-            persistence_values = get_persistence_cutoffs(
-                groups=maxima_groups,
-                group_vals=self.maxima_ref_values,
-                data=self.reference_grid.total,
-                use_minima=False,
-                )
-            # maxima_values = self.maxima_ref_values
-            # # get the lowest value that the maximum would connect to with the
-            # # current persistence tol
-            # persistence_values = []
-            # for group, max_val in zip(maxima_groups, maxima_values):
-            #     group_vals = self.reference_grid.total[
-            #         group[:,0],
-            #         group[:,1],
-            #         group[:,2],
-            #         ]
-            #     valid_mask = ((max_val - group_vals) / group_vals) < self.maxima_persistence_tol
-            #     best_val = group_vals[valid_mask].min()
-            #     # get lowest possible persistence below this value
-            #     # (max_val - val) / val < persistence_tol
-            #     # --> val = max_val / (1+persistence_tol)
-            #     persistence_values.append(
-            #         best_val / (1+self.maxima_persistence_tol)
-            #         )
+            maxima_values = self.maxima_ref_values
+            # get the lowest value that the maximum would connect to with the
+            # current persistence tol
+            persistence_values = []
+            for group, max_val in zip(maxima_groups, maxima_values):
+                group_vals = self.reference_grid.total[
+                    group[:,0],
+                    group[:,1],
+                    group[:,2],
+                    ]
+                valid_mask = ((max_val - group_vals) / group_vals) < self.maxima_persistence_tol
+                best_val = group_vals[valid_mask].min()
+                # get lowest possible persistence below this value
+                # (max_val - val) / val < persistence_tol
+                # --> val = max_val / (1+persistence_tol)
+                persistence_values.append(
+                    best_val / (1+self.maxima_persistence_tol)
+                    )
 
             self._maxima_persistence_values = np.array(persistence_values)
         return self._maxima_persistence_values
@@ -762,30 +757,24 @@ class Bader(BaseAnalysis):
             # self._run_minima_bader()
             # get groups
             minima_groups = self.minima_voxel_groups
-            persistence_values = get_persistence_cutoffs(
-                groups=minima_groups,
-                group_vals=self.minima_ref_values,
-                data=self.reference_grid.total,
-                use_minima=True,
-                )
-            # minima_values = self.minima_ref_values
-            # # get the lowest value that the maximum would connect to with the
-            # # current persistence tol
-            # persistence_values = []
-            # for group, min_val in zip(minima_groups, minima_values):
-            #     group_vals = self.reference_grid.total[
-            #         group[:,0],
-            #         group[:,1],
-            #         group[:,2],
-            #         ]
-            #     valid_mask = ((group_vals - min_val) / group_vals) < self.minima_persistence_tol
-            #     best_val = group_vals[valid_mask].max()
-            #     # get lowest possible persistence below this value
-            #     # (val - max_val) / val < persistence_tol
-            #     # --> val = max_val / (1+persistence_tol)
-            #     persistence_values.append(
-            #         best_val / (1-self.minima_persistence_tol)
-            #         )
+            minima_values = self.minima_ref_values
[…]
