+++ src/baderkit/core/utilities/betti_numbers.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
+
 from numba import njit, prange
 import numpy as np
 
 from baderkit.core.utilities.union_find import union, find_root
-from baderkit.core.utilities.basic import coords_to_flat
+from baderkit.core.utilities.basic import coords_to_flat, flat_to_coords
 
 # Full 26-neighborhood (excluding (0,0,0))
 NEIGHBOR_TRANSFORMS = np.array(
@@ -15,8 +16,23 @@ NEIGHBOR_TRANSFORMS = np.array(
     dtype=np.int8
 )
 
+NEIGHBOR_TRANSFORMS1 = NEIGHBOR_TRANSFORMS+1
+
+FACE_TRANSFORMS = np.array([
+    [1,0,0],
+    [0,1,0],
+    [0,0,1],
+    [-1,0,0],
+    [0,-1,0],
+    [0,0,-1],
+    ], dtype=np.int8)
+
+###############################################################################
+# Betti Numbers on Mask
+###############################################################################
+
 @njit(cache=True)
-def get_betty0(vol):
+def get_betti0(vol):
     nx, ny, nz = vol.shape
     num_vox = nx * ny * nz
     ny_nz = ny * nz
@@ -78,23 +94,14 @@ def get_betty0(vol):
     # Count unique connected components, excluding empty space
     return len(np.unique(roots)) - 1, roots
 
-FACE_TRANSFORMS = np.array([
-    [1,0,0],
-    [0,1,0],
-    [0,0,1],
-    [-1,0,0],
-    [0,-1,0],
-    [0,0,-1],
-    ], dtype=np.int8)
-
 @njit(cache=True)
-def get_betty2(vol):
+def get_betti2(vol):
     nx, ny, nz = vol.shape
-    num_vox = nx * ny * nz
+    num_vol = nx * ny * nz
     ny_nz = ny * nz
 
     flat_vol = vol.ravel()
-    labels = np.arange(num_vox, dtype=np.uint32)
+    labels = np.arange(num_vol, dtype=np.uint32)
 
     no_vol = True
 
@@ -136,21 +143,21 @@ def get_betty2(vol):
                     union(labels, label, labels[neigh_flat])
 
     if no_vol:
-        return 0
+        return 0, np.zeros_like(labels, dtype=np.uint32)
 
-    # Compress roots only for occupied voxels
+    # Compress roots only for occupied volels
     roots = np.empty_like(labels, dtype=np.uint32)
 
     for idx, is_vol in enumerate(flat_vol):
         if is_vol:
-            roots[idx] = num_vox
+            roots[idx] = num_vol
         else:
             roots[idx] = find_root(labels, labels[idx])
 
     # we want to return the number of unique roots that are not part of the
     # volume or the edges. By construction the edges are fully connected, and
     # the volume has 1 label, so we subtract 2.
-    return len(np.unique(roots)) - 2
+    return len(np.unique(roots)) - 2, roots
 
 @njit(cache=True)
 def get_euler_characteristic(vol):
@@ -295,34 +302,11 @@ def get_euler_characteristic(vol):
 
     return N0 - N1 + N2 - N3
 
+
 @njit(cache=True)
 def get_betti_numbers(vol):
-    b0, flat_labels = get_betty0(vol)
-    if b0 > 1:
-        nx, ny, nz = vol.shape
-        num_vox = nx * ny * nz
-        # take the largest portion and run again
-        a = np.sort(flat_labels)
-        n = len(a)
-        
-        unique = [a[0]]
-        counts = [1]
-        
-        for i in range(1, n):
-            # skip background
-            if a[i] == num_vox:
-                continue
-            if a[i] == a[i - 1]:
[…]
