### 2026-03-04 cc160c0
small updates for group meeting
