+++ src/baderkit/core/utilities/file_parsers.py
@@ -6,12 +6,13 @@ import math
 import mmap
 from enum import Enum
 from pathlib import Path
+import warnings
 
 import numpy as np
 from numba import njit
 from numpy.typing import NDArray
 from pymatgen.core import Lattice, Structure
-from pymatgen.io.vasp import Poscar
+from pymatgen.io.vasp import Poscar, Potcar
 
 
 # We list all available import/export formats that we've worked on for consistency
@@ -158,8 +159,47 @@ def read_vasp(filename, total_only: bool):
             * scale
         )
 
-        atom_types = f.readline().split()
-        atom_counts = list(map(int, f.readline().split()))
+        # Read the next line and check if we skipped species
+        line = f.readline().split()
+
+        def is_int(s):
+            try:
+                int(s)
+                return True
+            except ValueError:
+                return False
+
+        if all(is_int(x) for x in line):
+            logging.info("No species found in file. Trying to read from POTCAR")
+            # This is the atom-count line. We need to get our species  from
+            # the POTCAR
+            atom_counts = list(map(int, line))
+            potcar_path = path.parent / "POTCAR"
+            if not potcar_path.exists():
+                logging.warning(
+                    "No POTCAR found in directory. Atoms will be set as dummy atoms."
+                )
+                atom_types = [f"Z{i}" for i in range(len(atom_counts))]
+            else:
+                # try:
+                with warnings.catch_warnings(action="ignore"):
+                    potcar = Potcar.from_file(path.parent / "POTCAR")
+                atom_types = potcar.symbols
+                # except:
+                #     logging.warning("POTCAR could not be read, likely due to being malformed. Atoms will be set as dummy atoms.")
+                #     atom_types = [f"Z{i}" for i in range(len(atom_counts))]
+                #     breakpoint()
+
+        else:
+            # This is the species line
+            atom_types = line
+            atom_counts = list(map(int, f.readline().split()))
+
+        if len(atom_types) != len(atom_counts):
+            raise RuntimeError(
+                f"Mismatch: {len(atom_types)} species but {len(atom_counts)} atom counts"
+            )
+
         total_atoms = sum(atom_counts)
 
         # Skip the 'Direct' or 'Cartesian' line
