### 2026-03-07 7f79bae
small bug in gui
