### 2026-03-08 a914480
fix index bug in badelf dim
