### 2026-03-09 0065414
crit finder update
