+++ src/baderkit/core/critical_points/critical_points_numba.py
@@ -1,0 +1,1327 @@
+# -*- coding: utf-8 -*-
+from baderkit.core.utilities.basic import (
+    wrap_point,
+    wrap_point_w_shift, 
+    flat_to_coords
+    )
+
+import itertools
+import numpy as np
+from numpy.typing import NDArray
+from numba import njit, prange
+
+def get_trans_maps() -> tuple:
+    """
+
+    Returns
+    -------
+    tuple
+        A tuple with two parts:
+            trans_to_int
+                An array which maps the 6 unit vector transforms to a single
+                integer representation
+            int_to_trans
+                An array which maps the integer representation of a 6 unit vector
+                to the 3 entry transformation
+
+    """
+    trans_to_int = np.zeros((3, 3, 3), dtype=np.int8)
+    int_to_trans = np.zeros((6, 3), dtype=np.int8)
+    for i in range(6):
+        j = i // 2
+        value = (-1) ** (i & 1)  # -1 for even, 1 for odd
+        int_to_trans[i, j] = value
+        x, y, z = [0 if k != j else value for k in (0, 1, 2)]
+        trans_to_int[x, y, z] = i
+    return trans_to_int, int_to_trans
+
+
+TRANS_TO_INT, INT_TO_TRANS = get_trans_maps()
+
+IMAGE_TO_INT = np.empty([3,3,3], dtype=np.int64)
+INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
+for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
+    IMAGE_TO_INT[i,j,k] = shift_idx
+
+FACE_TRANSFORMS = np.array([
+    [1,0,0],
+    [-1,0,0],
+    [0,1,0],
+    [0,-1,0],
+    [0,0,1],
+    [0,0,-1],
+    ], dtype=np.int64)
+
+# PARITIES = np.array(
+#     [
+#         [0, 0, 0],  # 3D
+#         [0, 0, 1],  # 2D
+#         [0, 1, 0],  # 2D
+#         [0, 1, 1],  # 1D
+#         [1, 0, 0],  # 2D
+#         [1, 0, 1],  # 1D
+#         [1, 1, 0],  # 1D
+#         [1, 1, 1],  # 0D
+#     ],
+#     dtype=np.uint8,
+# )
+# PARITY_DIMS = np.array([3, 2, 2, 1, 2, 1, 1, 0], dtype=np.uint8)
+PARITIES = np.array(
+    [
+        [0, 0, 0],  # 0D
+        [0, 0, 1],  # 1D
+        [0, 1, 0],  # 1D
+        [0, 1, 1],  # 2D
+        [1, 0, 0],  # 1D
+        [1, 0, 1],  # 2D
+        [1, 1, 0],  # 2D
+        [1, 1, 1],  # 3D
+    ],
+    dtype=np.uint8,
+)
+PARITY_DIMS = np.array([0, 1, 1, 2, 1, 2, 2, 3], dtype=np.uint8)
+
+# Conversion from parity to its index
+PARITY_TO_INT = np.zeros((2, 2, 2), dtype=np.int8)
+for idx, (i, j, k) in enumerate(PARITIES):
+    PARITY_TO_INT[i, j, k] = idx
+    
+# Generate transformations to vertices for each parity type.
+CUBE_VERTICES = [
+    [-1, -1, -1],
+    [1, -1, -1],
+    [1, 1, -1],
+    [-1, 1, -1],
+    # [-1, -1, 1],
+    # [1, -1, 1],
+    # [1, 1, 1],
+    # [-1, 1, 1],
+]
+POLY_VERTICES = [
+    [-1, -1],
+    [1, -1],
[…]
