+++ src/baderkit/core/critical_points/hessian_based.py
@@ -1,0 +1,688 @@
+# -*- coding: utf-8 -*-
+import itertools
+
+from numba import njit, prange
+import numpy as np
+from numpy.typing import NDArray
+from baderkit.core.utilities.basic import coords_to_flat, get_norm, wrap_point_w_shift
+from baderkit.core.utilities.interpolation import (
+    newton_refine_critical, 
+    spline_grad_cart, 
+    spline_hess_cart, 
+    spline_grad_and_hess,
+    
+    )
+
+IMAGE_TO_INT = np.empty([3,3,3], dtype=np.int64)
+INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
+for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
+    IMAGE_TO_INT[i,j,k] = shift_idx
+
+@njit(inline='always', cache=True)
+def get_differing_neighs(
+    i, j, k,
+    nx, ny, nz,
+    labels,
+    images,
+    neighbor_transforms,
+    vacuum_mask,
+):
+    # get the label at this point
+    label0 = labels[i,j,k]
+    image0 = images[i,j,k]
+
+    # initialize potential alternative labels
+    label1 = -1; image1 = -1
+    unique = 0
+
+    # iterate over transforms
+    for trans in range(neighbor_transforms.shape[0]):
+        # if we've found more than two neighbors, immediately break
+        if unique == 2:
+            break
+        
+        # get shifts
+        si = neighbor_transforms[trans, 0]
+        sj = neighbor_transforms[trans, 1]
+        sk = neighbor_transforms[trans, 2]
+
+        # wrap around periodic edges and store shift
+        ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(
+            i+si, j+sj, k+sk, nx, ny, nz
+        )
+        
+        # skip points in the vacuum
+        if vacuum_mask[ii, jj, kk]:
+            continue
+        
+        # get the label and image of this neighbor
+        neigh_label = labels[ii, jj, kk]
+        neigh_image = images[ii, jj, kk]
+
+        # update image to be relative to the current points transformation
+        si1 = INT_TO_IMAGE[neigh_image, 0] + ssi
+        sj1 = INT_TO_IMAGE[neigh_image, 1] + ssj
+        sk1 = INT_TO_IMAGE[neigh_image, 2] + ssk
+        neigh_image = IMAGE_TO_INT[si1, sj1, sk1]
+
+        # compare to any previous labels and update our unique number
+        if unique == 0:
+            if neigh_label != label0 or neigh_image != image0:
+                label1 = neigh_label
+                image1 = neigh_image
+                unique = 1
+        elif unique == 1:
+            if ((neigh_label != label0 or neigh_image != image0) and
+                (neigh_label != label1 or neigh_image != image1)):
+                unique = 2
+
+    return unique
+
+@njit(parallel=True,  cache=True)
+def get_manifold_labels(
+    maxima_labels: NDArray[np.int64],
+    minima_labels: NDArray[np.int64],
+    maxima_images: NDArray[np.int64],
+    minima_images: NDArray[np.int64],
+    maxima_groups: list[NDArray],
+    minima_groups: list[NDArray],
+    neighbor_transforms: NDArray[np.int64],
+    vacuum_mask: NDArray[np.bool_],
+):
+    """
+    Takes the 3-manifolds of maxima and minima and determines the rough locations
+    of the following manifolds:
+        
+        0: minima
+        1: 1-saddle
+        2: 2-saddle
[…]

[…]
