+++ src/baderkit/core/elf_analysis/bifurcation_graph/graph_temp.py
@@ -1,1065 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import itertools
-import json
-import logging
-import time
-
-import numpy as np
-import plotly.graph_objects as go
-from numpy.typing import NDArray
-from pymatgen.analysis.local_env import CrystalNN
-
-from baderkit.core import Bader, Structure, Grid
-from baderkit.core.bader.methods.shared_numba import get_edges
-
-from .enum_and_styling import LINE_COLOR, DomainSubtype, FeatureType
-
-from baderkit.core.base.base_analysis import BaseAnalysis
-
-# from elf_analyzer.core.utilities import IonicRadiiTools
-from baderkit.core.bader import Bader
-from .graph_numba import (
-    find_domain_bifurcations,
-    find_domain_connections,
-    find_potential_saddle_points,
-    get_domains_surrounding_atoms,
-)
-from .nodes import IrreducibleNode, NodeBase, ReducibleNode
-
-
-class BifurcationGraph(BaseAnalysis):
-    """
-    A convenience class for storing the nodes of a bifurcation graph and gathering
-    data on them.
-
-    Parameters
-    ----------
-    structure : Structure
-        The PyMatGen structure of the chemical system.
-    basin_maxima_frac : NDArray[float]
-        The fractional coordinates of the Bader basins in the system.
-    basin_charges : NDArray[float]
-        The integrated charges of the Bader basins in the system.
-    basin_volumes : NDArray[float]
-        The volumes of the Bader basins in the system.
-    crystalnn_kwargs : dict
-        The keyword arguments for the CrystalNN object used for finding neighbors.
-        Particularly important when calculating charges assigned to each atom.
-    atomic_radii : NDArray[float], optional
-        The radii of each atom in the system used to calculate distance beyond
-        the atom for each feature. The default is None.
-    """
-
-    _reset_props = ["bifurcations"]
-
-    def __init__(
-        self,
-        reference_grid: Grid,
-        remove_cutoff: float = 0.05,
-        reduce_cutoff: float = 0.10,
-        calculate_surrounded_atoms: bool = False,
-        **kwargs,
-    ):
-        super().__init__(reference_grid=reference_grid, **kwargs)
-
-        # create bader object with ELF reference
-        self.bader = Bader(
-            charge_grid=self.charge_grid,
-            total_charge_grid=self.total_charge_grid,
-            reference_grid=reference_grid,
-            **kwargs,
-        )
-
-        self._calculate_surrounded_atoms = calculate_surrounded_atoms
-
-        self._root_nodes = []
-        self._nodes = []
-        self._node_keys = {}
-
-        self._remove_cutoff = remove_cutoff
-        self._reduce_cutoff = reduce_cutoff
-
-    ###########################################################################
-    # Base Operations
-    ###########################################################################
-
-    def __iter__(self):
-        return iter(self.nodes)
-
-    def __len__(self):
-        return len(self.nodes)
-
-    def __getitem__(self, key):
-        return self.nodes[key]
-
-    def __contains__(self, key):
-        return key in self.nodes
-
-    def __repr__(self):
-        return f"BifurcationGraph(num_nodes={len(self._nodes)})"
-
-    ###########################################################################
-    # Set Properties
-    ###########################################################################
-
-    @property
-    def root_nodes(self) -> list:
-        """
-
-        Returns
-        -------
-        list
-            A list of nodes that have no parents
-
-        """
-        return self._root_nodes
-
-    @property
-    def nodes(self) -> list:
-        """
-
-        Returns
-        -------
-        list
-            The list of nodes in the graph.
-
-        """
-        return self._nodes
-
-    def node_from_key(self, key: int) -> list:
-        """
[…]
