+++ src/baderkit/core/elf_analysis/bifurcation_graph/graph.py
@@ -1,19 +1,19 @@
 # -*- coding: utf-8 -*-
 
 import itertools
-import json
 import logging
 import time
 
 import numpy as np
 import plotly.graph_objects as go
 from numpy.typing import NDArray
-from pymatgen.analysis.local_env import CrystalNN
 
-from baderkit.core import Bader, Structure
+from baderkit.core import Bader, Structure, Grid
 from baderkit.core.bader.methods.shared_numba import get_edges
 
 from .enum_and_styling import LINE_COLOR, DomainSubtype, FeatureType
+
+from baderkit.core.base.base_analysis import BaseAnalysis
 
 # from elf_analyzer.core.utilities import IonicRadiiTools
 from .graph_numba import (
@@ -22,59 +22,57 @@ from .graph_numba import (
     find_potential_saddle_points,
     get_domains_surrounding_atoms,
 )
-from .nodes import IrreducibleNode, NodeBase, ReducibleNode
+from .nodes import IrreducibleNode, ReducibleNode
 
 
-class BifurcationGraph:
+class BifurcationGraph(BaseAnalysis):
     """
     A convenience class for storing the nodes of a bifurcation graph and gathering
     data on them.
 
-    The nodes themselves contain the information on their connectivity.
-
     Parameters
     ----------
-    structure : Structure
-        The PyMatGen structure of the chemical system.
-    labeler_type : str
-        The type of labeler used to mark irreducible features in the graph.
-    basin_maxima_frac : NDArray[float]
-        The fractional coordinates of the Bader basins in the system.
-    basin_charges : NDArray[float]
-        The integrated charges of the Bader basins in the system.
-    basin_volumes : NDArray[float]
-        The volumes of the Bader basins in the system.
-    crystalnn_kwargs : dict
-        The keyword arguments for the CrystalNN object used for finding neighbors.
-        Particularly important when calculating charges assigned to each atom.
-    atomic_radii : NDArray[float], optional
-        The radii of each atom in the system used to calculate distance beyond
-        the atom for each feature. The default is None.
+
     """
+
+    _reset_props = ["bifurcations"]
 
     def __init__(
         self,
-        structure: Structure,
-        labeler_type: str,
-        basin_maxima_frac: NDArray[float],
-        basin_charges: NDArray[float],
-        basin_volumes: NDArray[float],
-        crystalnn_kwargs: dict,
-        atomic_radii: NDArray[float] = None,
+        charge_grid: Grid,
+        reference_grid: Grid,
+        total_charge_grid: Grid = None,
+        remove_cutoff: float = 0.05,
+        reduce_cutoff: float = 0.10,
+        calculate_surrounded_atoms: bool = True,
         **kwargs,
     ):
+        super().__init__(
+            charge_grid=charge_grid,
+            total_charge_grid=total_charge_grid,
+            reference_grid=reference_grid,            
+            **kwargs)
+
+        # create bader object with ELF reference
+        self.bader = Bader(
+            charge_grid=charge_grid,
+            total_charge_grid=total_charge_grid,
+            reference_grid=reference_grid,
+            **kwargs,
+        )
+
+        self._calculate_surrounded_atoms = calculate_surrounded_atoms
 
         self._root_nodes = []
         self._nodes = []
         self._node_keys = {}
 
-        self.structure = structure
-        self.labeler_type = labeler_type
-        self.basin_maxima_frac = basin_maxima_frac
-        self.basin_charges = basin_charges
-        self.basin_volumes = basin_volumes
-        self.crystalnn_kwargs = crystalnn_kwargs
-        self.cnn = CrystalNN(**crystalnn_kwargs)
+        self._remove_cutoff = remove_cutoff
+        self._reduce_cutoff = reduce_cutoff
+
[…]
