+++ src/baderkit/core/elf_analysis/__init__.py
@@ -1,3 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from .elf_labeler import ElfLabeler, SpinElfLabeler
+from .bifurcation_graph import BifurcationGraph # isort:skip
+
+from .elf_labeler import ElfLabeler, SpinElfLabeler # isort:skip
+
+from .badelf import Badelf, SpinBadelf # isort:skip
