+++ src/baderkit/core/critical_points/hessian_based.py
@@ -1,38 +1,44 @@
 # -*- coding: utf-8 -*-
 import itertools
 
-from numba import njit, prange
 import numpy as np
+from numba import njit, prange
 from numpy.typing import NDArray
+
 from baderkit.core.utilities.basic import coords_to_flat, get_norm, wrap_point_w_shift
 from baderkit.core.utilities.interpolation import (
-    newton_refine_critical, 
-    spline_grad_cart, 
-    spline_hess_cart, 
+    newton_refine_critical,
     spline_grad_and_hess,
-    
-    )
+    spline_grad_cart,
+    spline_hess_cart,
+)
 
-IMAGE_TO_INT = np.empty([3,3,3], dtype=np.int64)
-INT_TO_IMAGE = np.array(list(itertools.product((-1,0,1), repeat=3)))
-for shift_idx, (i,j,k) in enumerate(INT_TO_IMAGE):
-    IMAGE_TO_INT[i,j,k] = shift_idx
+IMAGE_TO_INT = np.empty([3, 3, 3], dtype=np.int64)
+INT_TO_IMAGE = np.array(list(itertools.product((-1, 0, 1), repeat=3)))
+for shift_idx, (i, j, k) in enumerate(INT_TO_IMAGE):
+    IMAGE_TO_INT[i, j, k] = shift_idx
 
-@njit(inline='always', cache=True)
+
+@njit(inline="always", cache=True)
 def get_differing_neighs(
-    i, j, k,
-    nx, ny, nz,
+    i,
+    j,
+    k,
+    nx,
+    ny,
+    nz,
     labels,
     images,
     neighbor_transforms,
     vacuum_mask,
 ):
     # get the label at this point
-    label0 = labels[i,j,k]
-    image0 = images[i,j,k]
+    label0 = labels[i, j, k]
+    image0 = images[i, j, k]
 
     # initialize potential alternative labels
-    label1 = -1; image1 = -1
+    label1 = -1
+    image1 = -1
     unique = 0
 
     # iterate over transforms
@@ -40,7 +46,7 @@ def get_differing_neighs(
         # if we've found more than two neighbors, immediately break
         if unique == 2:
             break
-        
+
         # get shifts
         si = neighbor_transforms[trans, 0]
         sj = neighbor_transforms[trans, 1]
@@ -48,13 +54,13 @@ def get_differing_neighs(
 
         # wrap around periodic edges and store shift
         ii, jj, kk, ssi, ssj, ssk = wrap_point_w_shift(
-            i+si, j+sj, k+sk, nx, ny, nz
+            i + si, j + sj, k + sk, nx, ny, nz
         )
-        
+
         # skip points in the vacuum
         if vacuum_mask[ii, jj, kk]:
             continue
-        
+
         # get the label and image of this neighbor
         neigh_label = labels[ii, jj, kk]
         neigh_image = images[ii, jj, kk]
@@ -72,13 +78,15 @@ def get_differing_neighs(
                 image1 = neigh_image
                 unique = 1
         elif unique == 1:
-            if ((neigh_label != label0 or neigh_image != image0) and
-                (neigh_label != label1 or neigh_image != image1)):
+            if (neigh_label != label0 or neigh_image != image0) and (
+                neigh_label != label1 or neigh_image != image1
+            ):
                 unique = 2
 
     return unique
 
-@njit(parallel=True,  cache=True)
+
+@njit(parallel=True, cache=True)
 def get_manifold_labels(
     maxima_labels: NDArray[np.int64],
     minima_labels: NDArray[np.int64],
@@ -92,7 +100,7 @@ def get_manifold_labels(
     """
     Takes the 3-manifolds of maxima and minima and determines the rough locations
     of the following manifolds:
[…]
