+++ src/baderkit/core/utilities/persistence.py
@@ -6,21 +6,36 @@ import math
 import numpy as np
 from numba import njit, prange
 from numpy.typing import NDArray
+from numba.typed import Dict, List
+from numba.types import UniTuple, int16, int64
 
 from baderkit.core.utilities.basic import (
     dist,
+    wrap_point,
     wrap_point_w_shift,
 )
-from baderkit.core.utilities.interpolation import linear_slice
+from baderkit.core.utilities.interpolation import linear_slice, spline_hess
 from baderkit.core.utilities.union_find import find_root, union
 
 IMAGE_TO_INT = np.empty([3, 3, 3], dtype=np.int64)
 INT_TO_IMAGE = np.array(list(itertools.product((-1, 0, 1), repeat=3)))
 for shift_idx, (i, j, k) in enumerate(INT_TO_IMAGE):
     IMAGE_TO_INT[i, j, k] = shift_idx
+    
+FACE_TRANSFORMS = np.array(
+    [
+        [1, 0, 0],
+        [-1, 0, 0],
+        [0, 1, 0],
+        [0, -1, 0],
+        [0, 0, 1],
+        [0, 0, -1],
+    ],
+    dtype=np.int64,
+)
 
 
-# @njit(cache=True)
+@njit(cache=True)
 def compute_wrap_offset(point1, point2):
     """
     Computes wrap from point1 to point2
@@ -47,8 +62,67 @@ def compute_wrap_offset(point1, point2):
 
     return best_i, best_j, best_k
 
+# Predefine types
+key_type = UniTuple(int64, 3)
+value_type = int64
 
-# @njit(cache=True)
+@njit(cache=True)
+def unique_and_inverse_axis0(arr):
+    n = arr.shape[0]
+    arr64 = arr.astype(np.int64)
+
+    # Use the predeclared Numba types
+    d = Dict.empty(key_type=key_type, value_type=value_type)
+
+    inverse = np.empty(n, dtype=np.int64)
+    unique = List.empty_list(key_type)
+
+    next_idx = 0
+
+    for i in range(n):
+        key = (arr64[i,0], arr64[i,1], arr64[i,2])
+
+        if key in d:
+            inverse[i] = d[key]
+        else:
+            d[key] = next_idx
+            inverse[i] = next_idx
+            unique.append(key)
+            next_idx += 1
+
+    unique_arr = np.empty((next_idx, 3), dtype=arr.dtype)
+
+    for i in range(next_idx):
+        row = unique[i]
+        unique_arr[i,0] = row[0]
+        unique_arr[i,1] = row[1]
+        unique_arr[i,2] = row[2]
+
+    return unique_arr, inverse
+
+@njit(cache=True)
+def get_voxel_min_max(data, i, j, k, nx, ny, nz):
+    """
+    Estimate min/max value inside the voxel centered at (i,j,k)
+    using the center + 6 face centers (7 samples total).
+    """
+
+    c = data[i, j, k]
+
+    # set initial min/max
+    vmin = c
+    vmax = c
+    
+    for si, sj, sk in FACE_TRANSFORMS:
+        ni, nj, nk = wrap_point(i+si, j+sj, k+sk, nx, ny, nz)
+        v = 0.5 * (c + data[ni, nj, nk])
+        if v < vmin: vmin = v
+        if v > vmax: vmax = v
+
+    return vmin, vmax
+
+
+@njit(cache=True)
 def get_persistence_value(value1, value2, conn_value, p1, p2, p_conn):
     eps = 1e-12
     # get larger and smaller diffs
@@ -56,17 +130,21 @@ def get_persistence_value(value1, value2, conn_value, p1, p2, p_conn):
[…]
