### 2026-03-12 3597848
mostly working extrema combination
