+++ src/baderkit/core/utilities/persistence.py
@@ -4,7 +4,7 @@ import itertools
 import math
 
 import numpy as np
-from numba import njit, prange
+from numba import njit, prange, types
 from numpy.typing import NDArray
 from numba.typed import Dict, List
 from numba.types import UniTuple, int16, int64
@@ -21,7 +21,7 @@ IMAGE_TO_INT = np.empty([3, 3, 3], dtype=np.int64)
 INT_TO_IMAGE = np.array(list(itertools.product((-1, 0, 1), repeat=3)))
 for shift_idx, (i, j, k) in enumerate(INT_TO_IMAGE):
     IMAGE_TO_INT[i, j, k] = shift_idx
-    
+
 FACE_TRANSFORMS = np.array(
     [
         [1, 0, 0],
@@ -62,9 +62,11 @@ def compute_wrap_offset(point1, point2):
 
     return best_i, best_j, best_k
 
+
 # Predefine types
 key_type = UniTuple(int64, 3)
 value_type = int64
+
 
 @njit(cache=True)
 def unique_and_inverse_axis0(arr):
@@ -80,7 +82,7 @@ def unique_and_inverse_axis0(arr):
     next_idx = 0
 
     for i in range(n):
-        key = (arr64[i,0], arr64[i,1], arr64[i,2])
+        key = (arr64[i, 0], arr64[i, 1], arr64[i, 2])
 
         if key in d:
             inverse[i] = d[key]
@@ -94,33 +96,11 @@ def unique_and_inverse_axis0(arr):
 
     for i in range(next_idx):
         row = unique[i]
-        unique_arr[i,0] = row[0]
-        unique_arr[i,1] = row[1]
-        unique_arr[i,2] = row[2]
+        unique_arr[i, 0] = row[0]
+        unique_arr[i, 1] = row[1]
+        unique_arr[i, 2] = row[2]
 
     return unique_arr, inverse
-
-@njit(cache=True)
-def get_voxel_min_max(data, i, j, k, nx, ny, nz):
-    """
-    Estimate min/max value inside the voxel centered at (i,j,k)
-    using the center + 6 face centers (7 samples total).
-    """
-
-    c = data[i, j, k]
-
-    # set initial min/max
-    vmin = c
-    vmax = c
-    
-    for si, sj, sk in FACE_TRANSFORMS:
-        ni, nj, nk = wrap_point(i+si, j+sj, k+sk, nx, ny, nz)
-        v = 0.5 * (c + data[ni, nj, nk])
-        if v < vmin: vmin = v
-        if v > vmax: vmax = v
-
-    return vmin, vmax
-
 
 @njit(cache=True)
 def get_persistence_value(value1, value2, conn_value, p1, p2, p_conn):
@@ -133,52 +113,163 @@ def get_persistence_value(value1, value2, conn_value, p1, p2, p_conn):
     p1_dist = dist(p1, p_conn)
     p2_dist = dist(p2, p_conn)
     distance = p1_dist + p2_dist
-    
+
     # get approximate average value between extrema
-    avg = (p1_dist*(value1+conn_value)/2 + p2_dist*(value2+conn_value)/2) / distance
+    avg = (
+        p1_dist * (value1 + conn_value) / 2
+        + p2_dist * (value2 + conn_value) / 2
+    ) / distance
 
     # get persistence score:
     #   p = smaller_diff*dist / (average_value)
-    persistence_score = min(diff1, diff2) * distance / (abs(avg)+eps)
+    persistence_score = min(diff1, diff2) * distance / (abs(avg) + eps)
     # print(persistence_score)
     return persistence_score
 
+@njit(cache=True)
[…]
