+++ src/baderkit/core/bader/methods/shared_numba.py
@@ -13,7 +13,7 @@ from baderkit.core.utilities.basic import (
     wrap_point,
     wrap_point_w_shift,
 )
-from baderkit.core.utilities.interpolation import linear_slice
+from baderkit.core.utilities.interpolation import linear_slice, refine_extrema
 from baderkit.core.utilities.persistence import (
     compute_wrap_offset,
     get_conn_vals,
@@ -436,6 +436,7 @@ def get_basin_edges(
 
     return edges
 
+
 @njit(inline="always", cache=True)
 def get_differing_neighs_thin(
     i,
@@ -480,12 +481,14 @@ def get_differing_neighs_thin(
         # skip points in the vacuum
         if vacuum_mask[ii, jj, kk]:
             continue
-        
+
         # skip points with a higher value
         if (
-            use_minima and data[ii,jj,kk]>value0
-            or not use_minima and data[ii,jj,kk]<value0
-            ):
+            use_minima
+            and data[ii, jj, kk] > value0
+            or not use_minima
+            and data[ii, jj, kk] < value0
+        ):
             continue
 
         # get the label and image of this neighbor
@@ -511,6 +514,7 @@ def get_differing_neighs_thin(
                 unique = 2
 
     return unique
+
 
 @njit(parallel=True, cache=True)
 def get_thin_basin_edges(
@@ -562,6 +566,7 @@ def get_thin_basin_edges(
 
     return edges
 
+
 @njit(cache=True)
 def get_neighboring_basin_surface_area(
     labeled_array: NDArray[np.int64],
@@ -611,7 +616,9 @@ def get_neighboring_basin_surface_area(
                 # get this voxels label
                 label = labeled_array[i, j, k]
                 # iterate over the neighboring voxels
-                for (si, sj, sk), area in zip(neighbor_transforms, neighbor_areas):
+                for (si, sj, sk), area in zip(
+                    neighbor_transforms, neighbor_areas
+                ):
                     # wrap points
                     ii, jj, kk = wrap_point(i + si, j + sj, k + sk, nx, ny, nz)
                     # get neighbors label
@@ -716,31 +723,77 @@ def get_extrema(
                     extrema[i, j, k] = True
     return extrema
 
-
 @njit(cache=True)
+def reorder_labels(
+    labels,
+    data,
+    extrema_vox,
+    extrema_labels,
+    extrema_values,
+    use_minima,
+        ):
+    nx, ny, nz = data.shape
+    ny_nz = ny*nz
+
+    roots = np.empty(len(extrema_vox), dtype=labels.dtype)
+    for ext_idx in range(len(extrema_vox)):
+        label = extrema_labels[ext_idx]
+        root = find_root(labels, label)
+        roots[ext_idx] = root
+
+    unique_roots = np.unique(roots)
+    groups = []
+    for root_label in unique_roots:
+        i, j, k = flat_to_coords(root_label, ny_nz, nz)
+        group = []
+        best_point = root_label
+        best_value = data[i, j, k]
+        for ext_idx, (label, ext_root) in enumerate(zip(extrema_labels, roots)):
+            if ext_root != root_label:
+                continue
+            group.append(label)
+            value = extrema_values[ext_idx]
+            label = extrema_labels[ext_idx]
+            if (
+                not use_minima
+                and value > best_value
+                or use_minima
+                and value < best_value
+                or value == best_value
+                and label < root_label
+            ):
+                best_point = label
+                best_value = value
+        for label in group:
+            labels[label] = best_point
+        groups.append(group)
+    return labels, groups
+
+# @njit(cache=True)
 def initialize_labels_from_extrema(
     data,
     labels,
-    extrema_frac,
[…]

[…]
