### 2026-03-13 8b033ba
fix bug in basin reduction
