+++ src/baderkit/core/utilities/persistence.py
@@ -1,106 +1,66 @@
 # -*- coding: utf-8 -*-
 
-import itertools
 import math
 
 import numpy as np
-from numba import njit, prange, types
-from numpy.typing import NDArray
-from numba.typed import Dict, List
-from numba.types import UniTuple, int16, int64
+from numba import njit, prange
 
 from baderkit.core.utilities.basic import (
     dist,
     wrap_point,
-    wrap_point_w_shift,
+    coords_to_flat,
+    compute_wrap_offset,
 )
-from baderkit.core.utilities.interpolation import linear_slice, spline_hess
-from baderkit.core.utilities.union_find import find_root, union
-
-IMAGE_TO_INT = np.empty([3, 3, 3], dtype=np.int64)
-INT_TO_IMAGE = np.array(list(itertools.product((-1, 0, 1), repeat=3)))
-for shift_idx, (i, j, k) in enumerate(INT_TO_IMAGE):
-    IMAGE_TO_INT[i, j, k] = shift_idx
-
-FACE_TRANSFORMS = np.array(
-    [
-        [1, 0, 0],
-        [-1, 0, 0],
-        [0, 1, 0],
-        [0, -1, 0],
-        [0, 0, 1],
-        [0, 0, -1],
-    ],
-    dtype=np.int64,
-)
-
-
-@njit(cache=True)
-def compute_wrap_offset(point1, point2):
-    """
-    Computes wrap from point1 to point2
-
-    """
-    best_d2 = np.inf
-    best_i = 0
-    best_j = 0
-    best_k = 0
-
-    for i in (-1, 0, 1):
-        for j in (-1, 0, 1):
-            for k in (-1, 0, 1):
-                dx = (point2[0] + i) - point1[0]
-                dy = (point2[1] + j) - point1[1]
-                dz = (point2[2] + k) - point1[2]
-                d2 = dx * dx + dy * dy + dz * dz
-
-                if d2 < best_d2:
-                    best_d2 = d2
-                    best_i = i
-                    best_j = j
-                    best_k = k
-
-    return best_i, best_j, best_k
-
+from baderkit.core.utilities.interpolation import linear_slice
+from baderkit.core.utilities.union_find import union
+from baderkit.core.utilities.basins import (
+    get_best_neighbor,
+    reorder_labels,
+    )
+from baderkit.core.utilities.critical_points import refine_critical_points
+from baderkit.core.utilities.transforms import (
+    INT_TO_IMAGE,
+    ALL_NEIGHBOR_TRANSFORMS,
+    get_transform_dists
+    )
 
 # Predefine types
-key_type = UniTuple(int64, 3)
-value_type = int64
+# key_type = UniTuple(int64, 3)
+# value_type = int64
 
+# @njit(cache=True)
+# def unique_and_inverse_axis0(arr):
+#     n = arr.shape[0]
+#     arr64 = arr.astype(np.int64)
 
-@njit(cache=True)
-def unique_and_inverse_axis0(arr):
-    n = arr.shape[0]
-    arr64 = arr.astype(np.int64)
+#     # Use the predeclared Numba types
+#     d = Dict.empty(key_type=key_type, value_type=value_type)
 
-    # Use the predeclared Numba types
-    d = Dict.empty(key_type=key_type, value_type=value_type)
+#     inverse = np.empty(n, dtype=np.int64)
+#     unique = List.empty_list(key_type)
 
-    inverse = np.empty(n, dtype=np.int64)
-    unique = List.empty_list(key_type)
+#     next_idx = 0
 
-    next_idx = 0
+#     for i in range(n):
+#         key = (arr64[i, 0], arr64[i, 1], arr64[i, 2])
 
-    for i in range(n):
[…]
