+++ src/baderkit/core/utilities/basins.py
@@ -1,39 +1,21 @@
 # -*- coding: utf-8 -*-
 
-import itertools
-import math
-
 import numpy as np
 from numba import njit, prange
 from numpy.typing import NDArray
 
 from baderkit.core.utilities.basic import (
-    coords_to_flat,
     flat_to_coords,
     wrap_point,
     wrap_point_w_shift,
 )
-from baderkit.core.utilities.interpolation import linear_slice, refine_extrema
-from baderkit.core.utilities.persistence import (
-    compute_wrap_offset,
-    get_conn_vals,
-    reduce_by_conn,
-)
-from baderkit.core.utilities.union_find import (
-    find_root,
-    find_root_with_shift,
-    union,
-    union_with_shift,
-)
+from baderkit.core.utilities.union_find import find_root
+from baderkit.core.utilities.transforms import INT_TO_IMAGE, IMAGE_TO_INT
 
-IMAGE_TO_INT = np.empty([3, 3, 3], dtype=np.int64)
-INT_TO_IMAGE = np.array(list(itertools.product((-1, 0, 1), repeat=3)))
-for shift_idx, (i, j, k) in enumerate(INT_TO_IMAGE):
-    IMAGE_TO_INT[i, j, k] = shift_idx
-###############################################################################
-# General methods
-###############################################################################
 
+###############################################################################
+# Neighbor Finding Methods
+###############################################################################
 
 @njit(cache=True, inline="always")
 def get_best_neighbor(
@@ -196,6 +178,9 @@ def get_best_neighbor_with_shift(
         np.array((bsi, bsj, bsk), dtype=np.int8),
     )
 
+###############################################################################
+# Edge Finding Methods
+###############################################################################
 
 @njit(parallel=True, cache=True)
 def get_edges(
@@ -370,10 +355,14 @@ def get_differing_neighs(
         neigh_image = images[ii, jj, kk]
 
         # update image to be relative to the current points transformation
-        si1 = INT_TO_IMAGE[neigh_image, 0] + ssi
-        sj1 = INT_TO_IMAGE[neigh_image, 1] + ssj
-        sk1 = INT_TO_IMAGE[neigh_image, 2] + ssk
-        neigh_image = IMAGE_TO_INT[si1, sj1, sk1]
+        if ssi == 0 and ssj == 0 and ssk == 0:
+            neigh_image = images[ii, jj, kk]
+        else:
+            neigh_image = images[ii, jj, kk]
+            si1 = INT_TO_IMAGE[neigh_image, 0] + ssi
+            sj1 = INT_TO_IMAGE[neigh_image, 1] + ssj
+            sk1 = INT_TO_IMAGE[neigh_image, 2] + ssk
+            neigh_image = IMAGE_TO_INT[si1, sj1, sk1]
 
         # compare to any previous labels and update our unique number
         if unique == 0:
@@ -436,7 +425,6 @@ def get_basin_edges(
 
     return edges
 
-
 @njit(inline="always", cache=True)
 def get_differing_neighs_thin(
     i,
@@ -450,68 +438,57 @@ def get_differing_neighs_thin(
     images,
     neighbor_transforms,
     vacuum_mask,
-    use_minima=False,
+    use_minima,
 ):
-    # get the label at this point
+
     label0 = labels[i, j, k]
     image0 = images[i, j, k]
     value0 = data[i, j, k]
 
-    # initialize potential alternative labels
     label1 = -1
     image1 = -1
     unique = 0
 
-    # iterate over transforms
-    for trans in range(neighbor_transforms.shape[0]):
-        # if we've found more than two neighbors, immediately break
-        if unique == 2:
-            break
[…]
