+++ src/baderkit/core/bader/methods/neargrid/neargrid_method.py
@@ -5,8 +5,8 @@ import logging
 import numpy as np
 
 from baderkit.core.bader.methods.base import MethodBase
-from baderkit.core.utilities.basins import get_edges
 from baderkit.core.utilities.basic import get_lowest_int, get_lowest_uint
+from baderkit.core.utilities.basins import get_edges_w_flat_images
 
 from .neargrid_numba import (
     get_gradient_pointers_overdetermined,
@@ -65,22 +65,19 @@ class NeargridMethod(MethodBase):
             # get the pseudo inverse
             inv_norm_cart_trans = np.linalg.pinv(norm_cart_transforms[:13])
             # calculate gradients and pointers to best neighbors
-            labels, images, gradients, self._extrema_mask = (
-                get_gradient_pointers_overdetermined(
-                    data=reference_data,
-                    labels=labels,
-                    images=images,
-                    car2lat=self.car2lat,
-                    inv_norm_cart_trans=inv_norm_cart_trans,
-                    neighbor_dists=neighbor_dists,
-                    neighbor_transforms=neighbor_transforms,
-                    vacuum_mask=self.vacuum_mask,
-                    extrema_mask=self.extrema_mask,
-                    use_minima=self.use_minima,
-                )
+            labels, images, gradients = get_gradient_pointers_overdetermined(
+                data=reference_data,
+                labels=labels,
+                images=images,
+                car2lat=self.car2lat,
+                inv_norm_cart_trans=inv_norm_cart_trans,
+                neighbor_dists=neighbor_dists,
+                neighbor_transforms=neighbor_transforms,
+                vacuum_mask=self.vacuum_mask,
+                extrema_mask=self.extrema_mask,
+                use_minima=self.use_minima,
             )
         # Find roots
-        # breakpoint()
         logging.info("Finding Roots")
         labels, images = self.get_roots(labels, images)
 
@@ -96,16 +93,18 @@ class NeargridMethod(MethodBase):
 
         # Now we refine the edges with the neargrid method
         # Get our edges, not including edges on the vacuum.
-        refinement_mask = get_edges(
+        refinement_mask = get_edges_w_flat_images(
             labeled_array=labels,
+            images=images,
             neighbor_transforms=neighbor_transforms,
             vacuum_mask=self.vacuum_mask,
         )
+
         # remove extrema from refinement
         refinement_mask[self.extrema_mask] = False
         # note these labels and the vacuum should not be reassigned again in future cycles
-        labels[refinement_mask & self.vacuum_mask] = -labels[
-            refinement_mask & self.vacuum_mask
+        labels[refinement_mask | self.vacuum_mask] = -labels[
+            refinement_mask | self.vacuum_mask
         ]
         labels, images = refine_fast_neargrid(
             data=reference_data,
