### 2026-03-15 4b71c05
fix periodic bugs in initialization
