+++ src/baderkit/core/utilities/persistence.py
@@ -22,7 +22,11 @@ from baderkit.core.utilities.transforms import (
     INT_TO_IMAGE,
     get_transform_dists,
 )
-from baderkit.core.utilities.union_find import union
+from baderkit.core.utilities.union_find import (
+    find_root_with_shift1,
+    union,
+    union_with_shift1,
+)
 
 # Predefine types
 # key_type = UniTuple(int64, 3)
@@ -108,6 +112,24 @@ def grow_arrays(neighbors, values, conn_coords, neigh_coords):
 
 
 @njit(cache=True)
+def get_shift(coord):
+    shift = np.empty(3, dtype=np.int8)
+    if coord[0] < 0:
+        shift[0] = -1
+    elif coord[0] >= 0:
+        shift[0] = 1
+    if coord[1] < 0:
+        shift[1] = -1
+    elif coord[1] >= 0:
+        shift[1] = 1
+    if coord[2] < 0:
+        shift[2] = -1
+    elif coord[2] >= 0:
+        shift[2] = 1
+    return shift
+
+
+@njit(cache=True)
 def union_by_persistence(
     neighbors,
     values,
@@ -115,6 +137,7 @@ def union_by_persistence(
     neigh_coords,
     cursor,
     labels,
+    matrix,
     extrema_values,
     extrema_labels,
     use_minima,
@@ -141,21 +164,24 @@ def union_by_persistence(
                 continue
 
             if num == 1:
-                neigh = neighbors[ext_idx, 0]
-                union(labels, extrema_labels[ext_idx], extrema_labels[neigh])
+                neigh_idx = neighbors[ext_idx, 0]
+                union(labels, extrema_labels[ext_idx], extrema_labels[neigh_idx])
                 new_union = True
                 continue
 
             for i in range(num):
 
                 neigh_idx = neighbors[ext_idx, i]
+                neigh_val = extrema_values[neigh_idx]
+                # get shift
+                neigh_coord = neigh_coords[ext_idx, i]
+                neigh_cart = neigh_coord @ matrix
 
                 union(labels, extrema_labels[ext_idx], extrema_labels[neigh_idx])
                 new_union = True
 
                 conn_val = values[ext_idx, i]
                 conn_coord = conn_coords[ext_idx, i]
-                neigh_coord = neigh_coords[ext_idx, i]
 
                 neigh_count = cursor[neigh_idx]
 
@@ -165,6 +191,7 @@ def union_by_persistence(
                         continue
 
                     neigh_idx1 = neighbors[ext_idx, j]
+                    neigh_val1 = extrema_values[neigh_idx1]
 
                     exists = False
                     for k in range(neigh_count):
@@ -178,6 +205,7 @@ def union_by_persistence(
                     conn_val1 = values[ext_idx, j]
                     conn_coord1 = conn_coords[ext_idx, j]
                     neigh_coord1 = neigh_coords[ext_idx, j]
+                    neigh_cart1 = neigh_coord1 @ matrix
 
                     if (use_minima and conn_val >= conn_val1) or (
                         not use_minima and conn_val <= conn_val1
@@ -192,27 +220,52 @@ def union_by_persistence(
                         extrema_values[neigh_idx],
                         extrema_values[neigh_idx1],
                         val,
-                        neigh_coord,
-                        neigh_coord1,
+                        neigh_cart,
+                        neigh_cart1,
                         coord,
                     )
 
                     if score < persistence_tol:
+                        if (
+                            use_minima
+                            and neigh_val1 > neigh_val
+                            or not use_minima
+                            and neigh_val1 < neigh_val
+                            or neigh_val1 == neigh_val
+                            and neigh_idx1 < neigh_idx
+                        ):
+                            idx = cursor[neigh_idx1]
+                            # grow arrays if needed
[…]
