### 2026-03-16 1eb4162
fix memory explosion in basin reduction
