+++ src/baderkit/core/utilities/persistence.py
@@ -19,11 +19,14 @@ from baderkit.core.utilities.critical_points import refine_critical_points
 from baderkit.core.utilities.interpolation import linear_slice
 from baderkit.core.utilities.transforms import (
     ALL_NEIGHBOR_TRANSFORMS,
+    IMAGE_TO_INT,
     INT_TO_IMAGE,
     get_transform_dists,
 )
 from baderkit.core.utilities.union_find import (
+    find_root_no_compression,
     find_root_with_shift1,
+    find_root_with_shift_no_compression1,
     union,
     union_with_shift1,
 )
@@ -324,36 +327,140 @@ def get_approx_saddle_val(
     return conn_val
 
 
-@njit(parallel=True, cache=True)
+# @njit(parallel=True, cache=True)
+# def group_by_low_approx_persistence(
+#     data,
+#     labels,
+#     images,
+#     extrema_values,
+#     extrema_labels,
+#     extrema_vox,
+#     max_cart_offset,
+#     use_minima,
+#     persistence_tol,
+#     matrix,
+# ):
+#     shape = np.array(data.shape)
+#     nx, ny, nz = data.shape
+#     extrema_frac = extrema_vox / shape
+#     extrema_cart = extrema_frac @ matrix
+#     num_extrema = len(extrema_values)
+
+#     best_neighs = np.arange(num_extrema)
+#     best_shifts = np.zeros((num_extrema, 3), dtype=np.int8)
+
+#     # loop over each point and find the best neighbor
+#     for ext_idx in prange(num_extrema):
+#         ext_frac = extrema_frac[ext_idx]
+#         ext_cart = extrema_cart[ext_idx]
+#         ext_value = extrema_values[ext_idx]
+
+#         # create a tracker for the best neighbor
+#         best_label = int(ext_idx)
+#         best_persistence = np.inf
+#         best_shift = np.zeros(3, dtype=np.int8)
+#         for neigh_ext_idx in range(num_extrema):
+#             # skip the original index
+#             if ext_idx == neigh_ext_idx:
+#                 continue
+#             # skip neighbors with less extreme values
+#             neigh_ext_value = extrema_values[neigh_ext_idx]
+#             if (
+#                 use_minima
+#                 and neigh_ext_value > ext_value
+#                 or not use_minima
+#                 and neigh_ext_value < ext_value
+#                 or neigh_ext_value == ext_value
+#                 and neigh_ext_idx > ext_idx
+#             ):
+#                 continue
+
+#             # get neighbor frac coord
+#             neigh_frac = extrema_frac[neigh_ext_idx]
+#             shift = np.round(neigh_frac - ext_frac).astype(np.int8)
+#             wrapped_neigh_frac = neigh_frac - shift
+
+#             # check if this neighbor is outside our cutoff and continue if so
+#             neigh_cart = wrapped_neigh_frac @ matrix
+
+#             offset = neigh_cart - ext_cart
+#             dist = np.linalg.norm(offset)
+#             if dist > max_cart_offset:
+#                 continue
+
+#             # get the value these extrema connect at
+#             n_points = max(int(round(dist * 20)), 5)
+#             conn_val = get_approx_saddle_val(
+#                 ext_frac,
+#                 wrapped_neigh_frac,
+#                 n_points,
+#                 data,
+#                 use_minima,
+#                 nx,
+#                 ny,
+#                 nz,
+#             )
+
+#             if conn_val == np.inf:
+#                 continue
+
+#             conn_cart = ext_cart + offset / 2
+
+#             # get persistence
+#             persistence_score = get_persistence_value(
+#                 ext_value, neigh_ext_value, conn_val, ext_cart, neigh_cart, conn_cart
+#             )
+
+#             # if our persistence is below our tolerance and the current best,
[…]
