### 2026-03-17 2151f06
paper text
