### 2026-03-17 588adf7
fix gui dependencies
