+++ pyproject.toml
@@ -66,6 +66,7 @@ gui = [
     "pyvista>=0.44.0,<=0.46.3",
     "pyvistaqt>=0.6.0,<=0.11.3",
     "qtpy>=2.0.0,<3.0",
+    "PyQt5"
 ]
 
 # For development and testing. Taken straight from simmate
