### 2026-03-17 26583ed
move vacuum check to warning
