### 2026-03-17 099b569
docs and gui fixes (#46)
Fixes several bugs in the gui and updates documentation for clarity. Also adds initial framework for critical points.
