+++ src/baderkit/core/critical_points/critical_points.py
@@ -1,0 +1,599 @@
+# -*- coding: utf-8 -*-
+
+from typing import TypeVar
+
+import numpy as np
+from networkx import MultiDiGraph
+from numpy.typing import NDArray
+
+from baderkit.core.bader.bader import Bader
+from baderkit.core.base.base_analysis import BaseAnalysis
+from baderkit.core.toolkit import Grid
+from baderkit.core.utilities.basic import merge_frac_coords_weighted
+from baderkit.core.utilities.basins import get_manifold_labels, get_manifold_labels_thin
+from baderkit.core.utilities.transforms import INT_TO_IMAGE
+
+from .saddle_connections import (
+    get_saddle_saddle_connections,
+)
+
+# This allows for Self typing and is compatible with python 3.10
+Self = TypeVar("Self", bound="CriticalPoints")
+
+
+class CriticalPoints(BaseAnalysis):
+    """
+    Class for running Bader analysis on a regular grid. For information on each
+    method, see our [docs](https://sweav02.github.io/baderkit/)
+
+    Parameters
+    ----------
+    charge_grid : Grid
+        The Grid object with the charge density that will be integrated.
+    total_charge_grid : Grid | None, optional
+        The Grid object used for determining vacuum regions in the system. For
+        pseudopotential codes this represents the total electron density and should
+        be provided whenever possible. If None, defaults to the charge_grid.
+    reference_grid : Grid | None, optional
+        The Grid object whose values will be used to construct the basins. This
+        should typically only be set when partitioning functions other than the
+        charge density (e.g. ELI-D, ELF, etc.).If None, defaults to the
+        total_charge_grid.
+    vacuum_tol : float | bool, optional
+        If a float is provided, this is the value below which a point will
+        be considered part of the vacuum. If a bool is provided, no vacuum
+        will be used on False, and the default tolerance (0.001) will be used on True.
+    persistence_tol : float, optional
+        It is common for false maxima to be found using only nearest neighbor
+        points. To deal with this we combine pairs of basins that have low
+        topological persistence.
+
+        The persistence score is calculated as:
+
+            score = dist * (lower_max - saddle_value) / higher_max
+
+        where 'dist' is the cartesian distance between the maxima, lower_max
+        is the value at the maximum with a lower value, saddle_value is the
+        highest value at which there is a connection between the two maximas
+        descending manifold (basin), and higher_max is the value at the maximum
+        with a higher value. Any pairs that score below the tolerance will be
+        combined. The default is 0.01.
+    grid : Grid | Bader
+        A Grid or Bader type object to perform the analysis on. If a Bader object
+        is provided, the reference_grid will be used.
+
+    """
+
+    _reset_props = [
+        "minima_frac",
+        "minima_vox",
+        "minima_cart",
+        "minima_values",
+        "minima_persistence_groups",
+        "minima_group_types",
+        "saddle1_frac",
+        "saddle1_vox",
+        "saddle1_cart",
+        "saddle1_values",
+        "saddle1_persistence_groups",
+        "saddle1_group_types",
+        "saddle2_frac",
+        "saddle2_vox",
+        "saddle2_cart",
+        "saddle2_values",
+        "saddle2_persistence_groups",
+        "saddle2_group_types",
+        "maxima_frac",
+        "maxima_vox",
+        "maxima_cart",
+        "maxima_values",
+        "maxima_persistence_groups",
+        "maxima_group_types",
+        "saddle1_minima_connections",
+        "saddle2_maxima_connections",
+        "saddle_saddle_connections",
+        "manifold_labels",
+        "morse_graph",
+        "bond_paths",
+    ]
+
+    def __init__(
+        self,
[…]
