+++ pyproject.toml
@@ -55,7 +55,7 @@ dependencies = [
     "pandas>=2.1.0,<3.0",
     "pymatgen>=2022.1.9,<=2025.10.7",
     "scipy>=1.7.2,<2.0",
-    "numba>=0.59.0,<=0.62.1",
+    "numba>=0.59.0,<=0.64.0",
     "typer>=0.6.1",
     "rich>=11.0,<=14.2",
 ]
