### 2026-03-17 adc8866
Bump actions/setup-python from 5 to 6 (#50)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
