### 2026-03-17 31b406c
update basin reduction discussion
