+++ joss_paper/paper.md
@@ -46,9 +46,9 @@ A secondary goal of `BaderKit` is to update grid-based Bader algorithms to utili
 
 ## Speed and Parallelization
 
-BaderKit includes each of the algorithms from the original Fortran code including the `ongrid` [@Henkelman:2006], `neargrid` [@Henkelman:2009], and `weight` [@Yu:2011] methods. On a modern machine (AMD Ryzen™ Threadripper™ 1950X CPU with 16 cores), `BaderKit` runs two to three times as fast as the original Fortran code (\autoref{fig:fig1}). This speedup is primarily due to changes made to each method that allow them to be parallelized on multi-core architectures. Here, we briefly touch on the most significant changes.
+BaderKit includes each of the algorithms from the original Fortran code including the `ongrid` [@Henkelman:2006], `neargrid` [@Henkelman:2009], and `weight` [@Yu:2011] methods. On a modern machine (AMD Ryzen™ Threadripper™ 1950X CPU with 16 cores), `BaderKit` runs ~1.5-3 times as fast as the original Fortran code (\autoref{fig:fig1}). This speedup is primarily due to changes made to each method that allow them to be parallelized on multi-core architectures. Here, we briefly touch on the most significant changes.
 
-![Comparison of runtimes for BaderKit and the Henkelman Fortran code on the conventional cubic 8 atom NaCl structure (a=b=c=5.539Å) at varying grid densities. For a fair comparison, both methods were called through the command line and include reading files, running the algorithm, and writing outputs. The charge density is calculated by density functional theory (DFT) using the PBE functional[@Perdew:1996], an energy cutoff of 372.85 eV, a 3x3x3 Monkhorst–Pack k-point mesh, and VASP's default GW pseudo-potentials. \label{fig:fig1}](time_vs_grid_baderkit_henk_subplots.png)
+![Comparison of runtimes for BaderKit and the Henkelman Fortran code on the conventional cubic 8 atom NaCl structure (a=b=c=5.539Å) at varying grid densities. Runtimes include reading files, running the algorithm, and writing outputs. BaderKit additionally locates saddle points during this process. The charge density is calculated via density functional theory (DFT) using the PBE functional[@Perdew:1996], an energy cutoff of 372.85 eV, a 3x3x3 Monkhorst–Pack k-point mesh, and VASP's default GW pseudo-potentials. \label{fig:fig1}](time_vs_grid_baderkit_henk_subplots.png)
 
 The original `ongrid` and `neargrid` methods perform hill climbing algorithms that start at arbitrary points and climb the steepest gradient until either a maximum or previous point is reached. The `ongrid` method is fast, but its results are highly dependent on the rotational orientation of the system. The `neargrid` method is similar, but improves upon this by storing a correction vector from the current point to the true gradient and making adjustments when the vector is sufficiently large. It requires an additional edge refinement step as the correction vector is only correct for the initial starting point of the path. Instead of a path-building method, `BaderKit` loops over each point in parallel and establishes a pointer to the steepest neighbor. This creates a classic 'forest of trees' problem where the root of each tree corresponds to a Bader basin. `BaderKit` then finds these roots using a vectorized pointer jumper algorithm. For the `neargrid` method, the points along the edge are then iteratively refined in parallel using the original path method. This operation is significantly sped up by caching the gradients calculated during the initial pointer construction.
 
@@ -56,9 +56,9 @@ In contrast to the `ongrid` and `neargrid` methods, the `weight` method allows p
 
 ## Basin Reduction
 
[…]
