+++ joss_paper/paper.bib
@@ -78,6 +78,18 @@
     url={gaussian.com}
 }
 
+@article{Perdew:1996, 
+    title={Generalized gradient approximation made simple.}, 
+    volume={77}, 
+    DOI={10.1103/PhysRevLett.77.3865}, 
+    number={18}, 
+    journal={Physical Review Letters}, 
+    author={Perdew, J P and Burke, K and Ernzerhof, M}, 
+    year={1996}, 
+    month={Oct}, 
+    pages={3865–3868},
+    }
+
 @article{Giannozzi:2009, 
     title={QUANTUM ESPRESSO: a modular and open-source software project for quantum simulations of materials.}, 
     volume={21}, 
