### 2026-03-17 SWeav02/baderkit#45 @SWeav02
@mdavezac I've made updates based on your comments. Most of the updates were to the documentation which should now be [live](https://sweav02.github.io/baderkit/). I also made some changes to the paper, including updating the speed tests to match the current build and changing out figure 2 for a system that is hopefully a bit clearer. The new pdf should be [here](https://github.com/SWeav02/baderkit/actions/runs/23208592687)

Let me know if you have any other comments! I'm happy to update, and it only makes the package better. Thanks!
