### 2026-03-17 #9943 @SWeav02
@zhubonan @mdavezac @obaica I've made updates to the documentation and paper based on https://github.com/SWeav02/baderkit/issues/45. The updated docs should be [live](https://sweav02.github.io/baderkit/) and the new pdf of the paper should be [here](https://github.com/SWeav02/baderkit/issues/45).

Thanks for your attention and work!
