### 2026-03-18 3e1cbfa
initial overlap class
