### 2026-03-19 #9943 @zhubonan
@mdavezac Just pining to check how it is going. I can see that code changes have been applied to address: https://github.com/SWeav02/baderkit/issues/45 .
