+++ src/baderkit/core/bader/methods/neargrid_weight/neargrid_weight_numba.py
@@ -51,7 +51,7 @@ def get_edge_charges_volumes(
     labels,
     charges,
     volumes,
-    vacuum_mask,
+    vacuum_label,
     neighbor_transforms,
     neighbor_alpha,
     all_neighbor_transforms,
@@ -91,7 +91,8 @@ def get_edge_charges_volumes(
             # get neighbor and wrap around periodic boundary
             ii, jj, kk = wrap_point(i + si, j + sj, k + sk, nx, ny, nz)
             # if this point is part of the vacuum, continue
-            if vacuum_mask[ii, jj, kk]:
+            label = labels[ii,jj,kk]
+            if label == vacuum_label:
                 continue
             # get the neighbors value
             neigh_value = reference_data[ii, jj, kk]

[…]
