### 2026-03-19 b8a1247
start more complex overlap class
