+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler.py
@@ -1,11 +1,16 @@
 # -*- coding: utf-8 -*-
 
+from pathlib import Path
+
 import numpy as np
 from numpy.typing import NDArray
 
 from baderkit.core.base.base_analysis import BaseAnalysis
 from baderkit.core.toolkit import Grid
 from baderkit.core.elf_analysis.overlap import BasinOverlap
+from baderkit.core.utilities.coord_env import is_along_bond_all
+
+from .enum_and_styling import FeatureType
 
 
 class ElfLabeler(BaseAnalysis):
@@ -16,7 +21,10 @@ class ElfLabeler(BaseAnalysis):
     """
     
     _reset_props = [
-        "basin_types"
+        "basin_types",
+        "attractor_shapes",
+        "along_bond",
+        "heavily_polarized",
         ]
 
     def __init__(
@@ -24,6 +32,8 @@ class ElfLabeler(BaseAnalysis):
         charge_grid: Grid,
         reference_grid: Grid,
         total_charge_grid: Grid | None = None,
+        min_covalent_angle: float = 135,
+        max_covalent_polarization: float = 0.5,
         **kwargs,
     ):
         # create bader objects
@@ -35,6 +45,9 @@ class ElfLabeler(BaseAnalysis):
         )
         
         self.elf_bader = self.overlap.local_bader
+        
+        self._min_covalent_angle = min_covalent_angle
+        self._max_covalent_polarization = max_covalent_polarization
 
         super().__init__(
             charge_grid=charge_grid,
@@ -43,9 +56,61 @@ class ElfLabeler(BaseAnalysis):
             **kwargs,
         )
         
+    ###########################################################################
+    # Settings
+    ###########################################################################
+    @property
+    def min_covalent_angle(self) -> float:
+        return self._min_covalent_angle
+    
+    @min_covalent_angle.setter
+    def min_covalent_angle(self, value: float):
+        self._min_covalent_angle = value
+        self._reset_properties(
+            include_properties=[
+                "basin_types",
+                "along_bond",
+                ], 
+            )
+    
+    @property
+    def max_covalent_polarization(self) -> float:
+        return self._max_covalent_polarization
+    
+    @max_covalent_polarization.setter
+    def max_covalent_polarization(self, value: float):
+        self._max_covalent_polarization = value
+        self._reset_properties(
+            include_properties=[
+                "basin_types",
+                "heavily_polarized",
+                ], 
+            )
+    
+    ###########################################################################
+    # Properties
+    ###########################################################################
+     
     @property
     def maxima_frac(self) -> NDArray[np.float64]:
         return self.elf_bader.maxima_frac
+    
+    @property
+    def attractor_shapes(self) -> NDArray[str]:
+        if self._attractor_shapes is None:
+            betti_nums = self.elf_bader.maxima_betti_numbers
+            shapes = []
+            for i,j,k in betti_nums:
+                if i == 1 and j == 0 and k == 0:
+                    shapes.append("point")
+                elif i == 1 and j == 1 and k == 0:
+                    shapes.append("ring")
+                elif i == 1 and j == 0 and k == 1:
+                    shapes.append("cage")
+                else:
+                    raise Exception("Unknown shape found for attractor. This is a bug!")
+            self._attractor_shapes = np.array(shapes)
+        return self._attractor_shapes
     
     @property
     def basin_types(self) -> list[str]:
@@ -53,32 +118,176 @@ class ElfLabeler(BaseAnalysis):
             self._label_basins()
         return self._basin_types
     
+    @property
+    def along_bond(self):
+        if self._along_bond is None:
+            self._along_bond, _ = is_along_bond_all(
+                feature_frac_coords=self.maxima_frac,
[…]
