+++ src/baderkit/core/elf_analysis/overlap/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .overlap import BasinOverlap

[…]
