### 2026-03-20 004f53f
fix image bug in neargrid method
