+++ src/baderkit/core/elf_analysis/overlap/overlap.py
@@ -10,6 +10,7 @@ from baderkit.core.bader.methods import Method
 from baderkit.core.base.base_analysis import BaseAnalysis
 from baderkit.core.toolkit import Grid
 from .overlap_numba import get_overlaps
+from baderkit.core.utilities.coord_env import is_along_bond_all
 
 
 class BasinOverlap(BaseAnalysis):
@@ -24,9 +25,11 @@ class BasinOverlap(BaseAnalysis):
         "overlap_counts",
         "overlap_labels",
         "local_overlap_fractions",
-        "qtaim_overlap_fractions",
-        "shared_local_basins",
-        "unshared_local_basins",
+        "core_basins",
+        # "qtaim_overlap_fractions",
+        "along_bond",
+        "attractor_shapes",
+        "valence_basins",
         "polarization_indexes",
     ]
 
@@ -37,11 +40,13 @@ class BasinOverlap(BaseAnalysis):
         total_charge_grid: Grid | None = None,
         nna_cutoff: float = 1.0,
         method: str | Method = Method.neargrid,
+        polarization_cutoff: float = 0.5,
         **kwargs,
     ):
+        self.polarization_cutoff = polarization_cutoff
         
         # create bader objects
-        self.qtaim = Bader(
+        self.qtaim_bader = Bader(
             charge_grid=charge_grid,
             total_charge_grid=total_charge_grid,
             reference_grid=total_charge_grid,
@@ -64,6 +69,20 @@ class BasinOverlap(BaseAnalysis):
             reference_grid=reference_grid,
             **kwargs,
         )
+        
+    @property
+    def max_covalent_polarization(self) -> float:
+        return self._max_covalent_polarization
+    
+    @max_covalent_polarization.setter
+    def max_covalent_polarization(self, value: float):
+        self._max_covalent_polarization = value
+        self._reset_properties(
+            include_properties=[
+                "valence_basins",
+                ], 
+            )
+    
 
     ###########################################################################
     # Properties calculated by other classes
@@ -81,33 +100,8 @@ class BasinOverlap(BaseAnalysis):
 
         """
 
-        return self.qtaim.vacuum_mask
+        return self.qtaim_bader.vacuum_mask
 
-    @property
-    def local_basin_labels(self) -> NDArray[np.int64]:
-        """
-
-        Returns
-        -------
-        NDArray[np.int64]
-            A 3D grid representing the points in space assigned to local basins
-            such as those calculated from the ELF, ELI-D, LOL, etc.
-
-        """
-        return self.local_bader.maxima_basin_labels
-    
-    @property
-    def qtaim_basin_labels(self) -> NDArray[np.int64]:
-        """
-
-        Returns
-        -------
-        NDArray[np.int64]
-            A 3D grid representing the points in space assigned to Bader atoms.
-
-        """
-        return self.qtaim.atom_labels
-    
     @property
     def local_maxima_frac(self) -> NDArray[np.float64]:
         """
@@ -132,7 +126,19 @@ class BasinOverlap(BaseAnalysis):
             grid. The order corresponds to the atom labels and structure.
 
         """
-        return self.qtaim.maxima_frac
+        return self.qtaim_bader.maxima_frac
+    
+    @property
+    def along_bond(self):
+        if self._along_bond is None:
+            self._along_bond, _ = is_along_bond_all(
+                feature_frac_coords=self.maxima_frac,
+                atom_frac_coords=self.structure.frac_coords,
+                atom_cart_coords=self.structure.cart_coords,
+                matrix=self.reference_grid.matrix,
+                min_covalent_angle=self.min_covalent_angle,
+            )
+        return self._along_bond
     
     ###########################################################################
[…]
