+++ src/baderkit/core/elf_analysis/overlap/overlap.py
@@ -6,7 +6,6 @@ import numpy as np
 from numpy.typing import NDArray
 
 from baderkit.core.bader.bader import Bader
-from baderkit.core.bader.methods import Method
 from baderkit.core.base.base_analysis import BaseAnalysis
 from baderkit.core.toolkit import Grid
 from .overlap_numba import get_overlaps
@@ -26,10 +25,9 @@ class BasinOverlap(BaseAnalysis):
         "overlap_labels",
         "local_overlap_fractions",
         "core_basins",
-        # "qtaim_overlap_fractions",
+        "shared_basins",
         "along_bond",
         "attractor_shapes",
-        "valence_basins",
         "polarization_indexes",
     ]
 
@@ -39,11 +37,12 @@ class BasinOverlap(BaseAnalysis):
         reference_grid: Grid,
         total_charge_grid: Grid | None = None,
         nna_cutoff: float = 1.0,
-        method: str | Method = Method.neargrid,
+        min_covalent_angle: float = 135,
         polarization_cutoff: float = 0.5,
         **kwargs,
     ):
-        self.polarization_cutoff = polarization_cutoff
+        self._polarization_cutoff = polarization_cutoff
+        self._min_covalent_angle = min_covalent_angle
         
         # create bader objects
         self.qtaim_bader = Bader(
@@ -51,7 +50,6 @@ class BasinOverlap(BaseAnalysis):
             total_charge_grid=total_charge_grid,
             reference_grid=total_charge_grid,
             nna_cutoff=nna_cutoff,
-            method=method,
             **kwargs,
         )
         
@@ -59,7 +57,6 @@ class BasinOverlap(BaseAnalysis):
             charge_grid=charge_grid,
             total_charge_grid=total_charge_grid,
             reference_grid=reference_grid,
-            method=method,
             **kwargs,
         )
 
@@ -69,14 +66,31 @@ class BasinOverlap(BaseAnalysis):
             reference_grid=reference_grid,
             **kwargs,
         )
+
+    ###########################################################################
+    # Settings
+    ###########################################################################
+    @property
+    def min_covalent_angle(self) -> float:
+        return self._min_covalent_angle
+    
+    @min_covalent_angle.setter
+    def min_covalent_angle(self, value: float):
+        self._min_covalent_angle = value
+        self._reset_properties(
+            include_properties=[
+                "core_basins",
+                "along_bond",
+                ], 
+            )
         
     @property
-    def max_covalent_polarization(self) -> float:
-        return self._max_covalent_polarization
+    def polarization_cutoff(self) -> float:
+        return self._polarization_cutoff
     
-    @max_covalent_polarization.setter
-    def max_covalent_polarization(self, value: float):
-        self._max_covalent_polarization = value
+    @polarization_cutoff.setter
+    def polarization_cutoff(self, value: float):
+        self._polarization_cutoff = value
         self._reset_properties(
             include_properties=[
                 "valence_basins",
@@ -132,7 +146,7 @@ class BasinOverlap(BaseAnalysis):
     def along_bond(self):
         if self._along_bond is None:
             self._along_bond, _ = is_along_bond_all(
-                feature_frac_coords=self.maxima_frac,
+                feature_frac_coords=self.local_maxima_frac,
                 atom_frac_coords=self.structure.frac_coords,
                 atom_cart_coords=self.structure.cart_coords,
                 matrix=self.reference_grid.matrix,
@@ -197,20 +211,38 @@ class BasinOverlap(BaseAnalysis):
         return self._local_overlap_fractions
     
     @property
-    def core_basins(self) -> NDArray[bool]:
+    def core_basins(self) -> NDArray[int]:
         """
 
         Returns
         -------
-        NDArray[bool]
-            An array with an entry for each local basin that is True for basins
[…]
