+++ src/baderkit/core/elf_analysis/overlap/overlap.py
@@ -8,7 +8,7 @@ from numpy.typing import NDArray
 from baderkit.core.bader.bader import Bader
 from baderkit.core.base.base_analysis import BaseAnalysis
 from baderkit.core.toolkit import Grid
-from .overlap_numba import get_overlaps
+from .overlap_numba import get_overlaps, get_atom_access_sets
 from baderkit.core.utilities.coord_env import is_along_bond_all
 
 
@@ -21,14 +21,19 @@ class BasinOverlap(BaseAnalysis):
     
     _reset_props = [
         "atomicities",
-        "overlap_counts",
+        "overlap_charge_table",
         "overlap_labels",
-        "local_overlap_fractions",
+        "bond_fractions",
         "core_basins",
         "shared_basins",
         "along_bond",
         "attractor_shapes",
         "polarization_indexes",
+        "atom_core_populations",
+        "atom_valence_populations",
+        "atom_access_sets",
+        "atom_charge_claims",
+        
     ]
 
     def __init__(
@@ -142,6 +147,9 @@ class BasinOverlap(BaseAnalysis):
         """
         return self.qtaim_bader.maxima_frac
     
+    ###########################################################################
+    # Properties
+    ###########################################################################
     @property
     def along_bond(self):
         if self._along_bond is None:
@@ -153,10 +161,6 @@ class BasinOverlap(BaseAnalysis):
                 min_covalent_angle=self.min_covalent_angle,
             )
         return self._along_bond
-    
-    ###########################################################################
-    # Properties related to overlap
-    ###########################################################################
 
     @property
     def atomicities(self) -> NDArray[np.int64]:
@@ -174,41 +178,42 @@ class BasinOverlap(BaseAnalysis):
         if self._atomicities is None:
             # The number of atoms contributing to each label is the number of
             # non-zero entries in each row of our overlap_matrix
-            self._atomicities = np.array([len(i) for i in self.local_overlap_fractions])
+            self._atomicities = np.array([len(i) for i in self.bond_fractions])
         return self._atomicities
                
     @property
-    def overlap_counts(self) -> NDArray[np.int64]:
+    def overlap_charge_table(self) -> NDArray[np.int64]:
         """
 
         Returns
         -------
         NDArray[np.int64]
-            An Nx3 array where the columns represent the qtaim basin, local basin,
-            and number of overlapping grid points respectively.
+            An Nx4 array where each row represents a single overlap basin. The
+            columns represent the atom index, atom image, local basin index
+            and the integrated charge in the basin.
 
         """
-        if self._overlap_counts is None:
+        if self._overlap_charge_table is None:
             self._get_overlap()
-        return self._overlap_counts
+        return self._overlap_charge_table
 
     @property
-    def local_overlap_fractions(self) -> list[NDArray[np.float64]]:
+    def bond_fractions(self) -> list[NDArray[np.float64]]:
         """
 
         Returns
         -------
         list[NDArray[np.float64]]
             A list with the same length as the number of local basins. Each entry
-            is an Nx2 array where N is the number of overlapping qtaim basins
-            and each entry is the qtaim basin index and fraction of charge in the
-            overlapping basin respectively. For valence basins, this is equivalent 
-            to the fraction of the bond belonging to each overlapping atom.
+            is an Nx3 array where N is the number of overlapping qtaim basins
+            and each entry is the qtaim atom index, the qtaim atom's periodic
+            image (relative to the local basin's maximum), and the fraction of 
+            charge in the overlapping basin respectively. 
 
         """
-        if self._local_overlap_fractions is None:
+        if self._bond_fractions is None:
             self._get_overlap()
-        return self._local_overlap_fractions
[…]
