### 2026-03-22 #9943 @obaica
@zhubonan  @SWeav02  @editorialbot

I have completed the review of the `baderkit` software and its accompanying JOSS manuscript. I believe `baderkit` is a highly valuable and modern tool that will bring a much more efficient and user-friendly experience to the community performing **Bader charge analysis**. I recommend the software for acceptance after a few minor bugs and documentation improvements are addressed. Below are my comments and suggestions:

### Software Issues & Suggestions:
**1. JSON Serialization Bug during Execution**

When executing commands like  `baderkit run CHGCAR`, the program crashes during the result-saving step with the following error:

`TypeError: Object of type int64 is not JSON serializable.`

I was able to temporarily fix this by introducing a custom `NumpyEncoder` in `baderkit/core/base/base_analysis.py`. For the authors' reference, the fix involves:

- Adding a `NumpyEncoder` class (inheriting from json.JSONEncoder) above the `BaseAnalysis` class to handle `NumPy` data types .

```
class NumpyEncoder(json.JSONEncoder):
    def default(self, obj):
        if hasattr(obj, 'tolist'):
            return obj.tolist()
        return super().default(obj)
```

- Adding `cls=NumpyEncoder` to the` json.dumps` call in the `to_json` method.
- Adding `cls=NumpyEncoder` to the `json.dump` call in the `write_json` method.

**2. Example Files Accessibility**

Currently, the documentation provides a `Google Drive link` for the example files. External links like this are prone to expiration (link rot) and can be easily overlooked by users. I highly recommend creating an examples folder directly within the `GitHub` repository and hosting these files there.

**3. Reproducibility of Examples**

For the provided examples, it is difficult to reproduce the exact results without knowing the specific computational parameters. I suggest updating the `README` files within the examples to specify the exact types of `pseudopotentials` (e.g., `PBE`, `LDA`) and the **valence electron configurations** (e.g., `_sv`, `_pv` for VASP users) used to generate the input charge densities.

**4. Need for a Comprehensive/Literature-based Tutorial**
The current examples are quite basic and only provide raw output files, making it challenging for users to evaluate the correctness of their runs. It would be highly beneficial to include a comprehensive tutorial that reproduces the results of a published paper. This tutorial should include a detailed workflow: from the initial DFT calculation setup, running baderkit, post-processing the data, to visualizing the results or providing comparison tables against the literature.

**5. Inaccurate GUI Installation Prompt**

If a user attempts to run `baderkit gui` without the required GUI dependencies installed, the CLI prompts:

`WARNING Baderkits GUI requires additional dependencies. Please run 'pip install baderkit'`

This is slightly misleading and should be updated to direct the user to install the specific optional dependencies. It should be changed to:

`Please run 'pip install baderkit[gui]' (or 'pip install "baderkit[gui]"' depending on the OS/shell).`

**6. Capitalization in CLI Help Text**

In the command-line interface help menu (e.g., when running `baderkit --help`), `"bader"` should be capitalized as `"Bader"` when referring to the method. For instance, the description for the run command currently says: "Runs a bader analysis on the provided files." It should be updated to "Runs a Bader analysis...".

### Paper/Manuscript Issues:

**1. Typos and Grammar**:

Please correct the following minor typographical errors in the manuscript:

Line 12: "systems" -> "system's"

Line 26: "alorithms" -> "algorithms"

Line 30: "CPU's" -> "CPUs"

Line 89: "codes handling" -> "code's handling"

**2. Missing Journal Name in References**:

In the References section (Lines 135-137), the citation for Henkelman, G. et al. (2006) is missing the journal name. It should include the journal title: Computational Materials Science.
