### 2026-03-23 #9943 @SWeav02
@obaica @zhubonan @editorialbot
Thanks for the review! I'll comment here with some responses to your suggestions and start working on implementing them.

> **1. JSON Serialization Bug during Execution**
>
> When executing commands like `baderkit run CHGCAR`, the program crashes during the result-saving step with the following error:
>
> `TypeError: Object of type int64 is not JSON serializable.`
>
> I was able to temporarily fix this by introducing a custom `NumpyEncoder` in `baderkit/core/base/base_analysis.py`. For the authors' reference, the fix involves:
>
> * Adding a `NumpyEncoder` class (inheriting from json.JSONEncoder) above the `BaseAnalysis` class to handle `NumPy` data types .
>
> ```
> class NumpyEncoder(json.JSONEncoder):
>     def default(self, obj):
>         if hasattr(obj, 'tolist'):
>             return obj.tolist()
>         return super().default(obj)
> ```
> * Adding `cls=NumpyEncoder` to the` json.dumps` call in the `to_json` method.
> * Adding `cls=NumpyEncoder` to the `json.dump` call in the `write_json` method.

I did try and fix this error in the most recent version of BaderKit with some hardcoding, but this is a much more elegant solution that will prevent similar issues in the future. Thanks for the tip!

>
> **2. Example Files Accessibility**
>
> Currently, the documentation provides a `Google Drive link` for the example files. External links like this are prone to expiration (link rot) and can be easily overlooked by users. I highly recommend creating an examples folder directly within the `GitHub` repository and hosting these files there.
>
Good suggestion. I originally made the drive link as some of the example files are above githubs usual size limit. I know there are ways to add large files and prevent them from being tracked/downloaded but I've never used them before. I'll look into those and see if I can figure out a clean solution.

> **3. Reproducibility of Examples**
>
> For the provided examples, it is difficult to reproduce the exact results without knowing the specific computational parameters. I suggest updating the `README` files within the examples to specify the exact types of `pseudopotentials` (e.g., `PBE`, `LDA`) and the **valence electron configurations** (e.g., `_sv`, `_pv` for VASP users) used to generate the input charge densities.
>
Easy enough. I'll add the exact POTCAR versions and see if I'm missing anything else needed to replicate the data.

> **4. Need for a Comprehensive/Literature-based Tutorial** The current examples are quite basic and only provide raw output files, making it challenging for users to evaluate the correctness of their runs. It would be highly beneficial to include a comprehensive tutorial that reproduces the results of a published paper. This tutorial should include a detailed workflow: from the initial DFT calculation setup, running baderkit, post-processing the data, to visualizing the results or providing comparison tables against the literature.
>
This is a good suggestion, and one I would like to implement. I'm not sure how quickly I'll be able to get to this since it will take a bit of time to write out in a satisfactory way, but I'll try and get to it soon.

> **5. Inaccurate GUI Installation Prompt**
>
> If a user attempts to run `baderkit gui` without the required GUI dependencies installed, the CLI prompts:
>
> `WARNING Baderkits GUI requires additional dependencies. Please run 'pip install baderkit'`
>
> This is slightly misleading and should be updated to direct the user to install the specific optional dependencies. It should be changed to:
>
> `Please run 'pip install baderkit[gui]' (or 'pip install "baderkit[gui]"' depending on the OS/shell).`

This issue has plagued me for a bit as the actual logging message does say 'pip install baderkit[gui]'. The `[gui]` part shows up when I run the warning in an IPython shell, but not when run through bash. I tried an escape in a previous version but this gave a warning every time the gui command was run which seemed like it would be confusing to users.
Regardless, there is certainly a solution and I'll look a little harder to find it.

>
> **6. Capitalization in CLI Help Text**
>
> In the command-line interface help menu (e.g., when running `baderkit --help`), `"bader"` should be capitalized as `"Bader"` when referring to the method. For instance, the description for the run command currently says: "Runs a bader analysis on the provided files." It should be updated to "Runs a Bader analysis...".
>
Good call! Thanks for noticing that.

> ### Paper/Manuscript Issues:
> **1. Typos and Grammar**:
>
> Please correct the following minor typographical errors in the manuscript:
>
> Line 12: "systems" -> "system's"
>
> Line 26: "alorithms" -> "algorithms"
>
> Line 30: "CPU's" -> "CPUs"
>
> Line 89: "codes handling" -> "code's handling"
>
> **2. Missing Journal Name in References**:
>
> In the References section (Lines 135-137), the citation for Henkelman, G. et al. (2006) is missing the journal name. It should include the journal title: Computational Materials Science.

I can fix these sometime today. Thanks again!
