### 2026-03-23 2c5c644
fix gui logging bug
