### 2026-03-23 3959db7
add broader pseudopotential support
