+++ src/baderkit/core/toolkit/grid.py
@@ -22,7 +22,7 @@ from baderkit.core.toolkit.structure import Structure
 from baderkit.core.utilities.basic import get_transforms_in_radius
 from baderkit.core.utilities.file_parsers import (
     Format,
-    detect_format,
+    detect_volume_format,
     infer_significant_figures,
     read_cube,
     read_vasp,
@@ -1474,10 +1474,12 @@ class Grid(VolumetricData):
         try:
             import h5py
         except:
-            raise ImportError("""
+            raise ImportError(
+                """
                 The `h5py` package is required to read/write to the hdf5 format.
                 Please install with `conda install h5py` or `pip install h5py`.
-                """)
+                """
+            )
 
         logging.info(f"Loading {grid_file}")
         t0 = time.time()
@@ -1532,12 +1534,14 @@ class Grid(VolumetricData):
 
         if format is None:
             # guess format from file
-            format = detect_format(grid_file)
+            format = detect_volume_format(grid_file)
 
+        # make sure format is a Format enum
+        format = Format(format)
         # make sure format is an available option
         assert (
             format in Format
-        ), "Invalid provided format '{format}'. Options are: {[i.value for i in Format]}"
+        ), f"Invalid provided format '{format}'. Options are: {[i.value for i in Format]}"
 
         # get the reading method corresponding to this output format
         method_name = format.reader
@@ -1601,10 +1605,12 @@ class Grid(VolumetricData):
         try:
             import h5py
         except:
-            raise ImportError("""
+            raise ImportError(
+                """
                 The `h5py` package is required to read/write to the hdf5 format.
                 Please install with `conda install h5py` or `pip install h5py`.
-                """)
+                """
+            )
         filename = Path(filename)
         logging.info(f"Writing {filename.name}")
         super().to_hdf5(filename)
