+++ src/baderkit/core/bader/bader.py
@@ -3,14 +3,12 @@
 import importlib
 import logging
 import time
-import warnings
 from pathlib import Path
 from typing import TypeVar
 
 import numpy as np
 import pandas as pd
 from numpy.typing import NDArray
-from pymatgen.io.vasp import Potcar
 
 from baderkit.core.base.base_analysis import BaseAnalysis
 from baderkit.core.toolkit import Structure
@@ -54,6 +52,10 @@ class Bader(BaseAnalysis):
         should typically only be set when partitioning functions other than the
         charge density (e.g. ELI-D, ELF, etc.).If None, defaults to the
         total_charge_grid.
+    valence_counts : dict | None, optional
+        A dictionary where each key is an atomic species in the system and each
+        value is the number of valence electrons used in the pseudo potential.
+        This is used for methods that calculate oxidation states.
     method : str | Method, optional
         The algorithm to use for generating bader basins.
     vacuum_tol : float | bool, optional
@@ -66,51 +68,48 @@ class Bader(BaseAnalysis):
         found, dummy atoms will be appended to the structure and regarded as
         separate species. If a bool is provided, NNAs will be assigned to the
         nearest atom on False or a default value (1 Ang) will be used on True.
-    maxima_persistence_tol : float, optional
+    persistence_tol : float, optional
         It is common for false maxima to be found using only nearest neighbor
         points. To deal with this we combine pairs of basins that have low
         topological persistence.
 
         The persistence score is calculated as:
 
-            score = (lower_max - connection_value) / connection_value
-    maxima_persistence_tol : float, optional
-        It is common for false minima to be found using only nearest neighbor
-        points. To deal with this we combine pairs of basins that have low
-        topological persistence. We use a separate tolerance as minima typically
-        have lower persistence values.
+            score = abs(lower_max - connection_value) / connection_value
 
-        The persistence score is calculated as:
-
-            score = (connection_value - higher_min) / connection_value
 
 
     """
 
-    _reset_props = [
-        # assigned by _run_bader
-        "maxima_basin_labels",
-        "maxima_basin_images",
+    _method_kwargs = [
+        "method",
+        "nna_cutoff",
+        "persistence_tol",
+    ]
+
+    _maxima_results = [
         "maxima_frac",
         "maxima_cart",
         "maxima_vox",
-        "ongrid_maxima_groups",
         "maxima_charge_values",
         "maxima_ref_values",
-        "maxima_persistence_values",
         "basin_charges",
         "basin_volumes",
-        # assigned by _run_minima_bader
-        "minima_basin_labels",
-        "minima_basin_images",
+        "basin_min_surface_distances",
+        "basin_avg_surface_distances",
+        "basin_surface_areas",
+        "basin_contact_surface_areas",
+    ]
+
+    _minima_results = [
         "minima_frac",
         "minima_cart",
         "minima_vox",
-        "ongrid_minima_groups",
         "minima_charge_values",
         "minima_ref_values",
-        "minima_persistence_values",
-        # saddle props
+    ]
+
+    _saddle_results = [
         "saddle1_frac",
         "saddle1_cart",
         "saddle1_vox",
@@ -121,31 +120,58 @@ class Bader(BaseAnalysis):
         "saddle2_ref_values",
         "saddle1_connections",
         "saddle2_connections",
-        # Assigned by calling the property
-        "basin_min_surface_distances",
-        "basin_avg_surface_distances",
-        "basin_edges",
-        "atom_edges",
-        "basin_surface_areas",
-        "basin_contact_surface_areas",
-        "atom_surface_areas",
-        "atom_contact_surface_areas",
-        # Assigned by run_atom_assignment
+    ]
+
+    _atom_results = [
         "basin_atoms",
         "basin_atom_dists",
-        "atom_labels",
[…]
