+++ src/baderkit/command_line/base.py
@@ -78,6 +78,11 @@ def run(
         "-ref",
         help="The path to the reference file",
     ),
+    pseudopotential_file: list[Path] = typer.Option(
+        None,
+        "--pseudopotentials" "-pp",
+        help="The path to pseudopotential files for calculating oxidation states. If None, the current directory will be searched for files with common pseudopotential names (POTCAR, .UPF, .xml). Multiple files can be specified by calling this parameter multiple times (e.g. -pp file1 -pp file2 etc.)",
+    ),
     method: Method = typer.Option(
         Method.weight,
         "--method",
@@ -141,6 +146,7 @@ def run(
         charge_filename=charge_file,
         total_charge_filename=total_charge_file,
         reference_filename=reference_file,
+        pseudopotential_filename=pseudopotential_file,
         method=method,
         format=format,
         vacuum_tol=vacuum_tolerance,
@@ -224,7 +230,9 @@ def sum(
 
     shape1 = tuple(grid1.shape)
     shape2 = tuple(grid2.shape)
-    assert shape1 == shape2, f"""
+    assert (
+        shape1 == shape2
+    ), f"""
     Grids must have the same shape. {file1.name}: {shape1} differs from {file2.name}: {shape2}
     """
     # sum grids
@@ -343,13 +351,15 @@ def split(
     grid = Grid.from_dynamic(file, format=input_format, total_only=False)
 
     if not grid.is_spin_polarized:
-        raise Exception("""
+        raise Exception(
+            """
             This method only splits files that contain both the total and difference
             charge densities like VASP's CHGCAR format.
             If you need help splitting a charge density or ELF in a different format, please
             open a discussion on our [github](https://github.com/SWeav02/baderkit/discussions) and
             we will try and add an example to our documentation.
-            """)
+            """
+        )
 
     # split grid. raises errors internally
     spin_up, spin_down = grid.split_to_spin()

[…]
