+++ joss_paper/paper.bib
@@ -4,6 +4,7 @@
     DOI={10.1016/j.commatsci.2005.04.010}, 
     url={https://doi.org/10.1016/j.commatsci.2005.04.010},
     author={Henkelman, Graeme and Arnaldsson, Andri and Jónsson, Hannes}, 
+    journal={Computational Materials Science}, 
     year={2006}, 
     month={Jun}, 
     pages={354–360}
