### 2026-03-23 #9943 @SWeav02
@obaica @zhubonan
I've updated the paper with typo/grammar fixes and the missing journal: https://github.com/SWeav02/baderkit/actions/runs/23460088622

I've also made most of the fixes from https://github.com/openjournals/joss-reviews/issues/9943#issuecomment-4106291616 with the exception of a more comprehensive tutorial. These changes are currently in the `review2updates` branch and I'm planning to commit them to main after I update tests and linting. I should get to that in the next day or two and I'll comment here again once I've done so.
