### 2026-03-24 e0ea070
add more pseudopotential support (#56)
