+++ src/baderkit/core/base/base_analysis.py
@@ -3,7 +3,7 @@
 import copy
 import json
 import logging
-from abc import ABC, abstractclassmethod
+from abc import ABC
 from pathlib import Path
 from typing import Literal, TypeVar
 
@@ -11,13 +11,20 @@ import numpy as np
 from numpy.typing import NDArray
 
 from baderkit.core.toolkit import Grid, Structure
-from baderkit.core.utilities.file_parsers import Format
+from baderkit.core.utilities.file_parsers import Format, load_pseudo
 
 # # This allows for Self typing and is compatible with python 3.10
 Self = TypeVar("Self", bound="BaseAnalysis")
 
 # TODO:
 # - Add handling of non-nuclear attractors (e.g. those in Li metal)
+
+
+class NumpyEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if hasattr(obj, "tolist"):
+            return obj.tolist()
+        return super().default(obj)
 
 
 class BaseAnalysis(ABC):
@@ -38,6 +45,12 @@ class BaseAnalysis(ABC):
         should typically only be set when partitioning functions other than the
         charge density (e.g. ELI-D, ELF, etc.).If None, defaults to the
         total_charge_grid.
+    valence_counts : dict | None | False, optional
+        A dictionary where each key is an atomic species in the system and each
+        value is the number of valence electrons used in the pseudo potential.
+        This is used for methods that calculate oxidation states.
+        For all electron calculations, this can be set to False to automatically
+        calculate based on atomic numbers.
     vacuum_tol : float | bool, optional
         If a float is provided, this is the value below which a point will
         be considered part of the vacuum. If a bool is provided, no vacuum
@@ -53,12 +66,19 @@ class BaseAnalysis(ABC):
         "structure",
     ]
     _reset_props = []
+    _base_method_kwargs = [
+        "vacuum_tol",
+    ]
+    _method_kwargs = []
+
+    _summary_props = []
 
     def __init__(
         self,
         charge_grid: Grid,
         total_charge_grid: Grid | None = None,
         reference_grid: Grid | None = None,
+        valence_counts: dict | None = None,
         vacuum_tol: float | bool = 1.0e-3,
         **kwargs,
     ):
@@ -83,6 +103,13 @@ class BaseAnalysis(ABC):
 
         self._total_charge_grid = total_charge_grid
         self._reference_grid = reference_grid
+
+        if valence_counts is False:
+            valence_counts = {}
+            # The user requests a full electron calculation
+            for site in charge_grid.structure:
+                valence_counts[site.specie.symbol] = getattr(site.specie, "number", 0)
+        self._valence_counts = valence_counts
 
         # check if the total charge grid has values below 0
         if vacuum_tol is not False:
@@ -199,6 +226,24 @@ class BaseAnalysis(ABC):
         self._reset_properties()
 
     @property
+    def valence_counts(self) -> dict | None:
+        """
+
+        Returns
+        -------
+        dict | None
+            A dictionary where each key is an atomic species in the system and each
+            value is the number of valence electrons used in the pseudo potential.
+            This is used for methods that calculate oxidation states.
+
+        """
+        return self._valence_counts
+
+    @valence_counts.setter
+    def valence_counts(self, value: dict):
+        self._valence_counts = value
+
+    @property
     def vacuum_tol(self) -> float | bool:
         """
 
@@ -298,9 +343,6 @@ class BaseAnalysis(ABC):
     def from_vasp(
         cls,
         charge_filename: Path | str = "CHGCAR",
-        total_charge_filename: Path | None | str = None,
-        reference_filename: Path | None | str = None,
-        total_only: bool = True,
         **kwargs,
     ) -> Self:
         """
[…]
