### 2026-03-24 #9943 @SWeav02
@obaica @zhubonan
Updates from https://github.com/openjournals/joss-reviews/issues/9943#issuecomment-4106291616 have been pushed to main and pip. Let me know if you run into anything else or have other suggestions!
