### 2026-03-25 #9943 @obaica
> [@obaica](https://github.com/obaica) [@zhubonan](https://github.com/zhubonan) Updates from [#9943 (comment)](https://github.com/openjournals/joss-reviews/issues/9943#issuecomment-4106291616) have been pushed to main and pip. Let me know if you run into anything else or have other suggestions!

Thank you for the update. I have tested version 0.9.1, and the JSON serialization issue encountered during execution has been resolved on my end.

I also noticed that several typos in the software paper have been corrected.

I appreciate your prompt efforts in addressing these issues.
