+++ src/baderkit/core/elf_analysis/overlap/overlap_numba.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from numba import njit, prange
+from numba import njit, prange, types
 
 import numpy as np
 from numpy.typing import NDArray
@@ -8,27 +8,319 @@ from numpy.typing import NDArray
 from baderkit.core.utilities.transforms import INT_TO_IMAGE, IMAGE_TO_INT
 
 @njit(cache=True, parallel=True)
+def get_label_image_map(
+    labels,
+    images,
+    num_labels,
+        ):
+    nx, ny, nz = labels.shape
+    labels_w_images = np.zeros((num_labels, 27), dtype=np.bool_)
+    for i in prange(nx):
+        for j in range(ny):
+            for k in range(nz):
+                label = labels[i,j,k]
+                # skip vacuum
+                if label == num_labels:
+                    continue
+
+                # get the image
+                image = images[i,j,k]
+
+                labels_w_images[label, image] = True
+    # construct label map
+    pairs = np.argwhere(labels_w_images)
+    label_map = np.empty_like(labels_w_images, dtype=np.int16)
+    for idx in prange(len(pairs)):
+        i,j = pairs[idx]
+        label_map[i,j] = idx
+    return pairs, label_map
+
+@njit(cache=True)
+def get_unique_overlaps(
+    atom_labels,
+    local_labels,
+    atom_images,
+    local_images,
+    atom_image_map,
+    local_image_map,
+    num_atoms,
+    num_local,
+):
+    nx, ny, nz = atom_labels.shape
+
+    # create set to store pairs
+    atom_local_pairs = set()
+
+    # collect unique pairs
+    for i in range(nx):
+        for j in range(ny):
+            for k in range(nz):
+                atom_label = atom_labels[i, j, k]
+                local_label = local_labels[i, j, k]
+
+                # skip vacuum
+                if atom_label == num_atoms or local_label == num_local:
+                    continue
+
+                atom_image = atom_images[i, j, k]
+                local_image = local_images[i, j, k]
+
+                atom_pair = atom_image_map[atom_label, atom_image]
+                local_pair = local_image_map[local_label, local_image]
+
+                pair_val = szudzik_pair(atom_pair, local_pair)
+                atom_local_pairs.add(int(pair_val))
+
+    # convert set to array
+    out = np.empty(len(atom_local_pairs), dtype=np.int64)
+
+    idx = 0
+    for pair in atom_local_pairs:
+        out[idx] = pair
+        idx += 1
+
+    return out
+
+@njit(parallel=True, cache=True)
+def get_overlap_table(
+        atom_labels,
+        atom_images,
+        local_labels,
+        local_images,
+        num_atoms,
+        num_local,
+        ):
+    nx, ny, nz = local_labels.shape
+    # first we construct maps to and from each label/image pair to a
+    # single index. This reduces the memory needed to find overlaps
+    atom_image_pairs, atom_image_map = get_label_image_map(
+        labels=atom_labels,
+        images=atom_images,
+        num_labels=num_atoms,
+        )
+    local_image_pairs, local_image_map = get_label_image_map(
+        labels=local_labels,
+        images=local_images,
+        num_labels=num_local,
+        )
+    atom_local_pairs = get_unique_overlaps(
+        atom_labels,
+        local_labels,
+        atom_images,
+        local_images,
+        atom_image_map,
+        local_image_map,
+        num_atoms,
+        num_local,
+        )
+    # get the unique sets of charge/local pairs
+    unique_overlaps = np.empty((len(atom_local_pairs), 4), dtype=np.int64)
+    for idx in prange(len(atom_local_pairs)):
+        pair_idx = atom_local_pairs[idx]
[…]
