+++ src/baderkit/core/bader/bader.py
@@ -113,6 +113,9 @@ class Bader(BaseAnalysis):
         "minima_betti_groups",
         "minima_betti_cutoffs",
         "minima_betti_numbers",
+        ]
+
+    _saddle_results = [
         # saddle props
         "saddle1_frac",
         "saddle1_cart",
@@ -173,7 +176,7 @@ class Bader(BaseAnalysis):
         self,
         method: str | Method = Method.default,
         nna_cutoff: float | bool = False,
-        persistence_tol: float = 0.01,
+        persistence_tol: float = 0.1,
         **kwargs,
     ):
         super().__init__(**kwargs)
@@ -472,7 +475,7 @@ class Bader(BaseAnalysis):
         if self._ongrid_maxima_groups is None:
             self._run_bader()
         return self._ongrid_maxima_groups
-    
+
     @property
     def maxima_betti_groups(self) -> list[NDArray[np.int64]]:
         """
@@ -486,30 +489,29 @@ class Bader(BaseAnalysis):
         """
         # TODO: Update this to be the exact value at which each group has its
         # betti number
-        
+
         if self._maxima_betti_groups is None:
-            tol = max(self.maxima_persistence_tol, 0)
-            # self._run_maxima_bader()
+            tol = max(self.persistence_tol, 0)
             # get groups
             maxima_groups = self.ongrid_maxima_groups
             maxima_values = self.maxima_ref_values
             # get the lowest value that the maximum would connect to with the
             # current persistence tol
             persistence_values = []
-            for group, min_val in zip(maxima_groups, maxima_values):
+            for group, max_val in zip(maxima_groups, maxima_values):
                 group_vals = self.reference_grid.total[
                     group[:, 0],
                     group[:, 1],
                     group[:, 2],
                 ]
-                valid_mask = ((group_vals - min_val) / group_vals) - 1e-12 <= tol
-                best_val = group_vals[valid_mask].max()
+                valid_mask = (abs(max_val - group_vals) / group_vals) - 1e-12 <= tol
+                best_val = group_vals[valid_mask].min()
                 # get lowest possible persistence below this value
-                # (val - max_val) / val < persistence_tol
+                # (max_val - val) / val < persistence_tol
                 # --> val = max_val / (1+persistence_tol)
-                persistence_values.append(best_val / (1 - self.maxima_persistence_tol))
+                persistence_values.append(best_val / (1 + self.persistence_tol))
             persistence_values = np.array(persistence_values)
-            
+
             self._maxima_betti_groups = get_persistence_groups(
                 labels=self.maxima_basin_labels,
                 data=self.reference_grid.total,
@@ -518,7 +520,7 @@ class Bader(BaseAnalysis):
                 use_minima=False,
             )
         return self._maxima_betti_groups
-    
+
     @property
     def maxima_betti_cutoffs(self) -> NDArray[np.float64]:
         """
@@ -532,7 +534,7 @@ class Bader(BaseAnalysis):
             generally artifacts of the grid rather than significant features.
 
         """
-        
+
         if self._maxima_betti_cutoffs is None:
             self._get_betti_numbers(False)
         return self._maxima_betti_cutoffs
@@ -848,7 +850,7 @@ class Bader(BaseAnalysis):
         if self._ongrid_minima_groups is None:
             self._run_minima_bader()
         return self._ongrid_minima_groups
-    
+
     @property
     def minima_betti_groups(self) -> list[NDArray[np.int64]]:
         """
@@ -860,9 +862,9 @@ class Bader(BaseAnalysis):
             that are within each minimas persistence threshold.
 
         """
-        
+
         if self._minima_betti_groups is None:
-            tol = max(self.minima_persistence_tol, 0)
[…]

[…]
