+++ src/baderkit/core/elf_analysis/overlap/overlap.py
@@ -9,7 +9,7 @@ from numpy.typing import NDArray
 
 from baderkit.core.bader.bader import Bader
 from baderkit.core.base.base_analysis import BaseAnalysis
-from baderkit.core.toolkit import Grid
+from baderkit.core.toolkit import Grid, Structure
 from .overlap_numba import (
     get_atom_charge_claims,
     get_overlap_table,
@@ -34,8 +34,8 @@ class BasinOverlap(BaseAnalysis):
         "overlap_charges",
         "overlap_volumes",
         "overlap_labels",
-        "local_overlap_fractions",
-        "qtaim_overlap_fractions",
+        "bond_fractions",
+        "qtaim_overlap_groups",
         "core_basins",
         "shared_basins",
         "along_bond",
@@ -44,11 +44,26 @@ class BasinOverlap(BaseAnalysis):
         "atom_core_populations",
         "atom_valence_populations",
         "atom_access_sets",
+        "atom_access_electron_numbers",
         "atom_charge_claims",
         "atom_connection_indices",
         "atom_connection_index_labels",
-
     ]
+    
+    _summary_props = [
+        "atomicities",
+        "bond_fractions",
+        "core_basins",
+        "shared_basins",
+        "along_bond",
+        "attractor_shapes",
+        "polarization_indexes",
+        "atom_core_populations",
+        "atom_valence_populations",
+        "atom_charge_claims",
+        "atom_connection_indices",
+        "atom_connection_index_labels",
+        ]
 
     def __init__(
         self,
@@ -56,12 +71,10 @@ class BasinOverlap(BaseAnalysis):
         reference_grid: Grid,
         total_charge_grid: Grid | None = None,
         nna_cutoff: float = 1.0,
-        min_covalent_angle: float = 135,
-        polarization_cutoff: float = 0.5,
+        min_bond_angle: float = 135,
         **kwargs,
     ):
-        self._polarization_cutoff = polarization_cutoff
-        self._min_covalent_angle = min_covalent_angle
+        self._min_bond_angle = min_bond_angle
 
         # create bader objects
         self.qtaim_bader = Bader(
@@ -90,12 +103,12 @@ class BasinOverlap(BaseAnalysis):
     # Settings
     ###########################################################################
     @property
-    def min_covalent_angle(self) -> float:
-        return self._min_covalent_angle
+    def min_bond_angle(self) -> float:
+        return self._min_bond_angle
 
-    @min_covalent_angle.setter
-    def min_covalent_angle(self, value: float):
-        self._min_covalent_angle = value
+    @min_bond_angle.setter
+    def min_bond_angle(self, value: float):
+        self._min_bond_angle = value
         self._reset_properties(
             include_properties=[
                 "core_basins",
@@ -103,24 +116,21 @@ class BasinOverlap(BaseAnalysis):
                 ],
             )
 
-    @property
-    def polarization_cutoff(self) -> float:
-        return self._polarization_cutoff
-
-    @polarization_cutoff.setter
-    def polarization_cutoff(self, value: float):
-        self._polarization_cutoff = value
-        self._reset_properties(
-            include_properties=[
-                "valence_basins",
-                ],
-            )
-
-
     ###########################################################################
     # Properties calculated by other classes
     ###########################################################################
+    
+    @property
+    def structure(self) -> Structure:
+        """
 
+        Returns
+        -------
+        Structure
+            The pymatgen structure basins are assigned to.
+
+        """
+        return self.qtaim_bader.structure
 
     @property
     def vacuum_mask(self) -> NDArray[bool]:
@@ -159,7 +169,7 @@ class BasinOverlap(BaseAnalysis):
             grid. The order corresponds to the atom labels and structure.
 
         """
[…]
