### 2026-03-29 4ef04ba
refinement based combination
