+++ src/baderkit/core/utilities/critical_points.py
@@ -43,40 +43,29 @@ def classify_critical_point(H):
 
     return -1
 
+@njit(cache=True, inline="always")
+def compute_signature(evals, eig_rel_tol):
 
-# @njit(cache=True, fastmath=True)
-# def compute_signature(H, eig_rel_tol):
-#     evals, _ = np.linalg.eigh(H)
+    max_eval = 0.0
+    for i in range(3):
+        v = abs(evals[i])
+        if v > max_eval:
+            max_eval = v
 
-#     tol = eig_rel_tol * np.max(np.abs(evals))
+    tol = eig_rel_tol * max_eval + 1e-14
 
-#     n_neg = 0
-#     n_flat = 0
-#     for i in evals:
-#         if i < -tol:
-#             n_neg += 1
-#         elif abs(i) < tol:
-#             n_flat += 1
-
-
-#     return n_neg, n_flat
-@njit(cache=True)
-def compute_signature(H, eig_rel_tol):
-
-    evals = np.linalg.eigvalsh(H)
-    tol = eig_rel_tol * np.max(np.abs(evals))
-
-    n_neg = 0
+    morse = 0
     n_flat = 0
 
-    for lam in evals:
-        if lam < -tol:
-            n_neg += 1
-        elif abs(lam) <= tol:
+    for i in range(3):
+        lam = evals[i]
+
+        if abs(lam) < tol:
             n_flat += 1
+        elif lam < 0.0:
+            morse += 1
 
-    return n_neg, n_flat
-
+    return morse, n_flat
 
 @njit(cache=True)
 def is_ongrid_saddle(
@@ -112,7 +101,6 @@ def is_ongrid_saddle(
 
     return True
 
-
 @njit(fastmath=True, cache=True)
 def is_ongrid_newton_crit(
     coord,
@@ -122,11 +110,27 @@ def is_ongrid_newton_crit(
     H_frac_to_cart,
 ):
     """
-    Newton-step rejection using *full local neighborhood information*.
-    Uses full 3x3x3 Hessian and a conservative spectral bound.
+    Newton-step critical point classification with degeneracy handling.
+
+    Returns:
+        -1 : reject
+         0 : minimum
+         1 : index-1 saddle
+         2 : index-2 saddle
+         3 : maximum
+
+        10 : rank-1 valley  (1 positive, 2 ~0)
+        11 : rank-1 ridge   (1 negative, 2 ~0)
+
+        20 : rank-2 degenerate min-like (2 positive, 1 ~0)
+        21 : rank-2 degenerate saddle   (1 pos, 1 neg, 1 ~0)
+        22 : rank-2 degenerate max-like (2 negative, 1 ~0)
+
+        30 : rank-0 flat
     """
     nx, ny, nz = data.shape
     eps = 1e-12
+
     i, j, k = coord
     i = int(i)
     j = int(j)
@@ -154,7 +158,7 @@ def is_ongrid_newton_crit(
     gk = 0.5 * (n[1, 1, 2] - n[1, 1, 0])
 
     # -----------------------------
-    # Full Hessian (fractional)
+    # Hessian (fractional)
     # -----------------------------
     Hxx = n[2, 1, 1] - 2.0 * f0 + n[0, 1, 1]
     Hyy = n[1, 2, 1] - 2.0 * f0 + n[1, 0, 1]
@@ -165,69 +169,106 @@ def is_ongrid_newton_crit(
     Hyz = 0.25 * (n[1, 2, 2] - n[1, 2, 0] - n[1, 0, 2] + n[1, 0, 0])
 
     # -----------------------------
-    # Solve Newton step H s = -g
+    # Scale-aware tolerance
     # -----------------------------
[…]
