+++ src/baderkit/core/utilities/basins.py
@@ -179,7 +179,6 @@ def get_best_neighbor_with_shift(
         np.array((bsi, bsj, bsk), dtype=np.int8),
     )
 
-
 ###############################################################################
 # Edge Finding Methods
 ###############################################################################
