### 2026-03-29 #9943 @zhubonan
Hi @mdavezac just check how is the review going. I can see that the author has fixed the code based on your request in https://github.com/SWeav02/baderkit/issues/45. There are still a few unticked items in your checklist though. Could you please finish them up? Do let me and author know if you have and difficulties with the code.
