### 2026-03-29 #9943 @mdavezac
After review, I'm` happy to recommend Baderkit for publication in JOSS. It is a well designed tool that should prove useful to the community.
