### 2026-03-29 215a274
more optimized cutoffs
