### 2026-03-30 3a576ee
add volume ratio methods
