+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler_numba.py
@@ -1,12 +1,169 @@
 # -*- coding: utf-8 -*-
 
 from numba import njit, prange, types
+from math import erf
 
 import numpy as np
 from numpy.typing import NDArray
 
 from baderkit.core.utilities.transforms import INT_TO_IMAGE, IMAGE_TO_INT
 from baderkit.core.utilities.interpolation import linear_slice
+
+# @njit(cache=True)
+def get_core_gaussian_fit(
+    total_charge_data,
+    atom_labels,
+    atom_images,
+    basin_labels,
+    core_mask,
+    atom_frac_coords,
+    matrix,
+        ):
+    shape = np.array(total_charge_data.shape, dtype=np.int64)
+    nx, ny, nz = shape
+    num_points = nx*ny*nz
+    
+    num_atoms = len(atom_frac_coords)
+    num_basins = len(core_mask)
+    
+    atom_Q = np.zeros(num_atoms, dtype=np.float64)
+    atom_M2 = np.zeros(num_atoms, dtype=np.float64)
+    total_electrons = np.zeros(num_atoms, dtype=np.float64)
+    basin_electrons = np.zeros(len(core_mask), dtype=np.float64)
+    
+    for i in range(nx):
+        for j in range(ny):
+            for k in range(nz):
+                atom_label = atom_labels[i,j,k]
+                basin_label = basin_labels[i,j,k]
+                # skip vacuum/dummy atoms
+                if atom_label >= num_atoms or basin_label >= num_basins:
+                    continue
+                # add charge
+                charge = total_charge_data[i,j,k] / num_points
+                total_electrons[atom_label] += charge
+                basin_electrons[basin_label] += charge
+                
+                # skip points that are not cores
+                if core_mask[basin_label] == -1:
+                    continue
+                
+                # get coordinates
+                point_frac = np.array((i,j,k), dtype=np.float64) / shape
+                point_cart = point_frac @ matrix
+                
+                atom_frac = atom_frac_coords[atom_label] + INT_TO_IMAGE[atom_images[i,j,k]]
+                atom_cart = atom_frac @ matrix
+                
+                diff = point_cart - atom_cart
+                r2 = diff[0]**2 + diff[1]**2 + diff[2]**2
+
+                atom_Q[atom_label] += charge
+                atom_M2[atom_label] += charge*r2
+    atom_sigmas = np.zeros_like(atom_Q)
+    for a in range(num_atoms):
+        if atom_Q[a] > 1e-12:
+            atom_sigmas[a] = np.sqrt(atom_M2[a] / (3.0 * atom_Q[a]))
+        else:
+            atom_sigmas[a] = 0.0
+    total_electrons = total_electrons
+    basin_electrons = basin_electrons
+    return atom_sigmas, total_electrons, basin_electrons
+                
+                
+
+# @njit(cache=True, parallel=True)
+def get_basin_potential_energies(
+    basin_labels,
+    atom_frac_coords,
+    matrix,
+    nna_indices,
+    charge_bond_fracs,
+    total_charge_data,
+    atom_sigmas,
+    nuclei_charges,
+    electron_charges,
+    basin_electrons,
+        ):
+    shape = np.array(total_charge_data.shape, dtype=np.int64)
+    nx, ny, nz = shape
+    
+
+    num_points = nx*ny*nz
+
+    # get the total potential for each basin
+    potential_energies = np.zeros(len(nna_indices), dtype=np.float64)
+
+    for nna_idx in prange(len(nna_indices)):
+        local_idx = nna_indices[nna_idx]
+        # update the atom oxidation states as if this basins were not part of it
+        charge_bond_frac = charge_bond_fracs[local_idx]
+        
[…]
