+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler.py
@@ -12,7 +12,13 @@ from baderkit.core.toolkit import Grid
 from baderkit.core.elf_analysis.overlap import BasinOverlap
 
 from .enum_and_styling import FeatureType
-from .elf_labeler_numba import get_core_dist_ratios, get_approx_coulomb_potential, get_zeff_nna
+from .elf_labeler_numba import (
+    get_core_dist_ratios, 
+    get_approx_coulomb_potential, 
+    get_zeff_nna,
+    get_basin_potential_energies,
+    get_core_gaussian_fit,
+    )
 
 Self = TypeVar("Self", bound="ElfLabeler")
 
@@ -34,10 +40,12 @@ class ElfLabeler(BaseAnalysis):
         "basin_types",
         "attractor_shapes",
         "heavily_polarized",
-        "nna_core_volume_ratios",
-        "nna_core_distance_ratios",
-        "nna_core_distances",
-        "nna_coulombic_potentials",
+        "core_volume_ratios",
+        "nna_bond_fracs",
+        "nna_bond_dists",
+        "neighbor_zeffs",
+        "neighbor_veffs",
+        "nna_potential_energies",
         ]
     
     _reset_props = [
@@ -123,33 +131,19 @@ class ElfLabeler(BaseAnalysis):
         return self._heavily_polarized
     
     @property
-    def nna_core_volume_ratios(self):
-        if self._nna_core_volume_ratios is None:
-            # get each atom's core volume
-            atom_core_volumes = self.overlap.atom_core_volumes
-            # get the nna indices
-            indices = np.array([i for i, j in enumerate(self.basin_types) if j == "nna"], dtype=np.int64)
-            local_volume_ratios = np.zeros(len(indices), dtype=np.float64)
-            for nna_idx, (local_idx) in enumerate(indices):
-                volume_fracs = self.overlap.volume_bond_fractions[local_idx]
-                # get the volume of this basin
-                local_volume = self.elf_bader.basin_volumes[local_idx]
-                # add the volume of each atom's core
-                core_volume = 0.0    
-                for atom_idx, _, atom_frac in volume_fracs:
-                    core_volume += atom_core_volumes[int(atom_idx)] * atom_frac
-                if core_volume == 0:
-                    continue
-                local_volume_ratios[nna_idx] = local_volume / core_volume
-            self._nna_core_volume_ratios = local_volume_ratios
-        return self._nna_core_volume_ratios
+    def core_volume_ratios(self):
+        if self._core_volume_ratios is None:
+            veffs = self.neighbor_veffs
+            veffs[veffs==0] = 1
+            self._core_volume_ratios = self.overlap.local_bader.basin_volumes / veffs
+        return self._core_volume_ratios
     
     @property
-    def nna_core_distance_ratios(self):
-        "nna frac / core frac"
-        if self._nna_core_distance_ratios is None:
+    def nna_bond_fracs(self):
+        "weighted average of nna fracs of bond to nearest atoms"
+        if self._nna_bond_fracs is None:
             indices = np.array([i for i, j in enumerate(self.basin_types) if j == "nna"], dtype=np.int64)
-            self._nna_core_distance_ratios, self._nna_core_distances, self._nna_distance_ratios = get_core_dist_ratios(
+            self._nna_bond_dists, self._nna_bond_fracs = get_core_dist_ratios(
                 labels=self.elf_bader.maxima_basin_labels,
                 basin_frac_coords=self.elf_bader.maxima_frac,
                 atom_frac_coords=self.elf_bader.structure.frac_coords,
@@ -158,49 +152,87 @@ class ElfLabeler(BaseAnalysis):
                 core_basins=self.overlap.core_basins,
[…]

[…]
