+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler_numba.py
@@ -2,6 +2,7 @@
 
 from numba import njit, prange, types
 from math import erf
+from scipy.special import erf as scipy_erf
 
 import numpy as np
 from numpy.typing import NDArray
@@ -9,161 +10,224 @@ from numpy.typing import NDArray
 from baderkit.core.utilities.transforms import INT_TO_IMAGE, IMAGE_TO_INT
 from baderkit.core.utilities.interpolation import linear_slice
 
-# @njit(cache=True)
-def get_core_gaussian_fit(
-    total_charge_data,
-    atom_labels,
-    atom_images,
+@njit(cache=True)
+def get_valence_potentials(
+    charge_data,
+    potential_data,
     basin_labels,
-    core_mask,
-    atom_frac_coords,
-    matrix,
+    num_basins,
         ):
-    shape = np.array(total_charge_data.shape, dtype=np.int64)
+    shape = np.array(charge_data.shape, dtype=np.int64)
     nx, ny, nz = shape
     num_points = nx*ny*nz
-    
-    num_atoms = len(atom_frac_coords)
-    num_basins = len(core_mask)
-    
-    atom_Q = np.zeros(num_atoms, dtype=np.float64)
-    atom_M2 = np.zeros(num_atoms, dtype=np.float64)
-    total_electrons = np.zeros(num_atoms, dtype=np.float64)
-    basin_electrons = np.zeros(len(core_mask), dtype=np.float64)
-    
+    basin_potentials = np.zeros(num_basins, dtype=np.float64)
+
     for i in range(nx):
         for j in range(ny):
             for k in range(nz):
-                atom_label = atom_labels[i,j,k]
-                basin_label = basin_labels[i,j,k]
-                # skip vacuum/dummy atoms
-                if atom_label >= num_atoms or basin_label >= num_basins:
+                label = basin_labels[i,j,k]
+                if label >= num_basins:
                     continue
-                # add charge
-                charge = total_charge_data[i,j,k] / num_points
-                total_electrons[atom_label] += charge
-                basin_electrons[basin_label] += charge
-                
-                # skip points that are not cores
-                if core_mask[basin_label] == -1:
+                # add potential in this voxel
+                # (e*num_points) * eV = eV * num_points
+                basin_potentials[label] += 0.5 * charge_data[i,j,k] * potential_data[i,j,k]
+    # get the weighted potential felt by this basin
+    basin_potentials /= num_points
+    return basin_potentials
+
+@njit(cache=True)
+def get_avg_potentials(
+    potential_data,
+    basin_labels,
+    num_basins,
+        ):
+    shape = np.array(potential_data.shape, dtype=np.int64)
+    nx, ny, nz = shape
+    basin_potentials = np.zeros(num_basins, dtype=np.float64)
+    basin_counts = np.zeros(num_basins, dtype=np.int64)
+
+    for i in range(nx):
+        for j in range(ny):
+            for k in range(nz):
+                label = basin_labels[i,j,k]
+                if label >= num_basins:
                     continue
-                
-                # get coordinates
-                point_frac = np.array((i,j,k), dtype=np.float64) / shape
-                point_cart = point_frac @ matrix
-                
-                atom_frac = atom_frac_coords[atom_label] + INT_TO_IMAGE[atom_images[i,j,k]]
-                atom_cart = atom_frac @ matrix
-                
-                diff = point_cart - atom_cart
-                r2 = diff[0]**2 + diff[1]**2 + diff[2]**2
+                # add potential in this voxel
+                # (e*num_points) * eV = eV * num_points
[…]
