### 2026-04-16 SWeav02/baderkit#59 @zhubonan
Minor issues with the manuscript file
Hi @SWeav02 , I have found several minor issues with the manuscript related to https://github.com/openjournals/joss-reviews/issues/9943#issuecomment-4231643079:

1. For the address line, please use "United States of America" instead of "United States"
2. "Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules **(QTAIM)**" useful to add this acronym, also   perhaps it is useful to cite this (if there is a commonly used citation).
3. The sentence ”Most are written in Fortran making them cumbersome to automate and adapt for purposes outside their original scope“, perhaps consider   "Most are written in Fortran, which makes them cumbersome to automate and adapt for purposes outside their original scope"
3. There are missing hyphens such as: "memory efficient" -> "memory-efficient", "first principles" -> "first-principles", "hill climbing algorithms" -> "hill-climbing algorithms"
4. This sentence "BaderKit runs directly on the output most popular density functional theory codes" reads better as: ""BaderKit runs directly on the output **of the** most popular density functional theory codes"
5. "Because, this flux is only important at points that straddle multiple basins“ remove the comma -> "Because this flux is only important at points that straddle multiple
basins"

Please update the paper.md. Thanks!
