### 2026-04-16 #9943 @zhubonan
Hi @SWeav02 I have made some suggestion for the manuscript.

Once you have them addressed, please archive the package to get a DOI. The easiest why is to use Zenodo which you can link with your GitHub account. After you deposit the archive, please update the metadata so the title, authors (afflictions) and the software license match that of the JOSS paper. You can also check how other JOSS paper did it by looking at their archive links.

Once you archive the release, please post the DOI here.
