### 2026-04-16 b0d67d3
fix minor paper errors
