+++ joss_paper/paper.md
@@ -17,7 +17,7 @@ authors:
     corresponding: true
     affiliation: 1
 affiliations:
- - name: University of North Carolina Chapel Hill, United States
+ - name: University of North Carolina Chapel Hill, United States of America
    index: 1
    ror: 0130frc33
 date: 13 August 2025
@@ -27,20 +27,20 @@ bibliography: paper.bib
 
 # Summary
 
-The concept of oxidation states has existed for centuries, guiding and informing the decisions of generations of scientists. However, oxidation states are not observable, and cannot be uniquely derived from first principle calculations. This has led researchers to develop a variety of methods to recover oxidation states, each with their own unique theory and methodology. Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules, which derives oxidation states directly from a system's electron charge density. The `BaderKit` package brings Bader charge analysis into the modern Python ecosystem, and reworks the most popular grid-based algorithms to run in parallel.
+The concept of oxidation states has existed for centuries, guiding and informing the decisions of generations of scientists. However, oxidation states are not observable, and cannot be uniquely derived from first-principle calculations. This has led researchers to develop a variety of methods to recover oxidation states, each with their own unique theory and methodology. Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules (QTAIM), which derives oxidation states directly from a system's electron charge density.[@Bader:1994] The `BaderKit` package brings Bader charge analysis into the modern Python ecosystem, and reworks the most popular grid-based algorithms to run in parallel.
 
 # Statement of Need
 
-Bader Quantum Theory of Atoms in Molecules (QTAIM) charge analysis is among the most widely used methods in chemistry and materials science, with thousands of articles referencing the method each year. This popularity has given rise to many packages for performing QTAIM analysis. Despite the variety of implementations, none are well equipped for modern high-throughput workflows. Most are written in Fortran making them cumbersome to automate and adapt for purposes outside their original scope. The algorithms implemented in these codes are serial and do not utilize modern multi-core CPUs. `BaderKit` aims to resolve these issues, providing a fast, parallelized, and easily extended QTAIM implementation written in Python.
+QTAIM charge analysis is among the most widely used methods in chemistry and materials science, with thousands of articles referencing the method each year. This popularity has given rise to many packages for performing QTAIM analysis. Despite the variety of implementations, none are well equipped for modern high-throughput workflows. Most are written in Fortran, which makes them cumbersome to automate and adapt for purposes outside their original scope. The algorithms implemented in these codes are serial and do not utilize modern multi-core CPUs. `BaderKit` aims to resolve these issues, providing a fast, parallelized, and easily extended QTAIM implementation written in Python.
 
 
 # State of the Field
 
[…]
