+++ joss_paper/paper.bib
@@ -1,3 +1,10 @@
+@book{Bader:1994, 
+    title={Atoms in Molecules: A Quantum Theory}, 
+    publisher={Clarendon Press}, 
+    author={Bader, Richard}, 
+    year={1994},
+}
+
 @article{Henkelman:2006, 
     title={A fast and robust algorithm for Bader decomposition of charge density}, 
     volume={36}, 
