+++ joss_paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-The concept of oxidation states has existed for centuries, guiding and informing the decisions of generations of scientists. However, oxidation states are not observable, and cannot be uniquely derived from first-principle calculations. This has led researchers to develop a variety of methods to recover oxidation states, each with their own unique theory and methodology. Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules (QTAIM), which derives oxidation states directly from a system's electron charge density.[@Bader:1994] The `BaderKit` package brings Bader charge analysis into the modern Python ecosystem, and reworks the most popular grid-based algorithms to run in parallel.
+The concept of oxidation states has existed for centuries, guiding and informing the decisions of generations of scientists. However, oxidation states are not observable, and cannot be uniquely derived from first-principle calculations. This has led researchers to develop a variety of methods to recover oxidation states, each with their own unique theory and methodology. Chief among these methods is that described by Bader in his Quantum Theory of Atoms in Molecules (QTAIM) [@Bader:1994], which derives oxidation states directly from a system's electron charge density. The `BaderKit` package brings Bader charge analysis into the modern Python ecosystem, and reworks the most popular grid-based algorithms to run in parallel.
 
 # Statement of Need
 
@@ -36,7 +36,7 @@ QTAIM charge analysis is among the most widely used methods in chemistry and mat
 
 # State of the Field
 
-The most popular implementation is the `Bader v1.05` code developed by the Henkelman group at UT Austin [@Henkelman:2006] who pioneered the grid-based QTAIM algorithms. It is fast and memory-efficient, but its use of Fortran means that workflows using the code must call it as a subprocess. As a result, much of the useful information generated by the process must be read from file or is lost entirely. Additionally, it is fully serial and does not utilize modern CPUs to their full extent. Another popular Fortran implementation, `Critic2`[@Otero-de-la-Roza:2014] provides additional methods and a convenient graphical interface, but suffers from similar automation issues. There have been some previous attempts to alleviate these issues. In particular, `pybader` [@pybader] partially implemented the method into parallelized Python code. However, `pybader` is typically slower than the other implementations and requires a significant amount of boilerplate code. `BaderKit` aims to improve upon the areas where each of these codes falls short. It is fast, extensive, easy to use, and designed to be easily inserted into modern workflows.
+The most popular implementation is the `Bader v1.05` code developed by the Henkelman group at UT Austin [@Henkelman:2006] who pioneered the grid-based QTAIM algorithms. It is fast and memory-efficient, but its use of Fortran means that workflows using the code must call it as a subprocess. As a result, much of the useful information generated by the process must be read from file or is lost entirely. Additionally, it is fully serial and does not utilize modern CPUs to their full extent. Another popular Fortran implementation, `Critic2` [@Otero-de-la-Roza:2014] provides additional methods and a convenient graphical interface, but suffers from similar automation issues. There have been some previous attempts to alleviate these issues. In particular, `pybader` [@pybader] partially implemented the method into parallelized Python code. However, `pybader` is typically slower than the other implementations and requires a significant amount of boilerplate code. `BaderKit` aims to improve upon the areas where each of these codes falls short. It is fast, extensive, easy to use, and designed to be easily inserted into modern workflows.
 
 # Software Design
 
@@ -58,7 +58,7 @@ In contrast to the `ongrid` and `neargrid` methods, the `weight` method allows p
 
[…]
