+++ joss_paper/paper.md
@@ -40,7 +40,7 @@ The most popular implementation is the `Bader v1.05` code developed by the Henke
 
 # Software Design
 
-The foremost goal of `BaderKit` is to make Bader charge analysis broadly accessible and easy to implement. It is available through `github`, `PyPi`, and `conda-forge` and runs on the most popular operating systems (e.g. Windows, MacOS, Ubuntu, etc.). The object-oriented API is based on  the widely used `PyMatGen`[@Ong:2013] allowing users to obtain atomic charges with just three lines of code. `BaderKit` runs directly on the output of the most popular density functional theory codes (e.g. VASP [@Kresse:1996], Gaussian [@gaussian], Quantum Espresso [@Giannozzi:2009]), and can be easily extended to run on the output of others. These choices allow `BaderKit` to be easily inserted into complex workflows where QTAIM analysis is only one part of the process. For users who are less familiar with Python, `BaderKit` includes a command-line interface built with the `Typer` [@typer] package for quick one-off calculations, and a simple desktop application built with `PyQt5` [@PyQt5] and `PyVista` [@Sullivan:2019] for visualization.
+The foremost goal of `BaderKit` is to make Bader charge analysis broadly accessible and easy to implement. It is available through `github`, `PyPi`, and `conda-forge` and runs on the most popular operating systems (e.g. Windows, MacOS, Ubuntu, etc.). The object-oriented API is based on  the widely used `PyMatGen` [@Ong:2013] allowing users to obtain atomic charges with just three lines of code. `BaderKit` runs directly on the output of the most popular density functional theory codes (e.g. VASP [@Kresse:1996], Gaussian [@gaussian], Quantum Espresso [@Giannozzi:2009]), and can be easily extended to run on the output of others. These choices allow `BaderKit` to be easily inserted into complex workflows where QTAIM analysis is only one part of the process. For users who are less familiar with Python, `BaderKit` includes a command-line interface built with the `Typer` [@typer] package for quick one-off calculations, and a simple desktop application built with `PyQt5` [@PyQt5] and `PyVista` [@Sullivan:2019] for visualization.
 
 A secondary goal of `BaderKit` is to update grid-based Bader algorithms to utilize modern architectures. To achieve this, `BaderKit` uses the `Numba` [@Lam:2015] and `NumPy` [@harris:2020] packages to compile expensive operations to machine code and allow for fast, C-based, vectorized calculations. To improve speed further on modern multi-core CPUs, each Bader algorithm has been reworked from the ground up to allow for parallelization where possible.
 
@@ -48,7 +48,7 @@ A secondary goal of `BaderKit` is to update grid-based Bader algorithms to utili
 
 BaderKit includes each of the algorithms from the original Fortran code including the `ongrid` [@Henkelman:2006], `neargrid` [@Henkelman:2009], and `weight` [@Yu:2011] methods. On a modern machine (AMD Ryzen™ Threadripper™ 1950X CPU with 16 cores), `BaderKit` runs ~1.5-3 times as fast as the original Fortran code (\autoref{fig:fig1}). This speedup is primarily due to changes made to each method that allow them to be parallelized on multi-core architectures. Here, we briefly touch on the most significant changes.
 
-![Comparison of runtimes for BaderKit and the Henkelman Fortran code on the conventional cubic 8 atom NaCl structure (a=b=c=5.539Å) at varying grid densities. Runtimes include reading files, running the algorithm, and writing outputs. BaderKit additionally locates saddle points during this process. The charge density is calculated via density functional theory (DFT) using the PBE functional[@Perdew:1996], an energy cutoff of 372.85 eV, a 3x3x3 Monkhorst–Pack k-point mesh, and VASP's default GW pseudo-potentials. \label{fig:fig1}](time_vs_grid_baderkit_henk_subplots.png)
[…]
