### 2026-04-22 42a9087
fix bug in nna location
