+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler_numba.py
@@ -240,16 +240,13 @@ def get_core_dist_ratios(
     volume_bond_fracs,
         ):
 
-    basin_dists = np.zeros(len(nna_indices), dtype=np.float64)
     basin_fracs = np.zeros(len(nna_indices), dtype=np.float64)
 
     for nna_idx in prange(len(nna_indices)):
         # skip cores
         local_idx = nna_indices[nna_idx]
         local_coords = basin_frac_coords[local_idx]
-        local_cart_coords = local_coords @ matrix
         local_bond_frac = volume_bond_fracs[local_idx]
-        weighted_dist = 0.0
         total_basin_frac = 0.0
 
         total_frac = 0.0
@@ -277,6 +274,44 @@ def get_core_dist_ratios(
             # add fraction making up bond
             total_basin_frac += nna_frac * frac
 
+        # adjust for any fractions that had no cores
+        if total_frac == 0:
+            continue
+        frac_mult = 1/total_frac
+        # update arrays
+        basin_fracs[nna_idx] = total_basin_frac * frac_mult
+
+    return basin_fracs
+
+@njit(cache=True, parallel=True)
+def get_core_dists(
+    labels,
+    basin_frac_coords,
+    atom_frac_coords,
+    matrix,
+    nna_indices,
+    core_basins,
+    volume_bond_fracs,
+        ):
+
+    basin_dists = np.zeros(len(nna_indices), dtype=np.float64)
+
+    for nna_idx in prange(len(nna_indices)):
+        # skip cores
+        local_idx = nna_indices[nna_idx]
+        local_coords = basin_frac_coords[local_idx]
+        local_cart_coords = local_coords @ matrix
+        local_bond_frac = volume_bond_fracs[local_idx]
+        weighted_dist = 0.0
+
+        total_frac = 0.0
+        for atom_idx, atom_image, frac in local_bond_frac:
+            if atom_idx >= len(atom_frac_coords):
+                # this is an nna in the charge density and we don't want to include
+                # it.
+                continue
+            atom_coords = atom_frac_coords[int(atom_idx)] + INT_TO_IMAGE[int(atom_image)]
+            total_frac += frac
             # get distance to atom
             atom_cart_coords = atom_coords @ matrix
             dist = np.linalg.norm(atom_cart_coords - local_cart_coords)
@@ -289,8 +324,8 @@ def get_core_dist_ratios(
         frac_mult = 1/total_frac
         # update arrays
         basin_dists[nna_idx] = weighted_dist * frac_mult
-        basin_fracs[nna_idx] = total_basin_frac * frac_mult
-    return basin_dists, basin_fracs
+
+    return basin_dists
 
 @njit(cache=True)
 def get_zeff_nna(
