### 2026-04-24 ff675a6
fix bug in overlap vacuum check
