+++ src/baderkit/core/utilities/critical_points.py
@@ -18,7 +18,215 @@ from baderkit.core.utilities.transforms import (
     IMAGE_TO_INT,
     INT_TO_IMAGE,
 )
-from baderkit.core.utilities.union_find import find_root, union
+from baderkit.core.utilities.union_find import find_root, union, union_w_roots
+
+###############################################################################
+# Ongrid saddle detection
+###############################################################################
+
+@njit(cache=True)
+def estimate_voxel_extremum_periodic(
+        data,
+        i, j, k,
+        lattice,
+        find_max=False,
+        ):
+    """
+    Estimate extremum (min or max) inside a skewed voxel centered at (i,j,k)
+    using center + 6 face neighbors, with periodic boundary conditions.
+
+    Parameters
+    ----------
+    data : 3D array
+        Scalar field
+    i, j, k : int
+        Grid indices of center point
+    lattice : (3,3) array
+        Lattice vectors (unused here, kept for API consistency)
+    find_max : bool
+        If True, estimate maximum instead of minimum
+
+    Returns
+    -------
+    val : float
+        Estimated extremum value inside voxel
+    """
+
+    nx, ny, nz = data.shape
+
+    # periodic indices
+    ip = (i + 1) % nx
+    im = (i - 1) % nx
+    jp = (j + 1) % ny
+    jm = (j - 1) % ny
+    kp = (k + 1) % nz
+    km = (k - 1) % nz
+
+    # optionally flip sign for maxima
+    sign = -1.0 if find_max else 1.0
+
+    # center + neighbors (with sign applied)
+    f0 = sign * data[i, j, k]
+
+    f_px = sign * data[ip, j, k]
+    f_nx = sign * data[im, j, k]
+
+    f_py = sign * data[i, jp, k]
+    f_ny = sign * data[i, jm, k]
+
+    f_pz = sign * data[i, j, kp]
+    f_nz = sign * data[i, j, km]
+
+    # second derivatives
+    d2x = f_px - 2.0 * f0 + f_nx
+    d2y = f_py - 2.0 * f0 + f_ny
+    d2z = f_pz - 2.0 * f0 + f_nz
+
+    # first derivatives
+    dx = 0.5 * (f_px - f_nx)
+    dy = 0.5 * (f_py - f_ny)
+    dz = 0.5 * (f_pz - f_nz)
+
+    # numerical safety
+    eps = 1e-12
+    if abs(d2x) < eps:
+        d2x = eps
+    if abs(d2y) < eps:
+        d2y = eps
+    if abs(d2z) < eps:
+        d2z = eps
+
+    # quadratic minimizer (works for both via sign trick)
+    alpha = -dx / d2x
+    beta  = -dy / d2y
+    gamma = -dz / d2z
+
+    # clamp to voxel
+    if alpha > 0.5:
+        alpha = 0.5
+    elif alpha < -0.5:
+        alpha = -0.5
+
+    if beta > 0.5:
+        beta = 0.5
+    elif beta < -0.5:
+        beta = -0.5
+
+    if gamma > 0.5:
+        gamma = 0.5
+    elif gamma < -0.5:
+        gamma = -0.5
+
+    # evaluate quadratic
+    val = (
+        f0
+        + dx * alpha + dy * beta + dz * gamma
+        + 0.5 * (d2x * alpha * alpha +
+                 d2y * beta  * beta  +
+                 d2z * gamma * gamma)
+    )
+
+    # flip sign back if we computed a max
+    return sign * val
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def edge_extrema_values(
+        data,
+        edge_mask,
+        lattice,
+        use_minima=True,
+        ):
+    """
[…]
