### 2026-04-27 f02eacd
tweak saddle refinement vals
