### 2026-04-28 9dade3d
WIP on elf_overlap
