+++ scratch1.py
@@ -1,21 +1,165 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Apr  2 09:44:39 2026
-
-@author: sammw
-"""
-
-
-from baderkit.core.elf_analysis.elf_labeler1.elf_labeler import ElfLabeler
+import numpy as np
+from scipy.interpolate import interp1d
 from baderkit.core import Grid
-from pathlib import Path
+import re
 
-labeler = ElfLabeler.from_vasp(
-    persistence_tol=0.01,
-    # potential_filename="LOCPOT",
-    )
+# -----------------------------
+# User inputs
+# -----------------------------
+CHGCAR_FILE = "CHGCAR"
+POTCAR_FILE = "POTCAR"
+OUTPUT_FILE = "CHGCAR_total"
 
-# potential = labeler.electrostatic_potential
-test = labeler.nna_potential_energies
-print(test)
-# test = labeler.nna_approx_potential_energies
+# -----------------------------
+# Load CHGCAR
+# -----------------------------
+chgcar = Grid.from_vasp(CHGCAR_FILE)
+structure = chgcar.structure
+grid_valence = np.array(chgcar.total, dtype=float)
+nx, ny, nz = grid_valence.shape
+ngrid = nx * ny * nz
+lattice = structure.lattice
+
+# Cartesian grid
+x = np.linspace(0, 1, nx, endpoint=False)
+y = np.linspace(0, 1, ny, endpoint=False)
+z = np.linspace(0, 1, nz, endpoint=False)
+X, Y, Z = np.meshgrid(x, y, z, indexing="ij")
+frac_grid = np.stack([X, Y, Z], axis=-1)
+cart_grid = lattice.get_cartesian_coords(frac_grid.reshape(-1, 3)).reshape(nx, ny, nz, 3)
+
+# -----------------------------
+# Parse POTCAR core densities safely
+# -----------------------------
+def parse_paw_core_density(potcar_file):
+    core_data = {}
+    Z_dict = {}     # atomic number
+    ZVAL_dict = {}  # valence electrons
+
+    # read file
+    with open(potcar_file, "r") as f:
+        all_text = f.read()
+        # split to different POTCARS
+        potcar_lines = [i.splitlines() for i in all_text.split("End of Dataset")]
+
+    for lines in potcar_lines:
+        i = 0
+        element = None
+        Z = None
+        ZVAL = None
+        r = None
+        rho = None
+        while i < len(lines):
+            line = lines[i].strip()
+    
+            # Element name
+            if line.startswith("VRHFIN"):
+                element = line.split("=")[1].split()[0]
+    
+            # Atomic number and valence from POMASS/ZVAL
+            if "POMASS" in line:
+                m = re.search(r"ZVAL\s*=\s*([\d\.]+)", line)
+                if m:
+                    ZVAL = float(m.group(1))
+                if element is not None:
+                    try:
+                        from pymatgen.core.periodic_table import Element as pmgElement
+                        Z = pmgElement(element).Z
+                    except:
+                        Z = int(ZVAL + 10)  # fallback
+    
+                    
+            if "grid" == line:
+                i += 1
+                r = []
+                while i < len(lines):
+                    try:
+                        r.extend([float(x) for x in lines[i].strip().split()])
+                    except ValueError:
+                        break
+                    i += 1
+                r = np.array(r)
+    
+    
+            # Core charge-density (main, not partial)
+            if "core charge-density" in line and "(partial)" not in line:
+                i += 1
+                rho = []
+                while i < len(lines):
+                    try:
+                        rho.extend([float(x) for x in lines[i].strip().split()])
[…]
