+++ src/baderkit/core/utilities/basic.py
@@ -34,10 +34,8 @@ def dist(p1, p2):
 
 @njit(cache=True)
 def get_gradient_cart(i, j, k, data, dir2car):
-    # dir2car = np.linalg.inv(row_matrix).T
     nx, ny, nz = data.shape
 
-    c000 = data[i, j, k]
     c100 = data[(i + 1) % nx, j, k]
     c_100 = data[(i - 1) % nx, j, k]
     c010 = data[i, (j + 1) % ny, k]
@@ -45,23 +43,10 @@ def get_gradient_cart(i, j, k, data, dir2car):
     c001 = data[i, j, (k + 1) % nz]
     c00_1 = data[i, j, (k - 1) % nz]
 
-    # central differences in voxel coordinates
-    gi = (c100 - c_100) / (2.0)
-    gj = (c010 - c0_10) / (2.0)
-    gk = (c001 - c00_1) / (2.0)
-
-    # optional extrema clamping
-    if c100 <= c000 and c_100 <= c000:
-        gi = 0.0
-    if c010 <= c000 and c0_10 <= c000:
-        gj = 0.0
-    if c001 <= c000 and c00_1 <= c000:
-        gk = 0.0
-
-    # convert to fractional-coordinate gradient
-    gi *= nx
-    gj *= ny
-    gk *= nz
+    # central differences in fractional coordinates
+    gi = (c100 - c_100) * 0.5
+    gj = (c010 - c0_10) * 0.5
+    gk = (c001 - c00_1) * 0.5
 
     # convert to Cartesian gradient
     gx = dir2car[0, 0] * gi + dir2car[0, 1] * gj + dir2car[0, 2] * gk
