+++ src/baderkit/core/elf_analysis/elf_radii/elf_radii.py
@@ -1,0 +1,509 @@
+# -*- coding: utf-8 -*-
+
+from pathlib import Path
+from typing import TypeVar
+
+import numpy as np
+from numpy.typing import NDArray
+from pymatgen.analysis.local_env import CrystalNN
+from scipy.spatial import HalfspaceIntersection
+
+from baderkit.core.base.base_analysis import BaseAnalysis
+from baderkit.core.toolkit import Grid, Structure
+from baderkit.core.elf_analysis.elf_labeler1.elf_labeler import ElfLabeler
+
+from baderkit.core.elf_analysis.elf_labeler1.enum_and_styling import FeatureType
+from .elf_radii_numba import (
+    get_all_atom_elf_radii,
+    )
+from baderkit.core.utilities.voronoi import (
+    generate_symmetric_bonds,
+    get_canonical_bonds,
+    get_planes_on_surface,
+)
+
+
+Self = TypeVar("Self", bound="ElfRadii")
+
+# TODO:
+    # 1. very small maxima in the charge density throw off the rest of the calculation.
+    # These need to be separated properly. Update persistence method to focus on
+    # the lower of the two maxima only
+    # 2. Compare distance of electrides and metals based on the distance past
+    # the core. Only compare for atoms that have a core.
+
+class ElfRadii(BaseAnalysis):
+    """
+    A tool for calculating ionic/covalent radii based on a localization function
+    (ELF, ELI-D, LOL, etc.).
+
+    """
+
+    _summary_props = [
+        "bonding_basins",
+        ]
+
+    _reset_props = [
+        "bonding_structure",
+        ] + _summary_props
+
+
+
+    def __init__(
+        self,
+        charge_grid: Grid,
+        reference_grid: Grid,
+        total_charge_grid: Grid | None = None,
+        include_nnas: bool = False,
+        **kwargs,
+    ):
+        # create bader objects
+        self._labeler = ElfLabeler(
+            charge_grid=charge_grid,
+            total_charge_grid=total_charge_grid,
+            reference_grid=reference_grid,
+            **kwargs,
+        )
+        
+        self._include_nnas = include_nnas
+
+        super().__init__(
+            charge_grid=charge_grid,
+            total_charge_grid=total_charge_grid,
+            reference_grid=reference_grid,
+            **kwargs,
+        )
+        
+    ###########################################################################
+    # Settings
+    ###########################################################################
+
+    @property
+    def include_nnas(self) -> float:
+        return self._include_nnas
+
+    @include_nnas.setter
+    def include_nnas(self, value: float):
+        self._include_nnas = value
+        self._reset_properties()
+
+    ###########################################################################
+    # Properties
+    ###########################################################################
+    @property
+    def labeler(self) -> ElfLabeler:
+        return self._labeler
+    
+    @property
+    def bonding_basins(self) -> NDArray[int]:
+        if self._bonding_basins is None:
+            self._bonding_basins = np.array([i for i,j in self.labeler.basin_types if j in FeatureType.bonding], dtype=np.int64)
+        return self._bonding_basins
+    
+    @property
+    def bonding_structure(self) -> Structure:
+        if self._bonding_structure is None:
+            structure = self.reference_grid.structure.copy()
+            frac_coords = self.labeler.maxima_frac[self.bonding_basins]
+            for frac in frac_coords:
+                structure.append("x", frac)
+            self._bonding_structure = structure
+        return self._bonding_structure
+    
+    @property
+    def bonding_basin_labels(self) -> NDArray[int]:
+        if self._bonding_basin_labels is None:
+            if self._include_nnas:
+                (
+                    atom_labels,
+                    atom_images,
+                    atom_charges,
+                    atom_volumes,
[…]
