### 2026-04-29 7905add
initial radii class transfer
