+++ src/baderkit/core/elf_analysis/elf_radii/elf_radii_numba.py
@@ -9,18 +9,18 @@ from numpy.typing import NDArray
 from baderkit.core.utilities.interpolation import interp_nearest, interp_spline
 
 
-@njit(
-    parallel=True,
-    cache=True,
-)
-def get_elf_radius_frac(
+# @njit(
+#     parallel=True,
+#     cache=True,
+# )
+def get_elf_radius(
     data,
-    feature_labels,
+    local_labels,
     atom_idx,
     atom_coords,
     neigh_coords,
     all_frac_coords,
-    covalent_labels,
+    label_map,
     bond_dist,
     line_res: int = 20,
 ):
@@ -33,8 +33,8 @@ def get_elf_radius_frac(
     ----------
     data : NDArray
         The ELF data.
-    feature_labels : NDArray
-        The labeled grid.
+    local_labels : NDArray
+        The labeled grid for elf basins.
     atom_idx : int
         The index of the first atom in the bond.
     atom_coords : NDArray
@@ -43,8 +43,8 @@ def get_elf_radius_frac(
         The fractional coordinates of the second atom in the bond.
     all_frac_coords : NDArray
         The fractional coordinates of all atoms/dummy atoms in the structure
-    covalent_labels : NDArray
-        The indices in the feature_labels array that should be considered covalent.
+    label_map : NDArray
+        The atomic indices that each value in the local_labels array maps onto.
     bond_dist : float
         The lenght of the bond.
     line_res : int, optional
@@ -78,18 +78,24 @@ def get_elf_radius_frac(
         point = atom_coords + float(point_idx) * step_vec
         x, y, z = point
         values[point_idx] = interp_spline(x, y, z, data)
-        label = int(interp_nearest(x, y, z, feature_labels))
+        label = int(interp_nearest(x, y, z, local_labels))
+        label = label_map[label]
+        if label >= num_atoms:
+            labels[point_idx] = label
+            continue
+        
         # For small cells its possible for this label to be incorrect.
         # In poarticular it may assign to a translation of the actual atom. To
         # test for this, we check if the vector from the point to the site is
         # less than half a unit cell in any direction. This assumes the atom is
         # fairly spherical (which is also an assumption of badelf itself).
-
         vec = all_frac_coords[label] - point
         if np.max(np.abs(vec)) > 0.5:
-            label += num_atoms
+            label += num_atoms+1
 
         labels[point_idx] = label
+    # if label%(num_atoms+1) == 2:
+    #     breakpoint()
     # get the unique labels
     unique_labels = np.unique(labels)
     # TODO: If passes through vacuum, set to last point?
@@ -98,7 +104,7 @@ def get_elf_radius_frac(
     # The atom's nearest neighbor is a translation of itself. The radius
     # must always be halfway between the two and the bond must be covalent
     if len(unique_labels) == 1:
-        return 0.5, True, False
+        return bond_dist * 0.5, 0.5, True, False
 
     # SITUATION 2:
     # The bond passes through labels that are not the current site or neighbor.
@@ -111,15 +117,21 @@ def get_elf_radius_frac(
     site_idx = labels[0]
     neigh_idx = labels[-1]
     for label in unique_labels:
-        equiv_label = label % num_atoms
-        if covalent_labels[equiv_label]:
+        # note interatomic interactions
+        label = label % (num_atoms+1)
+        if label >= num_atoms:
             covalent = True
-        elif label != site_idx and label != neigh_idx:
+            continue
+        
+        # note if a different atom is involved in the bond
+        if label != site_idx and label != neigh_idx:
             other_atoms = True
[…]
