+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler.py
@@ -15,22 +15,10 @@ from .enum_and_styling import FeatureType
 from .elf_labeler_numba import (
     get_core_dist_ratios,
     get_core_dists,
-    get_approx_coulomb_potential,
-    get_zeff_nna,
-    get_valence_potentials,
-    solve_poisson,
-    get_avg_potentials,
-    get_test
     )
 
 Self = TypeVar("Self", bound="ElfLabeler")
 
-# TODO:
-    # 1. very small maxima in the charge density throw off the rest of the calculation.
-    # These need to be separated properly. Update persistence method to focus on
-    # the lower of the two maxima only
-    # 2. Compare distance of electrides and metals based on the distance past
-    # the core. Only compare for atoms that have a core.
 
 class ElfLabeler(BaseAnalysis):
     """
@@ -44,18 +32,17 @@ class ElfLabeler(BaseAnalysis):
         "basin_types",
         "attractor_shapes",
         "heavily_polarized",
-        "core_volume_ratios",
         "nna_bond_fracs",
-        "nna_bond_dists",
-        "neighbor_zeffs",
-        "neighbor_veffs",
-        "nna_potential_energies",
+        "nna_neighbor_dists",
         "nna_indices",
+        "nna_formula",
+        "nnas_per_formula",
+        "nnas_per_reduced_formula",
+        
         ]
 
     _reset_props = [
         "along_bond",
-        "electrostatic_potential",
         ] + _summary_props
 
 
@@ -66,9 +53,34 @@ class ElfLabeler(BaseAnalysis):
         reference_grid: Grid,
         total_charge_grid: Grid | None = None,
         polarization_cutoff: float = 0.5,
-        potential_grid: Grid | None =  None,
         **kwargs,
     ):
+        """
+        Labels each basin in the ELF as various chemical features.
+
+        This class is designed only for single spin or total spin charge densities
+        and ELF. For spin-dependent systems, use the SpinElfLabeler instead.
+
+        Parameters
+        ----------
+        charge_grid : Grid
+            The charge density grid used for integrating charge.
+        reference_grid : Grid
+            The ELF grid used to partition volumes.
+        total_charge_grid : Grid, optional
+            The total charge density used for bader integrations and vacuum masks. If
+            not provided, the charge_grid will be used instead.
+        polarization_cutoff: float, optional
+            The degree of polarization used for determining shared vs. unshared
+            behavior in a basin. O is more non-polar and 1 is more polar. This
+            is calculated from the two atoms that contribute the most to each
+            ELF basin.
+
+        
+        **kwargs : dict
+            Keyword arguments to pass to the Bader class.
+
+        """
         # create bader objects
         self._overlap = BasinOverlap(
             charge_grid=charge_grid,
@@ -76,8 +88,7 @@ class ElfLabeler(BaseAnalysis):
             reference_grid=reference_grid,
             **kwargs,
         )
-        self._potential_grid = potential_grid
-
+        
         self._elf_bader = self.overlap.local_bader
 
         super().__init__(
@@ -95,6 +106,17 @@ class ElfLabeler(BaseAnalysis):
 
     @property
     def polarization_cutoff(self) -> float:
+        """
+
+        Returns
+        -------
+        float
+            The degree of polarization used for determining shared vs. unshared
+            behavior in a basin. O is more non-polar and 1 is more polar. This
+            is calculated from the two atoms that contribute the most to each
+            ELF basin.
+
+        """
         return self._polarization_cutoff
 
     @polarization_cutoff.setter
@@ -112,195 +134,199 @@ class ElfLabeler(BaseAnalysis):
     ###########################################################################
     @property
     def overlap(self) -> BasinOverlap:
+        """
+
+        Returns
+        -------
+        BasinOverlap
[…]

[…]
