### 2026-04-29 d9f1212
start adding more radii types
