+++ src/baderkit/core/elf_analysis/badelf/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from .badelf import Badelf
-from .badelf_spin import SpinBadelf
+# from .badelf_spin import SpinBadelf
