+++ src/baderkit/core/elf_analysis/elf_labeler1/elf_labeler.py
@@ -1,496 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from pathlib import Path
-from typing import TypeVar
-
-import numpy as np
-from numpy.typing import NDArray
-
-from baderkit.core.base.base_analysis import BaseAnalysis
-from baderkit.core.bader.bader import Bader
-from baderkit.core.toolkit import Grid, Structure
-from baderkit.core.elf_analysis.overlap import BasinOverlap
-
-from .enum_and_styling import FeatureType
-from .elf_labeler_numba import (
-    get_core_dist_ratios,
-    get_core_dists,
-    )
-
-Self = TypeVar("Self", bound="ElfLabeler")
-
-# TODO: Add useful write methods?
-
-class ElfLabeler(BaseAnalysis):
-    """
-
-    A tool for labeling basins in a localization function (ELF, ELI-D, LOL, etc.)
-    as various chemical features.
-
-    """
-
-    spin_system = "total"
-
-    _summary_props = [
-        "basin_types",
-        "attractor_shapes",
-        "heavily_polarized",
-        "nna_bond_fracs",
-        "nna_neighbor_dists",
-        "nna_indices",
-        "nna_formula",
-        "nnas_per_formula",
-        "nnas_per_reduced_formula",
-        "label_structure",
-        ]
-
-    _reset_props = [
-        "nna_structure",
-        "along_bond",
-        ] + _summary_props
-
-
-
-    def __init__(
-        self,
-        charge_grid: Grid,
-        reference_grid: Grid,
-        total_charge_grid: Grid | None = None,
-        polarization_cutoff: float = 0.5,
-        **kwargs,
-    ):
-        """
-        Labels each basin in the ELF as various chemical features.
-
-        This class is designed only for single spin or total spin charge densities
-        and ELF. For spin-dependent systems, use the SpinElfLabeler instead.
-
-        Parameters
-        ----------
-        charge_grid : Grid
-            The charge density grid used for integrating charge.
-        reference_grid : Grid
-            The ELF grid used to partition volumes.
-        total_charge_grid : Grid, optional
-            The total charge density used for bader integrations and vacuum masks. If
-            not provided, the charge_grid will be used instead.
-        polarization_cutoff: float, optional
-            The degree of polarization used for determining shared vs. unshared
-            behavior in a basin. O is more non-polar and 1 is more polar. This
-            is calculated from the two atoms that contribute the most to each
-            ELF basin.
-
-
-        **kwargs : dict
-            Keyword arguments to pass to the Bader class.
-
-        """
-        # create bader objects
-        self._overlap = BasinOverlap(
-            charge_grid=charge_grid,
-            total_charge_grid=total_charge_grid,
-            reference_grid=reference_grid,
-            **kwargs,
-        )
-
-        self._elf_bader = self.overlap.local_bader
-
-        super().__init__(
-            charge_grid=charge_grid,
-            total_charge_grid=total_charge_grid,
-            reference_grid=reference_grid,
-            **kwargs,
-        )
-
-        self._polarization_cutoff = polarization_cutoff
-
-    ###########################################################################
-    # Settings
-    ###########################################################################
-
-    @property
-    def polarization_cutoff(self) -> float:
-        """
-
-        Returns
-        -------
-        float
-            The degree of polarization used for determining shared vs. unshared
-            behavior in a basin. O is more non-polar and 1 is more polar. This
-            is calculated from the two atoms that contribute the most to each
-            ELF basin.
-
-        """
-        return self._polarization_cutoff
-
-    @polarization_cutoff.setter
-    def polarization_cutoff(self, value: float):
-        self._polarization_cutoff = value
-        self._reset_properties(
-            include_properties=[
-                "heavily_polarized",
[…]
