+++ src/baderkit/core/elf_analysis/overlap/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .overlap import BasinOverlap
