+++ src/baderkit/core/__init__.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .toolkit import Grid, Structure  # isort:skip
-
-from .bader import Bader  # isort:skip
-
-# from .elf_analysis import (  # isort:skip
-#     Badelf,
-#     ElfLabeler,
-#     SpinBadelf,
-#     SpinElfLabeler,
-# )

[…]
