+++ src/baderkit/plotting/grid.py
@@ -1,0 +1,615 @@
+# -*- coding: utf-8 -*-
+
+from pathlib import Path
+
+import numpy as np
+import pyvista as pv
+from numpy.typing import NDArray
+
+from baderkit.toolkit.grid import Grid
+
+from .structure import StructurePlotter
+
+
+class GridPlotter(StructurePlotter):
+    def __init__(
+        self,
+        grid: Grid,
+        show_surface=True,
+        show_caps=True,
+        surface_opacity=0.8,
+        cap_opacity=0.8,
+        colormap="viridis",
+        use_solid_surface_color=False,
+        use_solid_cap_color=False,
+        surface_color="#BA8E23",
+        cap_color="#BA8E23",
+        iso_value=None,
+        **structure_kwargs,
+    ):
+        """
+        A convenience class for creating plots of crystal structures and isosurfaces
+        using pyvista's package for VTK.
+
+        Parameters
+        ----------
+        grid : Grid
+            The Grid object to use for isosurfaces. The structure will be pulled
+            from this grid.
+
+        Returns
+        -------
+        None.
+
+        """
+
+        self.grid = grid
+        self._show_surface = show_surface
+        self._show_caps = show_caps
+        self._surface_opacity = surface_opacity
+        self._cap_opacity = cap_opacity
+        self._colormap = colormap
+        self._use_solid_surface_color = use_solid_surface_color
+        self._use_solid_cap_color = use_solid_cap_color
+        self._surface_color = pv.Color(surface_color)
+        self._cap_color = pv.Color(cap_color)
+        self._hidden_mask = None  # for Bader class
+
+        self._min_val = grid.total.min() + 0.0000001
+        self._max_val = grid.total.max()
+        if iso_value is None:
+            self._iso_value = (self._min_val + self._max_val) / 4
+        else:
+            self._iso_value = max(self.min_val, min(iso_value, self.max_val))
+
+        # meshes
+        self._surface_mesh = None
+        self._cap_mesh = None
+        self._slice_meshes = {}
+        self._slice_planes = {}
+        self._slice_hkls = {}
+
+        # apply StructurePlotter kwargs
+        super().__init__(structure=grid.structure, **structure_kwargs)
+
+    @property
+    def show_surface(self) -> bool:
+        """
+
+        Returns
+        -------
+        bool
+            whether or not to display the isosurface.
+
+        """
+        return self._show_surface
+
+    @show_surface.setter
+    def show_surface(self, show_surface: bool):
+        if show_surface != self.show_surface:
+            self._show_surface = show_surface
+            self._update_surface_actor()
+
+    @property
+    def show_caps(self) -> bool:
+        """
+
+        Returns
+        -------
+        bool
+            Whether or not to display caps on the isosurface.
+
+        """
+        return self._show_caps
+
+    @show_caps.setter
+    def show_caps(self, show_caps: bool):
+        if show_caps != self.show_caps:
+            self._show_caps = show_caps
+            self._update_cap_actor()
+
+    @property
+    def surface_opacity(self) -> float:
+        """
+
+        Returns
+        -------
+        float
+            Opacity of the isosurface.
+
+        """
+        return self._surface_opacity
+
+    @surface_opacity.setter
+    def surface_opacity(self, surface_opacity: float):
+        if surface_opacity == self.surface_opacity:
+            return
+        actor = self.plotter.actors.get("iso", None)
+        if actor is not None:
+            actor.prop.opacity = surface_opacity
+        self._surface_opacity = surface_opacity
+
+    @property
+    def cap_opacity(self) -> float:
+        """
+
+        Returns
+        -------
+        float
+            Opacity of the caps.
+
+        """
[…]
