### 2026-05-01 028907c
update write methods
