+++ pyproject.toml
@@ -57,7 +57,7 @@ dependencies = [
     "scipy>=1.7.2,<2.0",
     "numba>=0.59.0,<=0.64.0",
     "typer>=0.6.1",
-    "rich>=11.0,<=14.3.3",
+    "rich>=11.0,<=15.0.0",
 ]
 
 
