### 2026-05-02 044a9ac
Update pyvista requirement from <=0.47.1,>=0.44.0 to >=0.44.0,<=0.47.3
Updates the requirements on [pyvista](https://github.com/pyvista/pyvista) to permit the latest version.
- [Release notes](https://github.com/pyvista/pyvista/releases)
- [Commits](https://github.com/pyvista/pyvista/compare/v0.44.0...v0.47.3)

---
updated-dependencies:
- dependency-name: pyvista
  dependency-version: 0.47.3
  dependency-type: direct:development
...

Signed-off-by: dependabot[bot] <support@github.com>
