### 2026-05-02 da4e213
Update pyvistaqt requirement from <=0.11.3,>=0.6.0 to >=0.6.0,<=0.11.4
Updates the requirements on [pyvistaqt](https://github.com/pyvista/pyvistaqt) to permit the latest version.
- [Release notes](https://github.com/pyvista/pyvistaqt/releases)
- [Commits](https://github.com/pyvista/pyvistaqt/compare/0.6.0...0.11.4)

---
updated-dependencies:
- dependency-name: pyvistaqt
  dependency-version: 0.11.4
  dependency-type: direct:development
...

Signed-off-by: dependabot[bot] <support@github.com>
