### 2026-05-02 a04c23e
Update numba requirement from <=0.64.0,>=0.59.0 to >=0.59.0,<=0.65.1
Updates the requirements on [numba](https://github.com/numba/numba) to permit the latest version.
- [Release notes](https://github.com/numba/numba/releases)
- [Commits](https://github.com/numba/numba/compare/0.59.0...0.65.1)

---
updated-dependencies:
- dependency-name: numba
  dependency-version: 0.65.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
