### 2026-05-02 ce66ae9
add crystalnn option for radii
