+++ src/baderkit/elf_analysis/elf_radii/elf_radii.py
@@ -32,9 +32,7 @@ class ElfRadii(BaseAnalysis):
 
     """
 
-    _method_kwargs = [
-        "include_nnas",
-    ]
+    _method_kwargs = ["include_nnas", "cnn_kwargs"]
 
     _radii_results = [
         "bonding_pairs",
@@ -65,6 +63,7 @@ class ElfRadii(BaseAnalysis):
         reference_grid: Grid,
         total_charge_grid: Grid | None = None,
         include_nnas: bool = False,
+        cnn_kwargs: dict | None = None,
         **kwargs,
     ):
         """
@@ -97,6 +96,11 @@ class ElfRadii(BaseAnalysis):
             Whether or not to treat non-nuclear attractors as quasi atoms. If
             set to true, they will be included as central points for the generated
             weighted voronoi surface and be given calculated radii.
+        cnn_kwargs : dict | None, optional
+            If provided, the nearest neighbors will be determined using PyMatGen's CrystalNN
+            class using the keyword arguments in this argument. If not provided, all neighbors
+            that share a weighted voronoi facet (constructed from the ELF radii) will be
+            included.
         **kwargs : dict
             Keyword arguments to pass to the ElfLabeler class.
 
@@ -107,10 +111,20 @@ class ElfRadii(BaseAnalysis):
             charge_grid=charge_grid,
             total_charge_grid=total_charge_grid,
             reference_grid=reference_grid,
+            cnn_kwargs=None,
             **kwargs,
         )
 
         self._include_nnas = include_nnas
+
+        if cnn_kwargs is not None:
+            self._use_cnn = True
+            self._cnn_kwargs = cnn_kwargs
+            self._cnn = CrystalNN(**cnn_kwargs)
+        else:
+            self._use_cnn = False
+            self._cnn_kwargs = None
+            self._cnn = None
 
         super().__init__(
             charge_grid=charge_grid,
@@ -140,6 +154,31 @@ class ElfRadii(BaseAnalysis):
     @include_nnas.setter
     def include_nnas(self, value: bool):
         self._include_nnas = value
+        self._reset_properties()
+
+    @property
+    def cnn_kwargs(self) -> dict | None:
+        """
+
+        Returns
+        -------
+        dict
+            The keyword arguments used to construct the CrystalNN
+            object.
+
+        """
+        return self._cnn_kwargs
+
+    @cnn_kwargs.setter
+    def cnn_kwargs(self, value: dict | None):
+        if value is not None:
+            self._cnn_kwargs = value
+            self._cnn = CrystalNN(**value)
+            self._use_cnn = True
+        else:
+            self._cnn_kwargs = value
+            self._cnn = value
+            self._use_cnn = False
         self._reset_properties()
 
     ###########################################################################
@@ -241,6 +280,19 @@ class ElfRadii(BaseAnalysis):
 
         return self._label_atom_map
 
+    @property
+    def cnn(self) -> CrystalNN:
+        """
+
+        Returns
+        -------
+        CrystalNN
+            If cnn_kwargs were provided, the CrystalNN object used to
+            determine coordination environments.
+
+        """
+        return self._cnn
+
     ###########################################################################
     # Radii Properties
     ###########################################################################
@@ -258,7 +310,7 @@ class ElfRadii(BaseAnalysis):
 
         """
         if self._bonding_pairs is None:
-            self._get_voronoi_radii()
+            self._get_radii()
         return self._bonding_pairs
 
     @property
@@ -273,7 +325,7 @@ class ElfRadii(BaseAnalysis):
 
         """
         if self._all_radii is None:
-            self._get_voronoi_radii()
+            self._get_radii()
         return self._all_radii
 
[…]
